unifies code path for particles and grid, and changes MPI communicator setup
