+++ src/VCluster/InVis.cpp
@@ -11,24 +11,39 @@
 #include <vector>
 #include <mpi.h>
 #include <dirent.h>
+#include <fstream>
+
 
 #include "InVis.hpp"
 #include "memory/ShmBuffer.hpp"
 
 #define USE_VULKAN true
 
-MPI_Comm comm;
+MPI_Comm renComm;
 int imgSize;
+double avg_recent = 0.0;
+int prevtime = 0;
+int count = 0;
+int totaltime = 0;
+int timeVDIwrite = 0;
 
-InVis::InVis(int wSize, int cPartners, MPI_Comm vComm, bool isHead) {
-    windowSize = wSize;
+InVis::InVis(int cPartners, MPI_Comm vComm, MPI_Comm rComm) {
+    windowSize = 700;
     computePartners = cPartners;
-    imageSize = windowSize*windowSize*7;
+    imageSize = windowSize*windowSize*7; //RGBA + 4 bytes for depth
     visComm = vComm;
-    comm = vComm;
-    MPI_Comm_size(visComm, &commSize);
-    std::cout<<"VisComm size is "<<commSize <<std::endl;
+    renderComm = rComm;
+    renComm = rComm;
+    MPI_Comm_size(renderComm, &rCommSize);
+    std::cout << "RenderComm size is " << rCommSize << std::endl;
+    MPI_Comm_size(visComm, &vCommSize);
+    std::cout << "VisComm size is " << vCommSize << std::endl;
     imgSize = imageSize;
+
+    for(int i = 0; i< computePartners; i++) {
+        std::vector<ShmBuffer *> tmp;
+        gridBuffers.push_back(tmp);
+    }
 
     DIR *dir;
     struct dirent *ent;
@@ -50,19 +65,20 @@ InVis::InVis(int wSize, int cPartners, MPI_Comm vComm, bool isHead) {
 
     std::cout<<"classpath is "<<classPath<<std::endl;
 
+    // determine whether particle or grid data is to be rendered
+    vtype = getVisType();
+
+    std::cout<<"Got the data type "<<std::endl;
+
+    std::string className;
+    if(vtype == vis_type::grid) className = "graphics/scenery/insitu/InVisSimpleVolumeRenderer";
+    else className = "graphics/scenery/insitu/InVisRenderer";
+
     //================== prepare loading of Java VM ============================
     JavaVMInitArgs vm_args;                        // Initialization arguments
     auto *options = new JavaVMOption[4];   // JVM invocation options
     options[0].optionString = (char *)classPath.c_str();
-//current one:
-// "-Djava.class.path=/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.base:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.compiler:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.datatransfer:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.desktop:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.instrument:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.logging:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.management:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.management.rmi:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.naming:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.net.http:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.prefs:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.rmi:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.scripting:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.se:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.security.jgss:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.security.sasl:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.smartcardio:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.sql:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.sql.rowset:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.transaction.xa:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.xml:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.xml.crypto:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.accessibility:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.aot:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.attach:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.charsets:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.compiler:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.crypto.cryptoki:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.crypto.ec:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.dynalink:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.editpad:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.hotspot.agent:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.httpserver:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.internal.ed:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.internal.jvmstat:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.internal.le:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.internal.opt:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.internal.vm.ci:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.internal.vm.compiler:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.internal.vm.compiler.management:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jartool:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.javadoc:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jcmd:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jconsole:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jdeps:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jdi:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jdwp.agent:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jfr:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jlink:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jshell:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jsobject:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jstatd:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.localedata:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.management:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.management.agent:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.management.jfr:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.naming.dns:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.naming.rmi:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.net:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.pack:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.rmic:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.scripting.nashorn:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.scripting.nashorn.shell:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.sctp:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.security.auth:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.security.jgss:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.unsupported:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.unsupported.desktop:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.xml.dom:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.zipfs:/home/aryaman/Repositories/scenery/target/test-classes:/home/aryaman/Repositories/scenery/target/classes:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-jdk8/1.4-M1/kotlin-stdlib-jdk8-1.4-M1.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib/1.4-M1/kotlin-stdlib-1.4-M1.jar:/home/aryaman/.m2/repository/org/jetbrains/annotations/13.0/annotations-13.0.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-jdk7/1.4-M1/kotlin-stdlib-jdk7-1.4-M1.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-reflect/1.4-M1/kotlin-reflect-1.4-M1.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlinx/kotlinx-coroutines-core/1.3.5-1.4-M1/kotlinx-coroutines-core-1.3.5-1.4-M1.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-common/1.4-M1/kotlin-stdlib-common-1.4-M1.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-linux-amd64.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-linux-i586.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-macosx-universal.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-windows-amd64.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-linux-amd64.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-linux-i586.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-macosx-universal.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-windows-amd64.jar:/home/aryaman/.m2/repository/junit/junit/4.13/junit-4.13.jar:/home/aryaman/.m2/repository/org/hamcrest/hamcrest-core/1.3/hamcrest-core-1.3.jar:/home/aryaman/.m2/repository/org/slf4j/slf4j-api/1.7.25/slf4j-api-1.7.25.jar:/home/aryaman/.m2/repository/org/slf4j/slf4j-simple/1.7.25/slf4j-simple-1.7.25.jar:/home/aryaman/.m2/repository/net/clearvolume/cleargl/2.2.10/cleargl-2.2.10.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt-main/2.3.2/gluegen-rt-main-2.3.2.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-android-aarch64.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-android-armv6.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-linux-armv6.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-linux-armv6hf.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-solaris-amd64.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-solaris-i586.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-windows-i586.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all-main/2.3.2/jogl-all-main-2.3.2.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-android-aarch64.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-android-armv6.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-linux-armv6.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-linux-armv6hf.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-solaris-amd64.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-solaris-i586.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-windows-i586.jar:/home/aryaman/.m2/repository/org/joml/joml/1.9.22/joml-1.9.22.jar:/home/aryaman/.m2/repository/com/github/scenerygraphics/vector/958f2e6/vector-958f2e6.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2/5.9.2/imglib2-5.9.2.jar:/home/aryaman/.m2/repository/net/clearcontrol/coremem/0.4.6/coremem-0.4.6.jar:/home/aryaman/.m2/repository/com/nativelibs4java/bridj/0.7.0/bridj-0.7.0.jar:/home/aryaman/.m2/repository/net/java/jinput/jinput/2.0.9/jinput-2.0.9.jar:/home/aryaman/.m2/repository/net/java/jinput/jinput/2.0.9/jinput-2.0.9-natives-all.jar:/home/aryaman/.m2/repository/org/scijava/scijava-common/2.83.2/scijava-common-2.83.2.jar:/home/aryaman/.m2/repository/org/scijava/parsington/2.0.0/parsington-2.0.0.jar:/home/aryaman/.m2/repository/org/bushe/eventbus/1.4/eventbus-1.4.jar:/home/aryaman/.m2/repository/org/scijava/script-editor/0.5.3/script-editor-0.5.3.jar:/home/aryaman/.m2/repository/org/scijava/scripting-java/0.4.1/scripting-java-0.4.1.jar:/home/aryaman/.m2/repository/org/scijava/minimaven/2.2.2/minimaven-2.2.2.jar:/home/aryaman/.m2/repository/com/fifesoft/rsyntaxtextarea/3.1.1/rsyntaxtextarea-3.1.1.jar:/home/aryaman/.m2/repository/com/fifesoft/languagesupport/3.1.0/languagesupport-3.1.0.jar:/home/aryaman/.m2/repository/com/fifesoft/autocomplete/3.1.0/autocomplete-3.1.0.jar:/home/aryaman/.m2/repository/org/mozilla/rhino/1.7.6/rhino-1.7.6.jar:/home/aryaman/.m2/repository/com/miglayout/miglayout/3.7.4/miglayout-3.7.4-swing.jar:/home/aryaman/.m2/repository/org/scijava/batch-processor/0.4.1/batch-processor-0.4.1.jar:/home/aryaman/.m2/repository/org/scijava/scijava-table/0.5.0/scijava-table-0.5.0.jar:/home/aryaman/.m2/repository/org/scijava/scijava-ui-swing/0.13.2/scijava-ui-swing-0.13.2.jar:/home/aryaman/.m2/repository/org/scijava/swing-checkbox-tree/1.0.2/swing-checkbox-tree-1.0.2.jar:/home/aryaman/.m2/repository/net/sourceforge/jdatepicker/jdatepicker/1.3.2/jdatepicker-1.3.2.jar:/home/aryaman/.m2/repository/org/scijava/scijava-search/0.7.0/scijava-search-0.7.0.jar:/home/aryaman/.m2/repository/org/ocpsoft/prettytime/prettytime/4.0.1.Final/prettytime-4.0.1.Final.jar:/home/aryaman/.m2/repository/commons-io/commons-io/2.6/commons-io-2.6.jar:/home/aryaman/.m2/repository/org/scijava/ui-behaviour/2.0.1/ui-behaviour-2.0.1.jar:/home/aryaman/.m2/repository/com/google/code/gson/gson/2.8.6/gson-2.8.6.jar:/home/aryaman/.m2/repository/org/yaml/snakeyaml/1.26/snakeyaml-1.26.jar:/home/aryaman/.m2/repository/net/sf/trove4j/trove4j/3.0.3/trove4j-3.0.3.jar:/home/aryaman/.m2/repository/org/scijava/scripting-javascript/0.5.0/scripting-javascript-0.5.0.jar:/home/aryaman/.m2/repository/org/scijava/scripting-jython/1.0.0/scripting-jython-1.0.0.jar:/home/aryaman/.m2/repository/org/python/jython-slim/2.7.2/jython-slim-2.7.2.jar:/home/aryaman/.m2/repository/org/antlr/antlr/3.5.2/antlr-3.5.2.jar:/home/aryaman/.m2/repository/javax/servlet/javax.servlet-api/3.1.0/javax.servlet-api-3.1.0.jar:/home/aryaman/.m2/repository/org/antlr/antlr-runtime/3.5.2/antlr-runtime-3.5.2.jar:/home/aryaman/.m2/repository/org/bouncycastle/bcpkix-jdk15on/1.62/bcpkix-jdk15on-1.62.jar:/home/aryaman/.m2/repository/org/bouncycastle/bcprov-jdk15on/1.62/bcprov-jdk15on-1.62.jar:/home/aryaman/.m2/repository/org/ow2/asm/asm/7.1/asm-7.1.jar:/home/aryaman/.m2/repository/org/ow2/asm/asm-commons/7.1/asm-commons-7.1.jar:/home/aryaman/.m2/repository/org/ow2/asm/asm-tree/7.1/asm-tree-7.1.jar:/home/aryaman/.m2/repository/org/ow2/asm/asm-analysis/7.1/asm-analysis-7.1.jar:/home/aryaman/.m2/repository/org/ow2/asm/asm-util/7.1/asm-util-7.1.jar:/home/aryaman/.m2/repository/com/google/guava/failureaccess/1.0.1/failureaccess-1.0.1.jar:/home/aryaman/.m2/repository/com/ibm/icu/icu4j/59.1/icu4j-59.1.jar:/home/aryaman/.m2/repository/com/carrotsearch/java-sizeof/0.0.5/java-sizeof-0.0.5.jar:/home/aryaman/.m2/repository/com/github/jnr/jffi/1.2.20/jffi-1.2.20.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-netdb/1.1.6/jnr-netdb-1.1.6.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-ffi/2.1.10/jnr-ffi-2.1.10.jar:/home/aryaman/.m2/repository/com/github/jnr/jffi/1.2.19/jffi-1.2.19-native.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-a64asm/1.0.0/jnr-a64asm-1.0.0.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-x86asm/1.0.2/jnr-x86asm-1.0.2.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-posix/3.0.50/jnr-posix-3.0.50.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-constants/0.9.12/jnr-constants-0.9.12.jar:/home/aryaman/.m2/repository/jline/jline/2.14.5/jline-2.14.5.jar:/home/aryaman/.m2/repository/io/netty/netty-buffer/4.1.45.Final/netty-buffer-4.1.45.Final.jar:/home/aryaman/.m2/repository/io/netty/netty-codec/4.1.45.Final/netty-codec-4.1.45.Final.jar:/home/aryaman/.m2/repository/io/netty/netty-common/4.1.45.Final/netty-common-4.1.45.Final.jar:/home/aryaman/.m2/repository/io/netty/netty-handler/4.1.45.Final/netty-handler-4.1.45.Final.jar:/home/aryaman/.m2/repository/io/netty/netty-resolver/4.1.45.Final/netty-resolver-4.1.45.Final.jar:/home/aryaman/.m2/repository/io/netty/netty-transport/4.1.45.Final/netty-transport-4.1.45.Final.jar:/home/aryaman/.m2/repository/org/apache/ant/ant/1.9.7/ant-1.9.7.jar:/home/aryaman/.m2/repository/org/apache/ant/ant-launcher/1.9.7/ant-launcher-1.9.7.jar:/home/aryaman/.m2/repository/net/java/dev/jna/jna/4.5.2/jna-4.5.2.jar:/home/aryaman/.m2/repository/net/java/dev/jna/jna-platform/4.5.2/jna-platform-4.5.2.jar:/home/aryaman/.m2/repository/org/jocl/jocl/2.0.1/jocl-2.0.1.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl/3.2.3/lwjgl-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-glfw/3.2.3/lwjgl-glfw-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-jemalloc/3.2.3/lwjgl-jemalloc-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-vulkan/3.2.3/lwjgl-vulkan-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-opengl/3.2.3/lwjgl-opengl-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-openvr/3.2.3/lwjgl-openvr-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-xxhash/3.2.3/lwjgl-xxhash-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl/3.2.3/lwjgl-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-opengl/3.2.3/lwjgl-opengl-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-glfw/3.2.3/lwjgl-glfw-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-jemalloc/3.2.3/lwjgl-jemalloc-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-openvr/3.2.3/lwjgl-openvr-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-xxhash/3.2.3/lwjgl-xxhash-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-remotery/3.2.3/lwjgl-remotery-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-remotery/3.2.3/lwjgl-remotery-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/core/jackson-databind/2.9.10.1/jackson-databind-2.9.10.1.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/core/jackson-annotations/2.9.10/jackson-annotations-2.9.10.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/core/jackson-core/2.9.10/jackson-core-2.9.10.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/module/jackson-module-kotlin/2.9.9/jackson-module-kotlin-2.9.9.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/dataformat/jackson-dataformat-yaml/2.9.9/jackson-dataformat-yaml-2.9.9.jar:/home/aryaman/.m2/repository/graphics/scenery/spirvcrossj/0.7.0-1.1.106.0/spirvcrossj-0.7.0-1.1.106.0.jar:/home/aryaman/.m2/repository/graphics/scenery/spirvcrossj/0.7.0-1.1.106.0/spirvcrossj-0.7.0-1.1.106.0-natives-linux.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-junit/1.4-M1/kotlin-test-junit-1.4-M1.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-annotations-common/1.4-M1/kotlin-test-annotations-common-1.4-M1.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test/1.4-M1/kotlin-test-1.4-M1.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-common/1.4-M1/kotlin-test-common-1.4-M1.jar:/home/aryaman/.m2/repository/org/zeromq/jeromq/0.4.3/jeromq-0.4.3.jar:/home/aryaman/.m2/repository/org/zeromq/jnacl/0.1.0/jnacl-0.1.0.jar:/home/aryaman/.m2/repository/com/esotericsoftware/kryo/4.0.2/kryo-4.0.2.jar:/home/aryaman/.m2/repository/com/esotericsoftware/reflectasm/1.11.3/reflectasm-1.11.3.jar:/home/aryaman/.m2/repository/com/esotericsoftware/minlog/1.3.0/minlog-1.3.0.jar:/home/aryaman/.m2/repository/org/objenesis/objenesis/2.5.1/objenesis-2.5.1.jar:/home/aryaman/.m2/repository/org/msgpack/msgpack-core/0.8.16/msgpack-core-0.8.16.jar:/home/aryaman/.m2/repository/org/msgpack/jackson-dataformat-msgpack/0.8.16/jackson-dataformat-msgpack-0.8.16.jar:/home/aryaman/.m2/repository/graphics/scenery/jvrpn/1.1.0/jvrpn-1.1.0.jar:/home/aryaman/.m2/repository/graphics/scenery/jvrpn/1.1.0/jvrpn-1.1.0-natives-linux.jar:/home/aryaman/.m2/repository/io/scif/scifio/0.40.0/scifio-0.40.0.jar:/home/aryaman/.m2/repository/io/scif/scifio-jai-imageio/1.1.1/scifio-jai-imageio-1.1.1.jar:/home/aryaman/.m2/repository/net/imagej/imagej-common/0.33.0/imagej-common-0.33.0.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-roi/0.10.2/imglib2-roi-0.10.2.jar:/home/aryaman/.m2/repository/edu/ucar/udunits/4.3.18/udunits-4.3.18.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-cache/1.0.0-beta-13/imglib2-cache-1.0.0-beta-13.jar:/home/aryaman/.m2/repository/com/github/ben-manes/caffeine/caffeine/2.4.0/caffeine-2.4.0.jar:/home/aryaman/.m2/repository/org/bytedeco/ffmpeg/4.2.1-1.5.2/ffmpeg-4.2.1-1.5.2.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp/1.5.2/javacpp-1.5.2.jar:/home/aryaman/.m2/repository/org/bytedeco/ffmpeg/4.2.1-1.5.2/ffmpeg-4.2.1-1.5.2-linux-x86_64.jar:/home/aryaman/.m2/repository/org/bytedeco/ffmpeg/4.2.1-1.5.2/ffmpeg-4.2.1-1.5.2-macosx-x86_64.jar:/home/aryaman/.m2/repository/org/bytedeco/ffmpeg/4.2.1-1.5.2/ffmpeg-4.2.1-1.5.2-windows-x86_64.jar:/home/aryaman/.m2/repository/org/reflections/reflections/0.9.11/reflections-0.9.11.jar:/home/aryaman/.m2/repository/com/google/guava/guava/21.0/guava-21.0.jar:/home/aryaman/.m2/repository/org/javassist/javassist/3.27.0-GA/javassist-3.27.0-GA.jar:/home/aryaman/.m2/repository/sc/fiji/bigvolumeviewer/0.1.8/bigvolumeviewer-0.1.8.jar:/home/aryaman/.m2/repository/commons-lang/commons-lang/2.6/commons-lang-2.6.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-algorithm/0.11.2/imglib2-algorithm-0.11.2.jar:/home/aryaman/.m2/repository/gov/nist/math/jama/1.0.3/jama-1.0.3.jar:/home/aryaman/.m2/repository/org/ojalgo/ojalgo/45.1.1/ojalgo-45.1.1.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-realtransform/3.0.0/imglib2-realtransform-3.0.0.jar:/home/aryaman/.m2/repository/jitk/jitk-tps/3.0.1/jitk-tps-3.0.1.jar:/home/aryaman/.m2/repository/com/googlecode/efficient-java-matrix-library/ejml/0.24/ejml-0.24.jar:/home/aryaman/.m2/repository/log4j/log4j/1.2.17/log4j-1.2.17.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-ui/2.0.1/imglib2-ui-2.0.1.jar:/home/aryaman/.m2/repository/org/antlr/ST4/4.0.8/ST4-4.0.8.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-math3/3.6.1/commons-math3-3.6.1.jar:/home/aryaman/.m2/repository/org/jdom/jdom2/2.0.6/jdom2-2.0.6.jar:/home/aryaman/.m2/repository/sc/fiji/bigdataviewer-core/9.0.2/bigdataviewer-core-9.0.2.jar:/home/aryaman/.m2/repository/cisd/jhdf5/14.12.6/jhdf5-14.12.6.jar:/home/aryaman/.m2/repository/org/scijava/scijava-listeners/1.0.0-beta-2/scijava-listeners-1.0.0-beta-2.jar:/home/aryaman/.m2/repository/sc/fiji/bigdataviewer-vistools/1.0.0-beta-20/bigdataviewer-vistools-1.0.0-beta-20.jar:/home/aryaman/.m2/repository/sc/fiji/spim_data/2.2.4/spim_data-2.2.4.jar:/home/aryaman/.m2/repository/org/lwjglx/lwjgl3-awt/0.1.7/lwjgl3-awt-0.1.7.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-jawt/3.2.3/lwjgl-jawt-3.2.3.jar:/home/aryaman/.m2/repository/net/imagej/imagej/2.0.0-rc-72/imagej-2.0.0-rc-72.jar:/home/aryaman/.m2/repository/net/imagej/imagej-notebook/0.7.1/imagej-notebook-0.7.1.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-text/1.8/commons-text-1.8.jar:/home/aryaman/.m2/repository/org/knowm/xchart/xchart/3.5.4/xchart-3.5.4.jar:/home/aryaman/.m2/repository/de/erichseifert/vectorgraphics2d/VectorGraphics2D/0.13/VectorGraphics2D-0.13.jar:/home/aryaman/.m2/repository/net/imagej/imagej-ops/0.45.5/imagej-ops-0.45.5.jar:/home/aryaman/.m2/repository/net/imagej/imagej-mesh/0.8.0/imagej-mesh-0.8.0.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-algorithm-fft/0.2.0/imglib2-algorithm-fft-0.2.0.jar:/home/aryaman/.m2/repository/edu/mines/mines-jtk/20151125/mines-jtk-20151125.jar:/home/aryaman/.m2/repository/net/imagej/imagej-updater/0.10.4/imagej-updater-0.10.4.jar:/home/aryaman/.m2/repository/net/imagej/imagej-plugins-batch/0.1.1/imagej-plugins-batch-0.1.1.jar:/home/aryaman/.m2/repository/net/imagej/imagej-deprecated/0.1.4/imagej-deprecated-0.1.4.jar:/home/aryaman/.m2/repository/net/imagej/imagej-plugins-commands/0.8.1/imagej-plugins-commands-0.8.1.jar:/home/aryaman/.m2/repository/com/github/sbridges/object-inspector/object-inspector/0.1/object-inspector-0.1.jar:/home/aryaman/.m2/repository/net/iharder/base64/2.3.8/base64-2.3.8.jar:/home/aryaman/.m2/repository/net/imagej/imagej-plugins-tools/0.3.1/imagej-plugins-tools-0.3.1.jar:/home/aryaman/.m2/repository/net/imagej/imagej-plugins-uploader-ssh/0.3.2/imagej-plugins-uploader-ssh-0.3.2.jar:/home/aryaman/.m2/repository/com/jcraft/jsch/0.1.55/jsch-0.1.55.jar:/home/aryaman/.m2/repository/net/imagej/imagej-plugins-uploader-webdav/0.3.2/imagej-plugins-uploader-webdav-0.3.2.jar:/home/aryaman/.m2/repository/org/apache/jackrabbit/jackrabbit-webdav/2.21.1/jackrabbit-webdav-2.21.1.jar:/home/aryaman/.m2/repository/org/apache/httpcomponents/httpclient/4.5.12/httpclient-4.5.12.jar:/home/aryaman/.m2/repository/commons-codec/commons-codec/1.14/commons-codec-1.14.jar:/home/aryaman/.m2/repository/org/apache/httpcomponents/httpcore/4.4.13/httpcore-4.4.13.jar:/home/aryaman/.m2/repository/net/imagej/imagej-scripting/0.8.2/imagej-scripting-0.8.2.jar:/home/aryaman/.m2/repository/org/scijava/scripting-beanshell/0.3.4/scripting-beanshell-0.3.4.jar:/home/aryaman/.m2/repository/org/apache-extras/beanshell/bsh/2.0b6/bsh-2.0b6.jar:/home/aryaman/.m2/repository/org/scijava/scripting-clojure/0.1.6/scripting-clojure-0.1.6.jar:/home/aryaman/.m2/repository/org/clojure/clojure/1.8.0/clojure-1.8.0.jar:/home/aryaman/.m2/repository/org/scijava/scripting-groovy/0.3.0/scripting-groovy-0.3.0.jar:/home/aryaman/.m2/repository/org/codehaus/groovy/groovy/3.0.4/groovy-3.0.4.jar:/home/aryaman/.m2/repository/org/codehaus/groovy/groovy-dateutil/3.0.4/groovy-dateutil-3.0.4.jar:/home/aryaman/.m2/repository/org/codehaus/groovy/groovy-json/3.0.4/groovy-json-3.0.4.jar:/home/aryaman/.m2/repository/org/codehaus/groovy/groovy-swing/3.0.4/groovy-swing-3.0.4.jar:/home/aryaman/.m2/repository/org/codehaus/groovy/groovy-templates/3.0.4/groovy-templates-3.0.4.jar:/home/aryaman/.m2/repository/org/codehaus/groovy/groovy-xml/3.0.4/groovy-xml-3.0.4.jar:/home/aryaman/.m2/repository/org/apache/ivy/ivy/2.4.0/ivy-2.4.0.jar:/home/aryaman/.m2/repository/org/codehaus/gpars/gpars/1.2.1/gpars-1.2.1.jar:/home/aryaman/.m2/repository/org/multiverse/multiverse-core/0.7.0/multiverse-core-0.7.0.jar:/home/aryaman/.m2/repository/org/codehaus/jsr166-mirror/jsr166y/1.7.0/jsr166y-1.7.0.jar:/home/aryaman/.m2/repository/org/scijava/scripting-jruby/0.3.1/scripting-jruby-0.3.1.jar:/home/aryaman/.m2/repository/org/jruby/jruby-core/9.1.17.0/jruby-core-9.1.17.0.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-enxio/0.16/jnr-enxio-0.16.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-unixsocket/0.17/jnr-unixsocket-0.17.jar:/home/aryaman/.m2/repository/org/jruby/joni/joni/2.1.15/joni-2.1.15.jar:/home/aryaman/.m2/repository/org/jruby/extras/bytelist/1.0.15/bytelist-1.0.15.jar:/home/aryaman/.m2/repository/org/jruby/jcodings/jcodings/1.0.27/jcodings-1.0.27.jar:/home/aryaman/.m2/repository/org/jruby/dirgra/0.3/dirgra-0.3.jar:/home/aryaman/.m2/repository/com/headius/invokebinder/1.10/invokebinder-1.10.jar:/home/aryaman/.m2/repository/com/headius/options/1.4/options-1.4.jar:/home/aryaman/.m2/repository/com/headius/unsafe-fences/1.0/unsafe-fences-1.0.jar:/home/aryaman/.m2/repository/com/jcraft/jzlib/1.1.3/jzlib-1.1.3.jar:/home/aryaman/.m2/repository/com/martiansoftware/nailgun-server/0.9.1/nailgun-server-0.9.1.jar:/home/aryaman/.m2/repository/joda-time/joda-time/2.10.6/joda-time-2.10.6.jar:/home/aryaman/.m2/repository/com/headius/modulator/1.0/modulator-1.0.jar:/home/aryaman/.m2/repository/org/jruby/jruby-stdlib/9.1.17.0/jruby-stdlib-9.1.17.0.jar:/home/aryaman/.m2/repository/org/scijava/scripting-renjin/0.2.3/scripting-renjin-0.2.3.jar:/home/aryaman/.m2/repository/org/renjin/renjin-script-engine/0.8.1906/renjin-script-engine-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/renjin-core/0.8.1906/renjin-core-0.8.1906.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-math/2.2/commons-math-2.2.jar:/home/aryaman/.m2/repository/org/renjin/gcc-runtime/0.8.1906/gcc-runtime-0.8.1906.jar:/home/aryaman/.m2/repository/org/netlib/netlib-java/0.9.3-renjin-patched-2/netlib-java-0.9.3-renjin-patched-2.jar:/home/aryaman/.m2/repository/org/netlib/lapack/0.8/lapack-0.8.jar:/home/aryaman/.m2/repository/org/netlib/f2jutil/0.8/f2jutil-0.8.jar:/home/aryaman/.m2/repository/org/netlib/xerbla/0.8/xerbla-0.8.jar:/home/aryaman/.m2/repository/org/netlib/blas/0.8/blas-0.8.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-vfs2/2.0/commons-vfs2-2.0.jar:/home/aryaman/.m2/repository/commons-logging/commons-logging/1.2/commons-logging-1.2.jar:/home/aryaman/.m2/repository/org/apache/maven/scm/maven-scm-api/1.4/maven-scm-api-1.4.jar:/home/aryaman/.m2/repository/org/codehaus/plexus/plexus-utils/1.5.6/plexus-utils-1.5.6.jar:/home/aryaman/.m2/repository/org/apache/maven/scm/maven-scm-provider-svnexe/1.4/maven-scm-provider-svnexe-1.4.jar:/home/aryaman/.m2/repository/org/apache/maven/scm/maven-scm-provider-svn-commons/1.4/maven-scm-provider-svn-commons-1.4.jar:/home/aryaman/.m2/repository/regexp/regexp/1.3/regexp-1.3.jar:/home/aryaman/.m2/repository/edu/emory/mathcs/jtransforms/2.4/jtransforms-2.4.jar:/home/aryaman/.m2/repository/net/sf/jung/jung-api/2.0.1/jung-api-2.0.1.jar:/home/aryaman/.m2/repository/net/sourceforge/collections/collections-generic/4.01/collections-generic-4.01.jar:/home/aryaman/.m2/repository/net/sf/jung/jung-graph-impl/2.0.1/jung-graph-impl-2.0.1.jar:/home/aryaman/.m2/repository/com/sun/codemodel/codemodel/2.6/codemodel-2.6.jar:/home/aryaman/.m2/repository/org/renjin/stats/0.8.1906/stats-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/renjin-appl/0.8.1906/renjin-appl-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/renjin-gnur-runtime/0.8.1906/renjin-gnur-runtime-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/methods/0.8.1906/methods-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/datasets/0.8.1906/datasets-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/utils/0.8.1906/utils-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/grDevices/0.8.1906/grDevices-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/graphics/0.8.1906/graphics-0.8.1906.jar:/home/aryaman/.m2/repository/org/scijava/scripting-scala/0.2.2/scripting-scala-0.2.2.jar:/home/aryaman/.m2/repository/org/scala-lang/scala-compiler/2.12.1/scala-compiler-2.12.1.jar:/home/aryaman/.m2/repository/org/scala-lang/scala-library/2.12.1/scala-library-2.12.1.jar:/home/aryaman/.m2/repository/org/scala-lang/scala-reflect/2.12.1/scala-reflect-2.12.1.jar:/home/aryaman/.m2/repository/org/scala-lang/modules/scala-xml_2.12/1.0.6/scala-xml_2.12-1.0.6.jar:/home/aryaman/.m2/repository/net/imagej/imagej-ui-swing/0.23.2/imagej-ui-swing-0.23.2.jar:/home/aryaman/.m2/repository/org/scijava/scijava-ui-awt/0.1.7/scijava-ui-awt-0.1.7.jar:/home/aryaman/.m2/repository/com/miglayout/miglayout-swing/5.2/miglayout-swing-5.2.jar:/home/aryaman/.m2/repository/com/miglayout/miglayout-core/5.2/miglayout-core-5.2.jar:/home/aryaman/.m2/repository/org/jfree/jfreechart/1.5.0/jfreechart-1.5.0.jar:/home/aryaman/.m2/repository/org/jhotdraw/jhotdraw/7.6.0/jhotdraw-7.6.0.jar:/home/aryaman/.m2/repository/net/imagej/imagej-ui-awt/0.3.1/imagej-ui-awt-0.3.1.jar:/home/aryaman/.m2/repository/org/scijava/scijava-io-http/0.2.0/scijava-io-http-0.2.0.jar:/home/aryaman/.m2/repository/com/squareup/okhttp3/okhttp/3.6.0/okhttp-3.6.0.jar:/home/aryaman/.m2/repository/com/squareup/okio/okio/1.11.0/okio-1.11.0.jar:/home/aryaman/.m2/repository/org/scijava/scijava-plugins-commands/0.2.3/scijava-plugins-commands-0.2.3.jar:/home/aryaman/.m2/repository/org/scijava/scijava-plugins-io-table/0.2.0/scijava-plugins-io-table-0.2.0.jar:/home/aryaman/.m2/repository/org/scijava/scijava-plugins-platforms/0.3.1/scijava-plugins-platforms-0.3.1.jar:/home/aryaman/.m2/repository/org/scijava/scijava-plugins-text-markdown/0.1.3/scijava-plugins-text-markdown-0.1.3.jar:/home/aryaman/.m2/repository/org/markdownj/markdownj/0.3.0-1.0.2b4/markdownj-0.3.0-1.0.2b4.jar:/home/aryaman/.m2/repository/org/scijava/scijava-plugins-text-plain/0.1.3/scijava-plugins-text-plain-0.1.3.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-lang3/3.10/commons-lang3-3.10.jar:/home/aryaman/.m2/repository/net/imagej/ij/1.52p/ij-1.52p.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-ij/2.0.0-beta-46/imglib2-ij-2.0.0-beta-46.jar:/home/aryaman/.m2/repository/org/janelia/saalfeldlab/n5/2.1.3/n5-2.1.3.jar:/home/aryaman/.m2/repository/org/tukaani/xz/1.8/xz-1.8.jar:/home/aryaman/.m2/repository/org/lz4/lz4-java/1.7.1/lz4-java-1.7.1.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-compress/1.20/commons-compress-1.20.jar:/home/aryaman/.m2/repository/org/janelia/saalfeldlab/n5-imglib2/3.4.1/n5-imglib2-3.4.1.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-label-multisets/0.9.0/imglib2-label-multisets-0.9.0.jar:/home/aryaman/Repositories/scenery-insitu/target/test-classes:/home/aryaman/Repositories/scenery-insitu/target/classes:/home/aryaman/.m2/repository/org/slf4j/slf4j-api/1.7.30/slf4j-api-1.7.30.jar:/home/aryaman/.m2/repository/org/joml/joml/1.9.24/joml-1.9.24.jar:/home/aryaman/.m2/repository/net/clearcontrol/coremem/0.4.8/coremem-0.4.8.jar:/home/aryaman/.m2/repository/org/scijava/script-editor/0.5.6/script-editor-0.5.6.jar:/home/aryaman/.m2/repository/org/janelia/saalfeldlab/n5/2.2.0/n5-2.2.0.jar:/home/aryaman/.m2/repository/org/janelia/saalfeldlab/n5-imglib2/3.5.0/n5-imglib2-3.5.0.jar:/home/aryaman/Repositories/scenery-insitu/mpi.jar:/home/aryaman/Repositories/insituclient/target/test-classes:/home/aryaman/Repositories/insituclient/target/classes:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-jdk8/1.3.60/kotlin-stdlib-jdk8-1.3.60.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib/1.3.60/kotlin-stdlib-1.3.60.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-common/1.3.60/kotlin-stdlib-common-1.3.60.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-jdk7/1.3.60/kotlin-stdlib-jdk7-1.3.60.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-reflect/1.3.60/kotlin-reflect-1.3.60.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlinx/kotlinx-coroutines-core/1.1.1/kotlinx-coroutines-core-1.1.1.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlinx/kotlinx-coroutines-core-common/1.1.1/kotlinx-coroutines-core-common-1.1.1.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-junit/1.3.60/kotlin-test-junit-1.3.60.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-annotations-common/1.3.60/kotlin-test-annotations-common-1.3.60.jar:/home/aryaman/.m2/repository/junit/junit/4.12/junit-4.12.jar:/home/aryaman/.m2/repository/org/msgpack/msgpack-core/0.8.20/msgpack-core-0.8.20.jar:/home/aryaman/.m2/repository/org/msgpack/jackson-dataformat-msgpack/0.8.20/jackson-dataformat-msgpack-0.8.20.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/core/jackson-databind/2.9.9.3/jackson-databind-2.9.9.3.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/core/jackson-annotations/2.9.0/jackson-annotations-2.9.0.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/core/jackson-core/2.9.9/jackson-core-2.9.9.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2/5.6.3/imglib2-5.6.3.jar:/home/aryaman/.m2/repository/org/scijava/minimaven/2.2.0/minimaven-2.2.0.jar:/home/aryaman/.m2/repository/com/fifesoft/rsyntaxtextarea/2.6.1/rsyntaxtextarea-2.6.1.jar:/home/aryaman/.m2/repository/com/fifesoft/languagesupport/2.6.0/languagesupport-2.6.0.jar:/home/aryaman/.m2/repository/com/fifesoft/autocomplete/2.6.0/autocomplete-2.6.0.jar:/home/aryaman/.m2/repository/org/scijava/batch-processor/0.3.0/batch-processor-0.3.0.jar:/home/aryaman/.m2/repository/org/scijava/scijava-table/0.2.0/scijava-table-0.2.0.jar:/home/aryaman/.m2/repository/org/scijava/scijava-ui-swing/0.11.0/scijava-ui-swing-0.11.0.jar:/home/aryaman/.m2/repository/org/scijava/scijava-ui-awt/0.1.6/scijava-ui-awt-0.1.6.jar:/home/aryaman/.m2/repository/org/scijava/scijava-search/0.5.0/scijava-search-0.5.0.jar:/home/aryaman/.m2/repository/com/google/code/gson/gson/2.8.5/gson-2.8.5.jar:/home/aryaman/.m2/repository/org/yaml/snakeyaml/1.23/snakeyaml-1.23.jar:/home/aryaman/.m2/repository/net/imagej/imagej-common/0.31.0/imagej-common-0.31.0.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-roi/0.7.0/imglib2-roi-0.7.0.jar:/home/aryaman/.m2/repository/com/google/guava/guava/20.0/guava-20.0.jar:/home/aryaman/.m2/repository/org/javassist/javassist/3.21.0-GA/javassist-3.21.0-GA.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-algorithm/0.11.1/imglib2-algorithm-0.11.1.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-realtransform/2.2.1/imglib2-realtransform-2.2.1.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-ui/2.0.0/imglib2-ui-2.0.0.jar:/home/aryaman/.m2/repository/sc/fiji/bigdataviewer-core/8.0.0/bigdataviewer-core-8.0.0.jar:/home/aryaman/.m2/repository/sc/fiji/bigdataviewer-vistools/1.0.0-beta-18/bigdataviewer-vistools-1.0.0-beta-18.jar:/home/aryaman/.m2/repository/sc/fiji/spim_data/2.2.2/spim_data-2.2.2.jar:/home/aryaman/.m2/repository/org/lz4/lz4-java/1.5.0/lz4-java-1.5.0.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-compress/1.19/commons-compress-1.19.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-label-multisets/0.8.1/imglib2-label-multisets-0.8.1.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test/1.3.60/kotlin-test-1.3.60.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-common/1.3.60/kotlin-test-common-1.3.60.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp-presets/ffmpeg/4.1-1.4.4/ffmpeg-4.1-1.4.4.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp/1.4.4/javacpp-1.4.4.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp-presets/ffmpeg/4.1-1.4.4/ffmpeg-4.1-1.4.4-linux-x86.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp-presets/ffmpeg/4.1-1.4.4/ffmpeg-4.1-1.4.4-linux-x86_64.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp-presets/ffmpeg/4.1-1.4.4/ffmpeg-4.1-1.4.4-macosx-x86_64.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp-presets/ffmpeg/4.1-1.4.4/ffmpeg-4.1-1.4.4-windows-x86_64.jar";
-//                                      "/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.base:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.compiler:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.datatransfer:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.desktop:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.instrument:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.logging:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.management:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.management.rmi:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.naming:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.net.http:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.prefs:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.rmi:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.scripting:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.se:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.security.jgss:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.security.sasl:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.smartcardio:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.sql:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.sql.rowset:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.transaction.xa:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.xml:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.xml.crypto:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.accessibility:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.aot:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.attach:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.charsets:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.compiler:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.crypto.cryptoki:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.crypto.ec:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.dynalink:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.editpad:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.hotspot.agent:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.httpserver:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.internal.ed:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.internal.jvmstat:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.internal.le:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.internal.opt:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.internal.vm.ci:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.internal.vm.compiler:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.internal.vm.compiler.management:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jartool:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.javadoc:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jcmd:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jconsole:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jdeps:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jdi:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jdwp.agent:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jfr:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jlink:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jshell:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jsobject:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jstatd:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.localedata:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.management:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.management.agent:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.management.jfr:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.naming.dns:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.naming.rmi:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.net:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.pack:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.rmic:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.scripting.nashorn:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.scripting.nashorn.shell:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.sctp:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.security.auth:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.security.jgss:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.unsupported:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.unsupported.desktop:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.xml.dom:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.zipfs:/home/aryaman/Repositories/scenery/target/test-classes:/home/aryaman/Repositories/scenery/target/classes:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-jdk8/1.4-M1/kotlin-stdlib-jdk8-1.4-M1.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib/1.4-M1/kotlin-stdlib-1.4-M1.jar:/home/aryaman/.m2/repository/org/jetbrains/annotations/13.0/annotations-13.0.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-jdk7/1.4-M1/kotlin-stdlib-jdk7-1.4-M1.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-reflect/1.4-M1/kotlin-reflect-1.4-M1.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlinx/kotlinx-coroutines-core/1.3.5-1.4-M1/kotlinx-coroutines-core-1.3.5-1.4-M1.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-common/1.4-M1/kotlin-stdlib-common-1.4-M1.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-linux-amd64.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-linux-i586.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-macosx-universal.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-windows-amd64.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-linux-amd64.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-linux-i586.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-macosx-universal.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-windows-amd64.jar:/home/aryaman/.m2/repository/junit/junit/4.13/junit-4.13.jar:/home/aryaman/.m2/repository/org/hamcrest/hamcrest-core/1.3/hamcrest-core-1.3.jar:/home/aryaman/.m2/repository/org/slf4j/slf4j-api/1.7.25/slf4j-api-1.7.25.jar:/home/aryaman/.m2/repository/org/slf4j/slf4j-simple/1.7.25/slf4j-simple-1.7.25.jar:/home/aryaman/.m2/repository/net/clearvolume/cleargl/2.2.10/cleargl-2.2.10.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt-main/2.3.2/gluegen-rt-main-2.3.2.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-android-aarch64.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-android-armv6.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-linux-armv6.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-linux-armv6hf.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-solaris-amd64.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-solaris-i586.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-windows-i586.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all-main/2.3.2/jogl-all-main-2.3.2.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-android-aarch64.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-android-armv6.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-linux-armv6.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-linux-armv6hf.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-solaris-amd64.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-solaris-i586.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-windows-i586.jar:/home/aryaman/.m2/repository/org/joml/joml/1.9.22/joml-1.9.22.jar:/home/aryaman/.m2/repository/com/github/scenerygraphics/vector/958f2e6/vector-958f2e6.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2/5.9.2/imglib2-5.9.2.jar:/home/aryaman/.m2/repository/net/clearcontrol/coremem/0.4.6/coremem-0.4.6.jar:/home/aryaman/.m2/repository/com/nativelibs4java/bridj/0.7.0/bridj-0.7.0.jar:/home/aryaman/.m2/repository/net/java/jinput/jinput/2.0.9/jinput-2.0.9.jar:/home/aryaman/.m2/repository/net/java/jinput/jinput/2.0.9/jinput-2.0.9-natives-all.jar:/home/aryaman/.m2/repository/org/scijava/scijava-common/2.83.2/scijava-common-2.83.2.jar:/home/aryaman/.m2/repository/org/scijava/parsington/2.0.0/parsington-2.0.0.jar:/home/aryaman/.m2/repository/org/bushe/eventbus/1.4/eventbus-1.4.jar:/home/aryaman/.m2/repository/org/scijava/script-editor/0.5.3/script-editor-0.5.3.jar:/home/aryaman/.m2/repository/org/scijava/scripting-java/0.4.1/scripting-java-0.4.1.jar:/home/aryaman/.m2/repository/org/scijava/minimaven/2.2.2/minimaven-2.2.2.jar:/home/aryaman/.m2/repository/com/fifesoft/rsyntaxtextarea/3.1.1/rsyntaxtextarea-3.1.1.jar:/home/aryaman/.m2/repository/com/fifesoft/languagesupport/3.1.0/languagesupport-3.1.0.jar:/home/aryaman/.m2/repository/com/fifesoft/autocomplete/3.1.0/autocomplete-3.1.0.jar:/home/aryaman/.m2/repository/org/mozilla/rhino/1.7.6/rhino-1.7.6.jar:/home/aryaman/.m2/repository/com/miglayout/miglayout/3.7.4/miglayout-3.7.4-swing.jar:/home/aryaman/.m2/repository/org/scijava/batch-processor/0.4.1/batch-processor-0.4.1.jar:/home/aryaman/.m2/repository/org/scijava/scijava-table/0.5.0/scijava-table-0.5.0.jar:/home/aryaman/.m2/repository/org/scijava/scijava-ui-swing/0.13.2/scijava-ui-swing-0.13.2.jar:/home/aryaman/.m2/repository/org/scijava/swing-checkbox-tree/1.0.2/swing-checkbox-tree-1.0.2.jar:/home/aryaman/.m2/repository/net/sourceforge/jdatepicker/jdatepicker/1.3.2/jdatepicker-1.3.2.jar:/home/aryaman/.m2/repository/org/scijava/scijava-search/0.7.0/scijava-search-0.7.0.jar:/home/aryaman/.m2/repository/org/ocpsoft/prettytime/prettytime/4.0.1.Final/prettytime-4.0.1.Final.jar:/home/aryaman/.m2/repository/commons-io/commons-io/2.6/commons-io-2.6.jar:/home/aryaman/.m2/repository/org/scijava/ui-behaviour/2.0.1/ui-behaviour-2.0.1.jar:/home/aryaman/.m2/repository/com/google/code/gson/gson/2.8.6/gson-2.8.6.jar:/home/aryaman/.m2/repository/org/yaml/snakeyaml/1.26/snakeyaml-1.26.jar:/home/aryaman/.m2/repository/net/sf/trove4j/trove4j/3.0.3/trove4j-3.0.3.jar:/home/aryaman/.m2/repository/org/scijava/scripting-javascript/0.5.0/scripting-javascript-0.5.0.jar:/home/aryaman/.m2/repository/org/scijava/scripting-jython/1.0.0/scripting-jython-1.0.0.jar:/home/aryaman/.m2/repository/org/python/jython-slim/2.7.2/jython-slim-2.7.2.jar:/home/aryaman/.m2/repository/org/antlr/antlr/3.5.2/antlr-3.5.2.jar:/home/aryaman/.m2/repository/javax/servlet/javax.servlet-api/3.1.0/javax.servlet-api-3.1.0.jar:/home/aryaman/.m2/repository/org/antlr/antlr-runtime/3.5.2/antlr-runtime-3.5.2.jar:/home/aryaman/.m2/repository/org/bouncycastle/bcpkix-jdk15on/1.62/bcpkix-jdk15on-1.62.jar:/home/aryaman/.m2/repository/org/bouncycastle/bcprov-jdk15on/1.62/bcprov-jdk15on-1.62.jar:/home/aryaman/.m2/repository/org/ow2/asm/asm/7.1/asm-7.1.jar:/home/aryaman/.m2/repository/org/ow2/asm/asm-commons/7.1/asm-commons-7.1.jar:/home/aryaman/.m2/repository/org/ow2/asm/asm-tree/7.1/asm-tree-7.1.jar:/home/aryaman/.m2/repository/org/ow2/asm/asm-analysis/7.1/asm-analysis-7.1.jar:/home/aryaman/.m2/repository/org/ow2/asm/asm-util/7.1/asm-util-7.1.jar:/home/aryaman/.m2/repository/com/google/guava/failureaccess/1.0.1/failureaccess-1.0.1.jar:/home/aryaman/.m2/repository/com/ibm/icu/icu4j/59.1/icu4j-59.1.jar:/home/aryaman/.m2/repository/com/carrotsearch/java-sizeof/0.0.5/java-sizeof-0.0.5.jar:/home/aryaman/.m2/repository/com/github/jnr/jffi/1.2.20/jffi-1.2.20.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-netdb/1.1.6/jnr-netdb-1.1.6.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-ffi/2.1.10/jnr-ffi-2.1.10.jar:/home/aryaman/.m2/repository/com/github/jnr/jffi/1.2.19/jffi-1.2.19-native.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-a64asm/1.0.0/jnr-a64asm-1.0.0.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-x86asm/1.0.2/jnr-x86asm-1.0.2.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-posix/3.0.50/jnr-posix-3.0.50.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-constants/0.9.12/jnr-constants-0.9.12.jar:/home/aryaman/.m2/repository/jline/jline/2.14.5/jline-2.14.5.jar:/home/aryaman/.m2/repository/io/netty/netty-buffer/4.1.45.Final/netty-buffer-4.1.45.Final.jar:/home/aryaman/.m2/repository/io/netty/netty-codec/4.1.45.Final/netty-codec-4.1.45.Final.jar:/home/aryaman/.m2/repository/io/netty/netty-common/4.1.45.Final/netty-common-4.1.45.Final.jar:/home/aryaman/.m2/repository/io/netty/netty-handler/4.1.45.Final/netty-handler-4.1.45.Final.jar:/home/aryaman/.m2/repository/io/netty/netty-resolver/4.1.45.Final/netty-resolver-4.1.45.Final.jar:/home/aryaman/.m2/repository/io/netty/netty-transport/4.1.45.Final/netty-transport-4.1.45.Final.jar:/home/aryaman/.m2/repository/org/apache/ant/ant/1.9.7/ant-1.9.7.jar:/home/aryaman/.m2/repository/org/apache/ant/ant-launcher/1.9.7/ant-launcher-1.9.7.jar:/home/aryaman/.m2/repository/net/java/dev/jna/jna/4.5.2/jna-4.5.2.jar:/home/aryaman/.m2/repository/net/java/dev/jna/jna-platform/4.5.2/jna-platform-4.5.2.jar:/home/aryaman/.m2/repository/org/jocl/jocl/2.0.1/jocl-2.0.1.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl/3.2.3/lwjgl-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-glfw/3.2.3/lwjgl-glfw-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-jemalloc/3.2.3/lwjgl-jemalloc-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-vulkan/3.2.3/lwjgl-vulkan-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-opengl/3.2.3/lwjgl-opengl-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-openvr/3.2.3/lwjgl-openvr-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-xxhash/3.2.3/lwjgl-xxhash-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl/3.2.3/lwjgl-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-opengl/3.2.3/lwjgl-opengl-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-glfw/3.2.3/lwjgl-glfw-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-jemalloc/3.2.3/lwjgl-jemalloc-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-openvr/3.2.3/lwjgl-openvr-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-xxhash/3.2.3/lwjgl-xxhash-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-remotery/3.2.3/lwjgl-remotery-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-remotery/3.2.3/lwjgl-remotery-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/core/jackson-databind/2.9.10.1/jackson-databind-2.9.10.1.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/core/jackson-annotations/2.9.10/jackson-annotations-2.9.10.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/core/jackson-core/2.9.10/jackson-core-2.9.10.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/module/jackson-module-kotlin/2.9.9/jackson-module-kotlin-2.9.9.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/dataformat/jackson-dataformat-yaml/2.9.9/jackson-dataformat-yaml-2.9.9.jar:/home/aryaman/.m2/repository/graphics/scenery/spirvcrossj/0.7.0-1.1.106.0/spirvcrossj-0.7.0-1.1.106.0.jar:/home/aryaman/.m2/repository/graphics/scenery/spirvcrossj/0.7.0-1.1.106.0/spirvcrossj-0.7.0-1.1.106.0-natives-linux.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-junit/1.4-M1/kotlin-test-junit-1.4-M1.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-annotations-common/1.4-M1/kotlin-test-annotations-common-1.4-M1.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test/1.4-M1/kotlin-test-1.4-M1.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-common/1.4-M1/kotlin-test-common-1.4-M1.jar:/home/aryaman/.m2/repository/org/zeromq/jeromq/0.4.3/jeromq-0.4.3.jar:/home/aryaman/.m2/repository/org/zeromq/jnacl/0.1.0/jnacl-0.1.0.jar:/home/aryaman/.m2/repository/com/esotericsoftware/kryo/4.0.2/kryo-4.0.2.jar:/home/aryaman/.m2/repository/com/esotericsoftware/reflectasm/1.11.3/reflectasm-1.11.3.jar:/home/aryaman/.m2/repository/com/esotericsoftware/minlog/1.3.0/minlog-1.3.0.jar:/home/aryaman/.m2/repository/org/objenesis/objenesis/2.5.1/objenesis-2.5.1.jar:/home/aryaman/.m2/repository/org/msgpack/msgpack-core/0.8.16/msgpack-core-0.8.16.jar:/home/aryaman/.m2/repository/org/msgpack/jackson-dataformat-msgpack/0.8.16/jackson-dataformat-msgpack-0.8.16.jar:/home/aryaman/.m2/repository/graphics/scenery/jvrpn/1.1.0/jvrpn-1.1.0.jar:/home/aryaman/.m2/repository/graphics/scenery/jvrpn/1.1.0/jvrpn-1.1.0-natives-linux.jar:/home/aryaman/.m2/repository/io/scif/scifio/0.40.0/scifio-0.40.0.jar:/home/aryaman/.m2/repository/io/scif/scifio-jai-imageio/1.1.1/scifio-jai-imageio-1.1.1.jar:/home/aryaman/.m2/repository/net/imagej/imagej-common/0.33.0/imagej-common-0.33.0.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-roi/0.10.2/imglib2-roi-0.10.2.jar:/home/aryaman/.m2/repository/edu/ucar/udunits/4.3.18/udunits-4.3.18.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-cache/1.0.0-beta-13/imglib2-cache-1.0.0-beta-13.jar:/home/aryaman/.m2/repository/com/github/ben-manes/caffeine/caffeine/2.4.0/caffeine-2.4.0.jar:/home/aryaman/.m2/repository/org/bytedeco/ffmpeg/4.2.1-1.5.2/ffmpeg-4.2.1-1.5.2.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp/1.5.2/javacpp-1.5.2.jar:/home/aryaman/.m2/repository/org/bytedeco/ffmpeg/4.2.1-1.5.2/ffmpeg-4.2.1-1.5.2-linux-x86_64.jar:/home/aryaman/.m2/repository/org/bytedeco/ffmpeg/4.2.1-1.5.2/ffmpeg-4.2.1-1.5.2-macosx-x86_64.jar:/home/aryaman/.m2/repository/org/bytedeco/ffmpeg/4.2.1-1.5.2/ffmpeg-4.2.1-1.5.2-windows-x86_64.jar:/home/aryaman/.m2/repository/org/reflections/reflections/0.9.11/reflections-0.9.11.jar:/home/aryaman/.m2/repository/com/google/guava/guava/21.0/guava-21.0.jar:/home/aryaman/.m2/repository/org/javassist/javassist/3.27.0-GA/javassist-3.27.0-GA.jar:/home/aryaman/.m2/repository/sc/fiji/bigvolumeviewer/0.1.8/bigvolumeviewer-0.1.8.jar:/home/aryaman/.m2/repository/commons-lang/commons-lang/2.6/commons-lang-2.6.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-algorithm/0.11.2/imglib2-algorithm-0.11.2.jar:/home/aryaman/.m2/repository/gov/nist/math/jama/1.0.3/jama-1.0.3.jar:/home/aryaman/.m2/repository/org/ojalgo/ojalgo/45.1.1/ojalgo-45.1.1.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-realtransform/3.0.0/imglib2-realtransform-3.0.0.jar:/home/aryaman/.m2/repository/jitk/jitk-tps/3.0.1/jitk-tps-3.0.1.jar:/home/aryaman/.m2/repository/com/googlecode/efficient-java-matrix-library/ejml/0.24/ejml-0.24.jar:/home/aryaman/.m2/repository/log4j/log4j/1.2.17/log4j-1.2.17.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-ui/2.0.1/imglib2-ui-2.0.1.jar:/home/aryaman/.m2/repository/org/antlr/ST4/4.0.8/ST4-4.0.8.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-math3/3.6.1/commons-math3-3.6.1.jar:/home/aryaman/.m2/repository/org/jdom/jdom2/2.0.6/jdom2-2.0.6.jar:/home/aryaman/.m2/repository/sc/fiji/bigdataviewer-core/9.0.2/bigdataviewer-core-9.0.2.jar:/home/aryaman/.m2/repository/cisd/jhdf5/14.12.6/jhdf5-14.12.6.jar:/home/aryaman/.m2/repository/org/scijava/scijava-listeners/1.0.0-beta-2/scijava-listeners-1.0.0-beta-2.jar:/home/aryaman/.m2/repository/sc/fiji/bigdataviewer-vistools/1.0.0-beta-20/bigdataviewer-vistools-1.0.0-beta-20.jar:/home/aryaman/.m2/repository/sc/fiji/spim_data/2.2.4/spim_data-2.2.4.jar:/home/aryaman/.m2/repository/org/lwjglx/lwjgl3-awt/0.1.7/lwjgl3-awt-0.1.7.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-jawt/3.2.3/lwjgl-jawt-3.2.3.jar:/home/aryaman/Repositories/scenery/mpi.jar:/home/aryaman/.m2/repository/net/imagej/imagej/2.0.0-rc-72/imagej-2.0.0-rc-72.jar:/home/aryaman/.m2/repository/net/imagej/imagej-notebook/0.7.1/imagej-notebook-0.7.1.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-text/1.8/commons-text-1.8.jar:/home/aryaman/.m2/repository/org/knowm/xchart/xchart/3.5.4/xchart-3.5.4.jar:/home/aryaman/.m2/repository/de/erichseifert/vectorgraphics2d/VectorGraphics2D/0.13/VectorGraphics2D-0.13.jar:/home/aryaman/.m2/repository/net/imagej/imagej-ops/0.45.5/imagej-ops-0.45.5.jar:/home/aryaman/.m2/repository/net/imagej/imagej-mesh/0.8.0/imagej-mesh-0.8.0.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-algorithm-fft/0.2.0/imglib2-algorithm-fft-0.2.0.jar:/home/aryaman/.m2/repository/edu/mines/mines-jtk/20151125/mines-jtk-20151125.jar:/home/aryaman/.m2/repository/net/imagej/imagej-updater/0.10.4/imagej-updater-0.10.4.jar:/home/aryaman/.m2/repository/net/imagej/imagej-plugins-batch/0.1.1/imagej-plugins-batch-0.1.1.jar:/home/aryaman/.m2/repository/net/imagej/imagej-deprecated/0.1.4/imagej-deprecated-0.1.4.jar:/home/aryaman/.m2/repository/net/imagej/imagej-plugins-commands/0.8.1/imagej-plugins-commands-0.8.1.jar:/home/aryaman/.m2/repository/com/github/sbridges/object-inspector/object-inspector/0.1/object-inspector-0.1.jar:/home/aryaman/.m2/repository/net/iharder/base64/2.3.8/base64-2.3.8.jar:/home/aryaman/.m2/repository/net/imagej/imagej-plugins-tools/0.3.1/imagej-plugins-tools-0.3.1.jar:/home/aryaman/.m2/repository/net/imagej/imagej-plugins-uploader-ssh/0.3.2/imagej-plugins-uploader-ssh-0.3.2.jar:/home/aryaman/.m2/repository/com/jcraft/jsch/0.1.55/jsch-0.1.55.jar:/home/aryaman/.m2/repository/net/imagej/imagej-plugins-uploader-webdav/0.3.2/imagej-plugins-uploader-webdav-0.3.2.jar:/home/aryaman/.m2/repository/org/apache/jackrabbit/jackrabbit-webdav/2.21.1/jackrabbit-webdav-2.21.1.jar:/home/aryaman/.m2/repository/org/apache/httpcomponents/httpclient/4.5.12/httpclient-4.5.12.jar:/home/aryaman/.m2/repository/commons-codec/commons-codec/1.14/commons-codec-1.14.jar:/home/aryaman/.m2/repository/org/apache/httpcomponents/httpcore/4.4.13/httpcore-4.4.13.jar:/home/aryaman/.m2/repository/net/imagej/imagej-scripting/0.8.2/imagej-scripting-0.8.2.jar:/home/aryaman/.m2/repository/org/scijava/scripting-beanshell/0.3.4/scripting-beanshell-0.3.4.jar:/home/aryaman/.m2/repository/org/apache-extras/beanshell/bsh/2.0b6/bsh-2.0b6.jar:/home/aryaman/.m2/repository/org/scijava/scripting-clojure/0.1.6/scripting-clojure-0.1.6.jar:/home/aryaman/.m2/repository/org/clojure/clojure/1.8.0/clojure-1.8.0.jar:/home/aryaman/.m2/repository/org/scijava/scripting-groovy/0.3.0/scripting-groovy-0.3.0.jar:/home/aryaman/.m2/repository/org/codehaus/groovy/groovy/3.0.4/groovy-3.0.4.jar:/home/aryaman/.m2/repository/org/codehaus/groovy/groovy-dateutil/3.0.4/groovy-dateutil-3.0.4.jar:/home/aryaman/.m2/repository/org/codehaus/groovy/groovy-json/3.0.4/groovy-json-3.0.4.jar:/home/aryaman/.m2/repository/org/codehaus/groovy/groovy-swing/3.0.4/groovy-swing-3.0.4.jar:/home/aryaman/.m2/repository/org/codehaus/groovy/groovy-templates/3.0.4/groovy-templates-3.0.4.jar:/home/aryaman/.m2/repository/org/codehaus/groovy/groovy-xml/3.0.4/groovy-xml-3.0.4.jar:/home/aryaman/.m2/repository/org/apache/ivy/ivy/2.4.0/ivy-2.4.0.jar:/home/aryaman/.m2/repository/org/codehaus/gpars/gpars/1.2.1/gpars-1.2.1.jar:/home/aryaman/.m2/repository/org/multiverse/multiverse-core/0.7.0/multiverse-core-0.7.0.jar:/home/aryaman/.m2/repository/org/codehaus/jsr166-mirror/jsr166y/1.7.0/jsr166y-1.7.0.jar:/home/aryaman/.m2/repository/org/scijava/scripting-jruby/0.3.1/scripting-jruby-0.3.1.jar:/home/aryaman/.m2/repository/org/jruby/jruby-core/9.1.17.0/jruby-core-9.1.17.0.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-enxio/0.16/jnr-enxio-0.16.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-unixsocket/0.17/jnr-unixsocket-0.17.jar:/home/aryaman/.m2/repository/org/jruby/joni/joni/2.1.15/joni-2.1.15.jar:/home/aryaman/.m2/repository/org/jruby/extras/bytelist/1.0.15/bytelist-1.0.15.jar:/home/aryaman/.m2/repository/org/jruby/jcodings/jcodings/1.0.27/jcodings-1.0.27.jar:/home/aryaman/.m2/repository/org/jruby/dirgra/0.3/dirgra-0.3.jar:/home/aryaman/.m2/repository/com/headius/invokebinder/1.10/invokebinder-1.10.jar:/home/aryaman/.m2/repository/com/headius/options/1.4/options-1.4.jar:/home/aryaman/.m2/repository/com/headius/unsafe-fences/1.0/unsafe-fences-1.0.jar:/home/aryaman/.m2/repository/com/jcraft/jzlib/1.1.3/jzlib-1.1.3.jar:/home/aryaman/.m2/repository/com/martiansoftware/nailgun-server/0.9.1/nailgun-server-0.9.1.jar:/home/aryaman/.m2/repository/joda-time/joda-time/2.10.6/joda-time-2.10.6.jar:/home/aryaman/.m2/repository/com/headius/modulator/1.0/modulator-1.0.jar:/home/aryaman/.m2/repository/org/jruby/jruby-stdlib/9.1.17.0/jruby-stdlib-9.1.17.0.jar:/home/aryaman/.m2/repository/org/scijava/scripting-renjin/0.2.3/scripting-renjin-0.2.3.jar:/home/aryaman/.m2/repository/org/renjin/renjin-script-engine/0.8.1906/renjin-script-engine-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/renjin-core/0.8.1906/renjin-core-0.8.1906.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-math/2.2/commons-math-2.2.jar:/home/aryaman/.m2/repository/org/renjin/gcc-runtime/0.8.1906/gcc-runtime-0.8.1906.jar:/home/aryaman/.m2/repository/org/netlib/netlib-java/0.9.3-renjin-patched-2/netlib-java-0.9.3-renjin-patched-2.jar:/home/aryaman/.m2/repository/org/netlib/lapack/0.8/lapack-0.8.jar:/home/aryaman/.m2/repository/org/netlib/f2jutil/0.8/f2jutil-0.8.jar:/home/aryaman/.m2/repository/org/netlib/xerbla/0.8/xerbla-0.8.jar:/home/aryaman/.m2/repository/org/netlib/blas/0.8/blas-0.8.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-vfs2/2.0/commons-vfs2-2.0.jar:/home/aryaman/.m2/repository/commons-logging/commons-logging/1.2/commons-logging-1.2.jar:/home/aryaman/.m2/repository/org/apache/maven/scm/maven-scm-api/1.4/maven-scm-api-1.4.jar:/home/aryaman/.m2/repository/org/codehaus/plexus/plexus-utils/1.5.6/plexus-utils-1.5.6.jar:/home/aryaman/.m2/repository/org/apache/maven/scm/maven-scm-provider-svnexe/1.4/maven-scm-provider-svnexe-1.4.jar:/home/aryaman/.m2/repository/org/apache/maven/scm/maven-scm-provider-svn-commons/1.4/maven-scm-provider-svn-commons-1.4.jar:/home/aryaman/.m2/repository/regexp/regexp/1.3/regexp-1.3.jar:/home/aryaman/.m2/repository/edu/emory/mathcs/jtransforms/2.4/jtransforms-2.4.jar:/home/aryaman/.m2/repository/net/sf/jung/jung-api/2.0.1/jung-api-2.0.1.jar:/home/aryaman/.m2/repository/net/sourceforge/collections/collections-generic/4.01/collections-generic-4.01.jar:/home/aryaman/.m2/repository/net/sf/jung/jung-graph-impl/2.0.1/jung-graph-impl-2.0.1.jar:/home/aryaman/.m2/repository/com/sun/codemodel/codemodel/2.6/codemodel-2.6.jar:/home/aryaman/.m2/repository/org/renjin/stats/0.8.1906/stats-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/renjin-appl/0.8.1906/renjin-appl-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/renjin-gnur-runtime/0.8.1906/renjin-gnur-runtime-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/methods/0.8.1906/methods-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/datasets/0.8.1906/datasets-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/utils/0.8.1906/utils-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/grDevices/0.8.1906/grDevices-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/graphics/0.8.1906/graphics-0.8.1906.jar:/home/aryaman/.m2/repository/org/scijava/scripting-scala/0.2.2/scripting-scala-0.2.2.jar:/home/aryaman/.m2/repository/org/scala-lang/scala-compiler/2.12.1/scala-compiler-2.12.1.jar:/home/aryaman/.m2/repository/org/scala-lang/scala-library/2.12.1/scala-library-2.12.1.jar:/home/aryaman/.m2/repository/org/scala-lang/scala-reflect/2.12.1/scala-reflect-2.12.1.jar:/home/aryaman/.m2/repository/org/scala-lang/modules/scala-xml_2.12/1.0.6/scala-xml_2.12-1.0.6.jar:/home/aryaman/.m2/repository/net/imagej/imagej-ui-swing/0.23.2/imagej-ui-swing-0.23.2.jar:/home/aryaman/.m2/repository/org/scijava/scijava-ui-awt/0.1.7/scijava-ui-awt-0.1.7.jar:/home/aryaman/.m2/repository/com/miglayout/miglayout-swing/5.2/miglayout-swing-5.2.jar:/home/aryaman/.m2/repository/com/miglayout/miglayout-core/5.2/miglayout-core-5.2.jar:/home/aryaman/.m2/repository/org/jfree/jfreechart/1.5.0/jfreechart-1.5.0.jar:/home/aryaman/.m2/repository/org/jhotdraw/jhotdraw/7.6.0/jhotdraw-7.6.0.jar:/home/aryaman/.m2/repository/net/imagej/imagej-ui-awt/0.3.1/imagej-ui-awt-0.3.1.jar:/home/aryaman/.m2/repository/org/scijava/scijava-io-http/0.2.0/scijava-io-http-0.2.0.jar:/home/aryaman/.m2/repository/com/squareup/okhttp3/okhttp/3.6.0/okhttp-3.6.0.jar:/home/aryaman/.m2/repository/com/squareup/okio/okio/1.11.0/okio-1.11.0.jar:/home/aryaman/.m2/repository/org/scijava/scijava-plugins-commands/0.2.3/scijava-plugins-commands-0.2.3.jar:/home/aryaman/.m2/repository/org/scijava/scijava-plugins-io-table/0.2.0/scijava-plugins-io-table-0.2.0.jar:/home/aryaman/.m2/repository/org/scijava/scijava-plugins-platforms/0.3.1/scijava-plugins-platforms-0.3.1.jar:/home/aryaman/.m2/repository/org/scijava/scijava-plugins-text-markdown/0.1.3/scijava-plugins-text-markdown-0.1.3.jar:/home/aryaman/.m2/repository/org/markdownj/markdownj/0.3.0-1.0.2b4/markdownj-0.3.0-1.0.2b4.jar:/home/aryaman/.m2/repository/org/scijava/scijava-plugins-text-plain/0.1.3/scijava-plugins-text-plain-0.1.3.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-lang3/3.10/commons-lang3-3.10.jar:/home/aryaman/.m2/repository/net/imagej/ij/1.52p/ij-1.52p.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-ij/2.0.0-beta-46/imglib2-ij-2.0.0-beta-46.jar:/home/aryaman/.m2/repository/org/janelia/saalfeldlab/n5/2.1.3/n5-2.1.3.jar:/home/aryaman/.m2/repository/org/tukaani/xz/1.8/xz-1.8.jar:/home/aryaman/.m2/repository/org/lz4/lz4-java/1.7.1/lz4-java-1.7.1.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-compress/1.20/commons-compress-1.20.jar:/home/aryaman/.m2/repository/org/janelia/saalfeldlab/n5-imglib2/3.4.1/n5-imglib2-3.4.1.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-label-multisets/0.9.0/imglib2-label-multisets-0.9.0.jar:/home/aryaman/Repositories/scenery-insitu/target/test-classes:/home/aryaman/Repositories/scenery-insitu/target/classes:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-jdk8/1.3.21/kotlin-stdlib-jdk8-1.3.21.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib/1.3.21/kotlin-stdlib-1.3.21.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-common/1.3.21/kotlin-stdlib-common-1.3.21.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-jdk7/1.3.21/kotlin-stdlib-jdk7-1.3.21.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-reflect/1.3.21/kotlin-reflect-1.3.21.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlinx/kotlinx-coroutines-core/1.1.1/kotlinx-coroutines-core-1.1.1.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlinx/kotlinx-coroutines-core-common/1.1.1/kotlinx-coroutines-core-common-1.1.1.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-junit/1.3.21/kotlin-test-junit-1.3.21.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-annotations-common/1.3.21/kotlin-test-annotations-common-1.3.21.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test/1.3.21/kotlin-test-1.3.21.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-common/1.3.21/kotlin-test-common-1.3.21.jar:/home/aryaman/.m2/repository/junit/junit/4.12/junit-4.12.jar:/home/aryaman/.m2/repository/org/joml/joml/1.9.12/joml-1.9.12.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2/5.6.3/imglib2-5.6.3.jar:/home/aryaman/.m2/repository/org/scijava/scijava-common/2.76.1/scijava-common-2.76.1.jar:/home/aryaman/.m2/repository/org/scijava/parsington/1.0.4/parsington-1.0.4.jar:/home/aryaman/.m2/repository/org/scijava/script-editor/0.4.0/script-editor-0.4.0.jar:/home/aryaman/.m2/repository/org/scijava/minimaven/2.2.1/minimaven-2.2.1.jar:/home/aryaman/.m2/repository/com/fifesoft/rsyntaxtextarea/2.6.1/rsyntaxtextarea-2.6.1.jar:/home/aryaman/.m2/repository/com/fifesoft/languagesupport/2.6.0/languagesupport-2.6.0.jar:/home/aryaman/.m2/repository/com/fifesoft/autocomplete/2.6.1/autocomplete-2.6.1.jar:/home/aryaman/.m2/repository/org/scijava/batch-processor/0.3.0/batch-processor-0.3.0.jar:/home/aryaman/.m2/repository/org/scijava/scijava-table/0.2.0/scijava-table-0.2.0.jar:/home/aryaman/.m2/repository/org/scijava/scijava-ui-swing/0.12.0/scijava-ui-swing-0.12.0.jar:/home/aryaman/.m2/repository/org/scijava/scijava-ui-awt/0.1.6/scijava-ui-awt-0.1.6.jar:/home/aryaman/.m2/repository/org/scijava/scijava-search/0.5.0/scijava-search-0.5.0.jar:/home/aryaman/.m2/repository/org/scijava/ui-behaviour/1.7.4/ui-behaviour-1.7.4.jar:/home/aryaman/.m2/repository/com/google/code/gson/gson/2.8.5/gson-2.8.5.jar:/home/aryaman/.m2/repository/org/yaml/snakeyaml/1.23/snakeyaml-1.23.jar:/home/aryaman/.m2/repository/org/scijava/scripting-javascript/0.4.4/scripting-javascript-0.4.4.jar:/home/aryaman/.m2/repository/org/scijava/scripting-jython/0.4.2/scripting-jython-0.4.2.jar:/home/aryaman/.m2/repository/org/scijava/jython-shaded/2.7.1/jython-shaded-2.7.1.jar:/home/aryaman/.m2/repository/org/ow2/asm/asm/5.0.4/asm-5.0.4.jar:/home/aryaman/.m2/repository/io/scif/scifio/0.37.3/scifio-0.37.3.jar:/home/aryaman/.m2/repository/net/imagej/imagej-common/0.28.0/imagej-common-0.28.0.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-roi/0.7.0/imglib2-roi-0.7.0.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-cache/1.0.0-beta-11/imglib2-cache-1.0.0-beta-11.jar:/home/aryaman/.m2/repository/org/javassist/javassist/3.24.0-GA/javassist-3.24.0-GA.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-algorithm/0.11.1/imglib2-algorithm-0.11.1.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-realtransform/2.1.0/imglib2-realtransform-2.1.0.jar:/home/aryaman/.m2/repository/jitk/jitk-tps/3.0.0/jitk-tps-3.0.0.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-ui/2.0.0/imglib2-ui-2.0.0.jar:/home/aryaman/.m2/repository/sc/fiji/bigdataviewer-core/5.1.0/bigdataviewer-core-5.1.0.jar:/home/aryaman/.m2/repository/sc/fiji/bigdataviewer-vistools/1.0.0-beta-11/bigdataviewer-vistools-1.0.0-beta-11.jar:/home/aryaman/.m2/repository/sc/fiji/spim_data/2.2.2/spim_data-2.2.2.jar:/home/aryaman/.m2/repository/org/lz4/lz4-java/1.5.0/lz4-java-1.5.0.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-compress/1.19/commons-compress-1.19.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-label-multisets/0.8.1/imglib2-label-multisets-0.8.1.jar:/home/aryaman/.m2/repository/org/msgpack/msgpack-core/0.8.20/msgpack-core-0.8.20.jar:/home/aryaman/.m2/repository/org/msgpack/jackson-dataformat-msgpack/0.8.20/jackson-dataformat-msgpack-0.8.20.jar:/home/aryaman/Repositories/insituclient/target/test-classes:/home/aryaman/Repositories/insituclient/target/classes:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-jdk8/1.3.60/kotlin-stdlib-jdk8-1.3.60.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib/1.3.60/kotlin-stdlib-1.3.60.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-common/1.3.60/kotlin-stdlib-common-1.3.60.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-jdk7/1.3.60/kotlin-stdlib-jdk7-1.3.60.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-reflect/1.3.60/kotlin-reflect-1.3.60.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-junit/1.3.60/kotlin-test-junit-1.3.60.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-annotations-common/1.3.60/kotlin-test-annotations-common-1.3.60.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/core/jackson-databind/2.9.9.3/jackson-databind-2.9.9.3.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/core/jackson-annotations/2.9.0/jackson-annotations-2.9.0.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/core/jackson-core/2.9.9/jackson-core-2.9.9.jar:/home/aryaman/.m2/repository/org/scijava/minimaven/2.2.0/minimaven-2.2.0.jar:/home/aryaman/.m2/repository/com/fifesoft/autocomplete/2.6.0/autocomplete-2.6.0.jar:/home/aryaman/.m2/repository/org/scijava/scijava-ui-swing/0.11.0/scijava-ui-swing-0.11.0.jar:/home/aryaman/.m2/repository/net/imagej/imagej-common/0.31.0/imagej-common-0.31.0.jar:/home/aryaman/.m2/repository/com/google/guava/guava/20.0/guava-20.0.jar:/home/aryaman/.m2/repository/org/javassist/javassist/3.21.0-GA/javassist-3.21.0-GA.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-realtransform/2.2.1/imglib2-realtransform-2.2.1.jar:/home/aryaman/.m2/repository/sc/fiji/bigdataviewer-core/8.0.0/bigdataviewer-core-8.0.0.jar:/home/aryaman/.m2/repository/sc/fiji/bigdataviewer-vistools/1.0.0-beta-18/bigdataviewer-vistools-1.0.0-beta-18.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test/1.3.60/kotlin-test-1.3.60.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-common/1.3.60/kotlin-test-common-1.3.60.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp-presets/ffmpeg/4.1-1.4.4/ffmpeg-4.1-1.4.4.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp/1.4.4/javacpp-1.4.4.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp-presets/ffmpeg/4.1-1.4.4/ffmpeg-4.1-1.4.4-linux-x86.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp-presets/ffmpeg/4.1-1.4.4/ffmpeg-4.1-1.4.4-linux-x86_64.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp-presets/ffmpeg/4.1-1.4.4/ffmpeg-4.1-1.4.4-macosx-x86_64.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp-presets/ffmpeg/4.1-1.4.4/ffmpeg-4.1-1.4.4-windows-x86_64.jar";
-                                      //"-Djava.class.path=/home/aryaman/jdk8u242-b08/jre/lib/charsets.jar:/home/aryaman/jdk8u242-b08/jre/lib/ext/cldrdata.jar:/home/aryaman/jdk8u242-b08/jre/lib/ext/dnsns.jar:/home/aryaman/jdk8u242-b08/jre/lib/ext/jaccess.jar:/home/aryaman/jdk8u242-b08/jre/lib/ext/localedata.jar:/home/aryaman/jdk8u242-b08/jre/lib/ext/nashorn.jar:/home/aryaman/jdk8u242-b08/jre/lib/ext/sunec.jar:/home/aryaman/jdk8u242-b08/jre/lib/ext/sunjce_provider.jar:/home/aryaman/jdk8u242-b08/jre/lib/ext/sunpkcs11.jar:/home/aryaman/jdk8u242-b08/jre/lib/ext/zipfs.jar:/home/aryaman/jdk8u242-b08/jre/lib/jce.jar:/home/aryaman/jdk8u242-b08/jre/lib/jsse.jar:/home/aryaman/jdk8u242-b08/jre/lib/management-agent.jar:/home/aryaman/jdk8u242-b08/jre/lib/resources.jar:/home/aryaman/jdk8u242-b08/jre/lib/rt.jar:/home/aryaman/Repositories/scenery/target/test-classes:/home/aryaman/Repositories/scenery/target/classes:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-jdk8/1.3.61/kotlin-stdlib-jdk8-1.3.61.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib/1.3.61/kotlin-stdlib-1.3.61.jar:/home/aryaman/.m2/repository/org/jetbrains/annotations/13.0/annotations-13.0.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-jdk7/1.3.61/kotlin-stdlib-jdk7-1.3.61.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-reflect/1.3.61/kotlin-reflect-1.3.61.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlinx/kotlinx-coroutines-core/1.3.2/kotlinx-coroutines-core-1.3.2.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-common/1.3.50/kotlin-stdlib-common-1.3.50.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-linux-amd64.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-linux-i586.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-macosx-universal.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-windows-amd64.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-linux-amd64.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-linux-i586.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-macosx-universal.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-windows-amd64.jar:/home/aryaman/.m2/repository/junit/junit/4.12/junit-4.12.jar:/home/aryaman/.m2/repository/org/hamcrest/hamcrest-core/1.3/hamcrest-core-1.3.jar:/home/aryaman/.m2/repository/org/slf4j/slf4j-api/1.7.25/slf4j-api-1.7.25.jar:/home/aryaman/.m2/repository/org/slf4j/slf4j-simple/1.7.25/slf4j-simple-1.7.25.jar:/home/aryaman/.m2/repository/net/clearvolume/cleargl/2.2.10/cleargl-2.2.10.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt-main/2.3.2/gluegen-rt-main-2.3.2.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-android-aarch64.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-android-armv6.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-linux-armv6.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-linux-armv6hf.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-solaris-amd64.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-solaris-i586.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-windows-i586.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all-main/2.3.2/jogl-all-main-2.3.2.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-android-aarch64.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-android-armv6.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-linux-armv6.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-linux-armv6hf.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-solaris-amd64.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-solaris-i586.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-windows-i586.jar:/home/aryaman/.m2/repository/net/clearcontrol/coremem/0.4.6/coremem-0.4.6.jar:/home/aryaman/.m2/repository/com/nativelibs4java/bridj/0.7.0/bridj-0.7.0.jar:/home/aryaman/.m2/repository/net/java/jinput/jinput/2.0.9/jinput-2.0.9.jar:/home/aryaman/.m2/repository/net/java/jinput/jinput/2.0.9/jinput-2.0.9-natives-all.jar:/home/aryaman/.m2/repository/org/scijava/scijava-common/2.77.0/scijava-common-2.77.0.jar:/home/aryaman/.m2/repository/org/scijava/parsington/1.0.4/parsington-1.0.4.jar:/home/aryaman/.m2/repository/org/bushe/eventbus/1.4/eventbus-1.4.jar:/home/aryaman/.m2/repository/org/scijava/script-editor/0.5.3/script-editor-0.5.3.jar:/home/aryaman/.m2/repository/org/scijava/scripting-java/0.4.1/scripting-java-0.4.1.jar:/home/aryaman/.m2/repository/org/scijava/minimaven/2.2.1/minimaven-2.2.1.jar:/home/aryaman/.m2/repository/com/fifesoft/rsyntaxtextarea/2.6.1/rsyntaxtextarea-2.6.1.jar:/home/aryaman/.m2/repository/com/fifesoft/languagesupport/2.6.0/languagesupport-2.6.0.jar:/home/aryaman/.m2/repository/com/fifesoft/autocomplete/2.6.1/autocomplete-2.6.1.jar:/home/aryaman/.m2/repository/org/mozilla/rhino/1.7.6/rhino-1.7.6.jar:/home/aryaman/.m2/repository/com/miglayout/miglayout/3.7.4/miglayout-3.7.4-swing.jar:/home/aryaman/.m2/repository/org/scijava/batch-processor/0.3.0/batch-processor-0.3.0.jar:/home/aryaman/.m2/repository/org/scijava/scijava-table/0.4.0/scijava-table-0.4.0.jar:/home/aryaman/.m2/repository/org/scijava/scijava-ui-swing/0.12.0/scijava-ui-swing-0.12.0.jar:/home/aryaman/.m2/repository/org/scijava/scijava-ui-awt/0.1.6/scijava-ui-awt-0.1.6.jar:/home/aryaman/.m2/repository/org/scijava/swing-checkbox-tree/1.0.2/swing-checkbox-tree-1.0.2.jar:/home/aryaman/.m2/repository/com/github/sbridges/object-inspector/object-inspector/0.1/object-inspector-0.1.jar:/home/aryaman/.m2/repository/net/sourceforge/jdatepicker/jdatepicker/1.3.2/jdatepicker-1.3.2.jar:/home/aryaman/.m2/repository/org/scijava/scijava-search/0.6.0/scijava-search-0.6.0.jar:/home/aryaman/.m2/repository/org/ocpsoft/prettytime/prettytime/4.0.1.Final/prettytime-4.0.1.Final.jar:/home/aryaman/.m2/repository/commons-io/commons-io/2.6/commons-io-2.6.jar:/home/aryaman/.m2/repository/org/scijava/ui-behaviour/1.7.4/ui-behaviour-1.7.4.jar:/home/aryaman/.m2/repository/com/google/code/gson/gson/2.8.5/gson-2.8.5.jar:/home/aryaman/.m2/repository/org/yaml/snakeyaml/1.23/snakeyaml-1.23.jar:/home/aryaman/.m2/repository/net/sf/trove4j/trove4j/3.0.3/trove4j-3.0.3.jar:/home/aryaman/.m2/repository/org/scijava/scripting-javascript/0.4.4/scripting-javascript-0.4.4.jar:/home/aryaman/.m2/repository/net/java/dev/jna/jna/4.5.2/jna-4.5.2.jar:/home/aryaman/.m2/repository/net/java/dev/jna/jna-platform/4.5.2/jna-platform-4.5.2.jar:/home/aryaman/.m2/repository/org/jocl/jocl/2.0.1/jocl-2.0.1.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl/3.2.3/lwjgl-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-glfw/3.2.3/lwjgl-glfw-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-jemalloc/3.2.3/lwjgl-jemalloc-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-vulkan/3.2.3/lwjgl-vulkan-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-opengl/3.2.3/lwjgl-opengl-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-openvr/3.2.3/lwjgl-openvr-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-xxhash/3.2.3/lwjgl-xxhash-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl/3.2.3/lwjgl-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-opengl/3.2.3/lwjgl-opengl-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-glfw/3.2.3/lwjgl-glfw-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-jemalloc/3.2.3/lwjgl-jemalloc-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-openvr/3.2.3/lwjgl-openvr-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-xxhash/3.2.3/lwjgl-xxhash-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-remotery/3.2.3/lwjgl-remotery-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-remotery/3.2.3/lwjgl-remotery-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/core/jackson-databind/2.9.10.1/jackson-databind-2.9.10.1.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/core/jackson-annotations/2.9.10/jackson-annotations-2.9.10.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/core/jackson-core/2.9.10/jackson-core-2.9.10.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/module/jackson-module-kotlin/2.9.9/jackson-module-kotlin-2.9.9.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/dataformat/jackson-dataformat-yaml/2.9.9/jackson-dataformat-yaml-2.9.9.jar:/home/aryaman/.m2/repository/graphics/scenery/spirvcrossj/0.7.0-1.1.106.0/spirvcrossj-0.7.0-1.1.106.0.jar:/home/aryaman/.m2/repository/graphics/scenery/spirvcrossj/0.7.0-1.1.106.0/spirvcrossj-0.7.0-1.1.106.0-natives-linux.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-junit/1.3.61/kotlin-test-junit-1.3.61.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-annotations-common/1.3.61/kotlin-test-annotations-common-1.3.61.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test/1.3.61/kotlin-test-1.3.61.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-common/1.3.61/kotlin-test-common-1.3.61.jar:/home/aryaman/.m2/repository/org/zeromq/jeromq/0.4.3/jeromq-0.4.3.jar:/home/aryaman/.m2/repository/org/zeromq/jnacl/0.1.0/jnacl-0.1.0.jar:/home/aryaman/.m2/repository/com/esotericsoftware/kryo/4.0.2/kryo-4.0.2.jar:/home/aryaman/.m2/repository/com/esotericsoftware/reflectasm/1.11.3/reflectasm-1.11.3.jar:/home/aryaman/.m2/repository/org/ow2/asm/asm/5.0.4/asm-5.0.4.jar:/home/aryaman/.m2/repository/com/esotericsoftware/minlog/1.3.0/minlog-1.3.0.jar:/home/aryaman/.m2/repository/org/objenesis/objenesis/2.5.1/objenesis-2.5.1.jar:/home/aryaman/.m2/repository/org/msgpack/msgpack-core/0.8.16/msgpack-core-0.8.16.jar:/home/aryaman/.m2/repository/org/msgpack/jackson-dataformat-msgpack/0.8.16/jackson-dataformat-msgpack-0.8.16.jar:/home/aryaman/.m2/repository/graphics/scenery/jvrpn/1.1.0/jvrpn-1.1.0.jar:/home/aryaman/.m2/repository/graphics/scenery/jvrpn/1.1.0/jvrpn-1.1.0-natives-linux.jar:/home/aryaman/.m2/repository/io/scif/scifio/0.37.3/scifio-0.37.3.jar:/home/aryaman/.m2/repository/io/scif/scifio-jai-imageio/1.1.1/scifio-jai-imageio-1.1.1.jar:/home/aryaman/.m2/repository/net/imagej/imagej-common/0.28.2/imagej-common-0.28.2.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-roi/0.7.0/imglib2-roi-0.7.0.jar:/home/aryaman/.m2/repository/edu/ucar/udunits/4.3.18/udunits-4.3.18.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2/5.6.3/imglib2-5.6.3.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-cache/1.0.0-beta-11/imglib2-cache-1.0.0-beta-11.jar:/home/aryaman/.m2/repository/com/github/ben-manes/caffeine/caffeine/2.4.0/caffeine-2.4.0.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp-presets/ffmpeg/4.1-1.4.4/ffmpeg-4.1-1.4.4.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp/1.4.4/javacpp-1.4.4.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp-presets/ffmpeg/4.1-1.4.4/ffmpeg-4.1-1.4.4-linux-x86.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp-presets/ffmpeg/4.1-1.4.4/ffmpeg-4.1-1.4.4-linux-x86_64.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp-presets/ffmpeg/4.1-1.4.4/ffmpeg-4.1-1.4.4-macosx-x86_64.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp-presets/ffmpeg/4.1-1.4.4/ffmpeg-4.1-1.4.4-windows-x86_64.jar:/home/aryaman/.m2/repository/org/reflections/reflections/0.9.11/reflections-0.9.11.jar:/home/aryaman/.m2/repository/com/google/guava/guava/21.0/guava-21.0.jar:/home/aryaman/.m2/repository/org/javassist/javassist/3.24.0-GA/javassist-3.24.0-GA.jar:/home/aryaman/.m2/repository/sc/fiji/bigvolumeviewer/0.1.6/bigvolumeviewer-0.1.6.jar:/home/aryaman/.m2/repository/org/joml/joml/1.9.12/joml-1.9.12.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-math3/3.6.1/commons-math3-3.6.1.jar:/home/aryaman/.m2/repository/sc/fiji/bigdataviewer-core/5.1.0/bigdataviewer-core-5.1.0.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-realtransform/2.2.1/imglib2-realtransform-2.2.1.jar:/home/aryaman/.m2/repository/gov/nist/math/jama/1.0.3/jama-1.0.3.jar:/home/aryaman/.m2/repository/jitk/jitk-tps/3.0.1/jitk-tps-3.0.1.jar:/home/aryaman/.m2/repository/com/googlecode/efficient-java-matrix-library/ejml/0.24/ejml-0.24.jar:/home/aryaman/.m2/repository/log4j/log4j/1.2.17/log4j-1.2.17.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-ui/2.0.0/imglib2-ui-2.0.0.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-algorithm/0.11.1/imglib2-algorithm-0.11.1.jar:/home/aryaman/.m2/repository/org/ojalgo/ojalgo/45.1.1/ojalgo-45.1.1.jar:/home/aryaman/.m2/repository/sc/fiji/spim_data/2.2.2/spim_data-2.2.2.jar:/home/aryaman/.m2/repository/cisd/jhdf5/14.12.6/jhdf5-14.12.6.jar:/home/aryaman/.m2/repository/commons-lang/commons-lang/2.6/commons-lang-2.6.jar:/home/aryaman/.m2/repository/org/jdom/jdom2/2.0.6/jdom2-2.0.6.jar:/home/aryaman/.m2/repository/sc/fiji/bigdataviewer-vistools/1.0.0-beta-11/bigdataviewer-vistools-1.0.0-beta-11.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-ij/2.0.0-beta-44/imglib2-ij-2.0.0-beta-44.jar:/home/aryaman/.m2/repository/net/imagej/ij/1.52o/ij-1.52o.jar:/home/aryaman/.m2/repository/net/imagej/imagej/2.0.0-rc-71/imagej-2.0.0-rc-71.jar:/home/aryaman/.m2/repository/net/imagej/imagej-notebook/0.7.1/imagej-notebook-0.7.1.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-text/1.6/commons-text-1.6.jar:/home/aryaman/.m2/repository/org/knowm/xchart/xchart/3.5.4/xchart-3.5.4.jar:/home/aryaman/.m2/repository/de/erichseifert/vectorgraphics2d/VectorGraphics2D/0.13/VectorGraphics2D-0.13.jar:/home/aryaman/.m2/repository/net/imagej/imagej-ops/0.43.2/imagej-ops-0.43.2.jar:/home/aryaman/.m2/repository/net/imagej/imagej-mesh/0.7.1/imagej-mesh-0.7.1.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-algorithm-fft/0.2.0/imglib2-algorithm-fft-0.2.0.jar:/home/aryaman/.m2/repository/edu/mines/mines-jtk/20151125/mines-jtk-20151125.jar:/home/aryaman/.m2/repository/net/imagej/imagej-updater/0.9.3/imagej-updater-0.9.3.jar:/home/aryaman/.m2/repository/net/imagej/imagej-plugins-batch/0.1.1/imagej-plugins-batch-0.1.1.jar:/home/aryaman/.m2/repository/net/imagej/imagej-deprecated/0.1.2/imagej-deprecated-0.1.2.jar:/home/aryaman/.m2/repository/net/imagej/imagej-plugins-commands/0.8.1/imagej-plugins-commands-0.8.1.jar:/home/aryaman/.m2/repository/net/iharder/base64/2.3.8/base64-2.3.8.jar:/home/aryaman/.m2/repository/net/imagej/imagej-plugins-tools/0.3.1/imagej-plugins-tools-0.3.1.jar:/home/aryaman/.m2/repository/net/imagej/imagej-plugins-uploader-ssh/0.3.2/imagej-plugins-uploader-ssh-0.3.2.jar:/home/aryaman/.m2/repository/com/jcraft/jsch/0.1.55/jsch-0.1.55.jar:/home/aryaman/.m2/repository/net/imagej/imagej-plugins-uploader-webdav/0.2.2/imagej-plugins-uploader-webdav-0.2.2.jar:/home/aryaman/.m2/repository/net/imagej/imagej-scripting/0.8.2/imagej-scripting-0.8.2.jar:/home/aryaman/.m2/repository/org/scijava/scripting-beanshell/0.3.3/scripting-beanshell-0.3.3.jar:/home/aryaman/.m2/repository/org/apache-extras/beanshell/bsh/2.0b6/bsh-2.0b6.jar:/home/aryaman/.m2/repository/org/scijava/scripting-clojure/0.1.6/scripting-clojure-0.1.6.jar:/home/aryaman/.m2/repository/org/clojure/clojure/1.8.0/clojure-1.8.0.jar:/home/aryaman/.m2/repository/org/scijava/scripting-groovy/0.2.7/scripting-groovy-0.2.7.jar:/home/aryaman/.m2/repository/org/codehaus/groovy/groovy/2.4.8/groovy-2.4.8.jar:/home/aryaman/.m2/repository/org/apache/ivy/ivy/2.4.0/ivy-2.4.0.jar:/home/aryaman/.m2/repository/org/codehaus/gpars/gpars/1.2.1/gpars-1.2.1.jar:/home/aryaman/.m2/repository/org/multiverse/multiverse-core/0.7.0/multiverse-core-0.7.0.jar:/home/aryaman/.m2/repository/org/codehaus/jsr166-mirror/jsr166y/1.7.0/jsr166y-1.7.0.jar:/home/aryaman/.m2/repository/org/scijava/scripting-jruby/0.3.0/scripting-jruby-0.3.0.jar:/home/aryaman/.m2/repository/org/jruby/jruby-core/9.1.17.0/jruby-core-9.1.17.0.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-netdb/1.1.6/jnr-netdb-1.1.6.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-enxio/0.16/jnr-enxio-0.16.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-x86asm/1.0.2/jnr-x86asm-1.0.2.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-unixsocket/0.17/jnr-unixsocket-0.17.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-posix/3.0.44/jnr-posix-3.0.44.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-constants/0.9.9/jnr-constants-0.9.9.jar:/home/aryaman/.m2/repository/com/github/jnr/jffi/1.2.16/jffi-1.2.16.jar:/home/aryaman/.m2/repository/com/github/jnr/jffi/1.2.16/jffi-1.2.16-native.jar:/home/aryaman/.m2/repository/org/jruby/joni/joni/2.1.15/joni-2.1.15.jar:/home/aryaman/.m2/repository/org/jruby/extras/bytelist/1.0.15/bytelist-1.0.15.jar:/home/aryaman/.m2/repository/org/jruby/jcodings/jcodings/1.0.27/jcodings-1.0.27.jar:/home/aryaman/.m2/repository/org/jruby/dirgra/0.3/dirgra-0.3.jar:/home/aryaman/.m2/repository/com/headius/invokebinder/1.10/invokebinder-1.10.jar:/home/aryaman/.m2/repository/com/headius/options/1.4/options-1.4.jar:/home/aryaman/.m2/repository/com/headius/unsafe-fences/1.0/unsafe-fences-1.0.jar:/home/aryaman/.m2/repository/com/jcraft/jzlib/1.1.3/jzlib-1.1.3.jar:/home/aryaman/.m2/repository/com/martiansoftware/nailgun-server/0.9.1/nailgun-server-0.9.1.jar:/home/aryaman/.m2/repository/joda-time/joda-time/2.10.1/joda-time-2.10.1.jar:/home/aryaman/.m2/repository/com/headius/modulator/1.0/modulator-1.0.jar:/home/aryaman/.m2/repository/org/jruby/jruby-stdlib/9.1.17.0/jruby-stdlib-9.1.17.0.jar:/home/aryaman/.m2/repository/org/scijava/scripting-jython/0.4.2/scripting-jython-0.4.2.jar:/home/aryaman/.m2/repository/org/scijava/jython-shaded/2.7.1.1/jython-shaded-2.7.1.1.jar:/home/aryaman/.m2/repository/org/scijava/scripting-renjin/0.2.2/scripting-renjin-0.2.2.jar:/home/aryaman/.m2/repository/org/renjin/renjin-script-engine/0.8.1906/renjin-script-engine-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/renjin-core/0.8.1906/renjin-core-0.8.1906.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-math/2.2/commons-math-2.2.jar:/home/aryaman/.m2/repository/org/renjin/gcc-runtime/0.8.1906/gcc-runtime-0.8.1906.jar:/home/aryaman/.m2/repository/org/netlib/netlib-java/0.9.3-renjin-patched-2/netlib-java-0.9.3-renjin-patched-2.jar:/home/aryaman/.m2/repository/org/netlib/lapack/0.8/lapack-0.8.jar:/home/aryaman/.m2/repository/org/netlib/f2jutil/0.8/f2jutil-0.8.jar:/home/aryaman/.m2/repository/org/netlib/xerbla/0.8/xerbla-0.8.jar:/home/aryaman/.m2/repository/org/netlib/blas/0.8/blas-0.8.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-vfs2/2.0/commons-vfs2-2.0.jar:/home/aryaman/.m2/repository/commons-logging/commons-logging/1.2/commons-logging-1.2.jar:/home/aryaman/.m2/repository/org/apache/maven/scm/maven-scm-api/1.4/maven-scm-api-1.4.jar:/home/aryaman/.m2/repository/org/codehaus/plexus/plexus-utils/1.5.6/plexus-utils-1.5.6.jar:/home/aryaman/.m2/repository/org/apache/maven/scm/maven-scm-provider-svnexe/1.4/maven-scm-provider-svnexe-1.4.jar:/home/aryaman/.m2/repository/org/apache/maven/scm/maven-scm-provider-svn-commons/1.4/maven-scm-provider-svn-commons-1.4.jar:/home/aryaman/.m2/repository/regexp/regexp/1.3/regexp-1.3.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-compress/1.4.1/commons-compress-1.4.1.jar:/home/aryaman/.m2/repository/org/tukaani/xz/1.0/xz-1.0.jar:/home/aryaman/.m2/repository/edu/emory/mathcs/jtransforms/2.4/jtransforms-2.4.jar:/home/aryaman/.m2/repository/net/sf/jung/jung-api/2.0.1/jung-api-2.0.1.jar:/home/aryaman/.m2/repository/net/sourceforge/collections/collections-generic/4.01/collections-generic-4.01.jar:/home/aryaman/.m2/repository/net/sf/jung/jung-graph-impl/2.0.1/jung-graph-impl-2.0.1.jar:/home/aryaman/.m2/repository/org/ow2/asm/asm-util/5.0.4/asm-util-5.0.4.jar:/home/aryaman/.m2/repository/org/ow2/asm/asm-tree/5.0.4/asm-tree-5.0.4.jar:/home/aryaman/.m2/repository/org/ow2/asm/asm-commons/5.0.4/asm-commons-5.0.4.jar:/home/aryaman/.m2/repository/com/sun/codemodel/codemodel/2.6/codemodel-2.6.jar:/home/aryaman/.m2/repository/org/renjin/stats/0.8.1906/stats-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/renjin-appl/0.8.1906/renjin-appl-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/renjin-gnur-runtime/0.8.1906/renjin-gnur-runtime-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/methods/0.8.1906/methods-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/datasets/0.8.1906/datasets-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/utils/0.8.1906/utils-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/grDevices/0.8.1906/grDevices-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/graphics/0.8.1906/graphics-0.8.1906.jar:/home/aryaman/.m2/repository/org/scijava/scripting-scala/0.2.1/scripting-scala-0.2.1.jar:/home/aryaman/.m2/repository/org/scala-lang/scala-compiler/2.12.1/scala-compiler-2.12.1.jar:/home/aryaman/.m2/repository/org/scala-lang/scala-library/2.12.1/scala-library-2.12.1.jar:/home/aryaman/.m2/repository/org/scala-lang/scala-reflect/2.12.1/scala-reflect-2.12.1.jar:/home/aryaman/.m2/repository/org/scala-lang/modules/scala-xml_2.12/1.0.6/scala-xml_2.12-1.0.6.jar:/home/aryaman/.m2/repository/net/imagej/imagej-ui-swing/0.22.0/imagej-ui-swing-0.22.0.jar:/home/aryaman/.m2/repository/org/jfree/jfreechart/1.5.0/jfreechart-1.5.0.jar:/home/aryaman/.m2/repository/org/jhotdraw/jhotdraw/7.6.0/jhotdraw-7.6.0.jar:/home/aryaman/.m2/repository/net/imagej/imagej-ui-awt/0.3.1/imagej-ui-awt-0.3.1.jar:/home/aryaman/.m2/repository/org/scijava/scijava-plugins-commands/0.2.3/scijava-plugins-commands-0.2.3.jar:/home/aryaman/.m2/repository/org/scijava/scijava-plugins-platforms/0.3.1/scijava-plugins-platforms-0.3.1.jar:/home/aryaman/.m2/repository/org/scijava/scijava-plugins-text-markdown/0.1.3/scijava-plugins-text-markdown-0.1.3.jar:/home/aryaman/.m2/repository/org/markdownj/markdownj/0.3.0-1.0.2b4/markdownj-0.3.0-1.0.2b4.jar:/home/aryaman/.m2/repository/org/scijava/scijava-plugins-text-plain/0.1.3/scijava-plugins-text-plain-0.1.3.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-lang3/3.8.1/commons-lang3-3.8.1.jar:/home/aryaman/.m2/repository/org/antlr/ST4/4.0.8/ST4-4.0.8.jar:/home/aryaman/.m2/repository/org/antlr/antlr-runtime/3.5.2/antlr-runtime-3.5.2.jar:/home/aryaman/.m2/repository/com/github/lwjglx/lwjgl3-awt/45a85af/lwjgl3-awt-45a85af.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-jawt/3.2.3/lwjgl-jawt-3.2.3.jar";
-//                                      "/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.base:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.compiler:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.datatransfer:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.desktop:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.instrument:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.logging:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.management:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.management.rmi:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.naming:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.net.http:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.prefs:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.rmi:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.scripting:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.se:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.security.jgss:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.security.sasl:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.smartcardio:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.sql:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.sql.rowset:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.transaction.xa:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.xml:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.xml.crypto:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.accessibility:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.aot:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.attach:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.charsets:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.compiler:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.crypto.cryptoki:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.crypto.ec:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.dynalink:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.editpad:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.hotspot.agent:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.httpserver:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.internal.ed:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.internal.jvmstat:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.internal.le:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.internal.opt:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.internal.vm.ci:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.internal.vm.compiler:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.internal.vm.compiler.management:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jartool:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.javadoc:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jcmd:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jconsole:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jdeps:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jdi:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jdwp.agent:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jfr:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jlink:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jshell:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jsobject:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jstatd:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.localedata:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.management:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.management.agent:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.management.jfr:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.naming.dns:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.naming.rmi:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.net:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.pack:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.rmic:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.scripting.nashorn:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.scripting.nashorn.shell:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.sctp:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.security.auth:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.security.jgss:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.unsupported:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.unsupported.desktop:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.xml.dom:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.zipfs:/home/aryaman/Repositories/scenery/target/test-classes:/home/aryaman/Repositories/scenery/target/classes:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-jdk8/1.3.61/kotlin-stdlib-jdk8-1.3.61.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib/1.3.61/kotlin-stdlib-1.3.61.jar:/home/aryaman/.m2/repository/org/jetbrains/annotations/13.0/annotations-13.0.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-jdk7/1.3.61/kotlin-stdlib-jdk7-1.3.61.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-reflect/1.3.61/kotlin-reflect-1.3.61.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlinx/kotlinx-coroutines-core/1.3.2/kotlinx-coroutines-core-1.3.2.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-common/1.3.50/kotlin-stdlib-common-1.3.50.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-linux-amd64.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-linux-i586.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-macosx-universal.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-windows-amd64.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-linux-amd64.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-linux-i586.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-macosx-universal.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-windows-amd64.jar:/home/aryaman/.m2/repository/junit/junit/4.12/junit-4.12.jar:/home/aryaman/.m2/repository/org/hamcrest/hamcrest-core/1.3/hamcrest-core-1.3.jar:/home/aryaman/.m2/repository/org/slf4j/slf4j-api/1.7.25/slf4j-api-1.7.25.jar:/home/aryaman/.m2/repository/org/slf4j/slf4j-simple/1.7.25/slf4j-simple-1.7.25.jar:/home/aryaman/.m2/repository/net/clearvolume/cleargl/2.2.10/cleargl-2.2.10.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt-main/2.3.2/gluegen-rt-main-2.3.2.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-android-aarch64.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-android-armv6.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-linux-armv6.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-linux-armv6hf.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-solaris-amd64.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-solaris-i586.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-windows-i586.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all-main/2.3.2/jogl-all-main-2.3.2.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-android-aarch64.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-android-armv6.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-linux-armv6.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-linux-armv6hf.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-solaris-amd64.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-solaris-i586.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-windows-i586.jar:/home/aryaman/.m2/repository/net/clearcontrol/coremem/0.4.6/coremem-0.4.6.jar:/home/aryaman/.m2/repository/com/nativelibs4java/bridj/0.7.0/bridj-0.7.0.jar:/home/aryaman/.m2/repository/net/java/jinput/jinput/2.0.9/jinput-2.0.9.jar:/home/aryaman/.m2/repository/net/java/jinput/jinput/2.0.9/jinput-2.0.9-natives-all.jar:/home/aryaman/.m2/repository/org/scijava/scijava-common/2.77.0/scijava-common-2.77.0.jar:/home/aryaman/.m2/repository/org/scijava/parsington/1.0.4/parsington-1.0.4.jar:/home/aryaman/.m2/repository/org/bushe/eventbus/1.4/eventbus-1.4.jar:/home/aryaman/.m2/repository/org/scijava/script-editor/0.5.3/script-editor-0.5.3.jar:/home/aryaman/.m2/repository/org/scijava/scripting-java/0.4.1/scripting-java-0.4.1.jar:/home/aryaman/.m2/repository/org/scijava/minimaven/2.2.1/minimaven-2.2.1.jar:/home/aryaman/.m2/repository/com/fifesoft/rsyntaxtextarea/2.6.1/rsyntaxtextarea-2.6.1.jar:/home/aryaman/.m2/repository/com/fifesoft/languagesupport/2.6.0/languagesupport-2.6.0.jar:/home/aryaman/.m2/repository/com/fifesoft/autocomplete/2.6.1/autocomplete-2.6.1.jar:/home/aryaman/.m2/repository/org/mozilla/rhino/1.7.6/rhino-1.7.6.jar:/home/aryaman/.m2/repository/com/miglayout/miglayout/3.7.4/miglayout-3.7.4-swing.jar:/home/aryaman/.m2/repository/org/scijava/batch-processor/0.3.0/batch-processor-0.3.0.jar:/home/aryaman/.m2/repository/org/scijava/scijava-table/0.4.0/scijava-table-0.4.0.jar:/home/aryaman/.m2/repository/org/scijava/scijava-ui-swing/0.12.0/scijava-ui-swing-0.12.0.jar:/home/aryaman/.m2/repository/org/scijava/scijava-ui-awt/0.1.6/scijava-ui-awt-0.1.6.jar:/home/aryaman/.m2/repository/org/scijava/swing-checkbox-tree/1.0.2/swing-checkbox-tree-1.0.2.jar:/home/aryaman/.m2/repository/com/github/sbridges/object-inspector/object-inspector/0.1/object-inspector-0.1.jar:/home/aryaman/.m2/repository/net/sourceforge/jdatepicker/jdatepicker/1.3.2/jdatepicker-1.3.2.jar:/home/aryaman/.m2/repository/org/scijava/scijava-search/0.6.0/scijava-search-0.6.0.jar:/home/aryaman/.m2/repository/org/ocpsoft/prettytime/prettytime/4.0.1.Final/prettytime-4.0.1.Final.jar:/home/aryaman/.m2/repository/commons-io/commons-io/2.6/commons-io-2.6.jar:/home/aryaman/.m2/repository/org/scijava/ui-behaviour/1.7.4/ui-behaviour-1.7.4.jar:/home/aryaman/.m2/repository/com/google/code/gson/gson/2.8.5/gson-2.8.5.jar:/home/aryaman/.m2/repository/org/yaml/snakeyaml/1.23/snakeyaml-1.23.jar:/home/aryaman/.m2/repository/net/sf/trove4j/trove4j/3.0.3/trove4j-3.0.3.jar:/home/aryaman/.m2/repository/org/scijava/scripting-javascript/0.4.4/scripting-javascript-0.4.4.jar:/home/aryaman/.m2/repository/net/java/dev/jna/jna/4.5.2/jna-4.5.2.jar:/home/aryaman/.m2/repository/net/java/dev/jna/jna-platform/4.5.2/jna-platform-4.5.2.jar:/home/aryaman/.m2/repository/org/jocl/jocl/2.0.1/jocl-2.0.1.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl/3.2.3/lwjgl-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-glfw/3.2.3/lwjgl-glfw-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-jemalloc/3.2.3/lwjgl-jemalloc-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-vulkan/3.2.3/lwjgl-vulkan-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-opengl/3.2.3/lwjgl-opengl-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-openvr/3.2.3/lwjgl-openvr-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-xxhash/3.2.3/lwjgl-xxhash-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl/3.2.3/lwjgl-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-opengl/3.2.3/lwjgl-opengl-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-glfw/3.2.3/lwjgl-glfw-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-jemalloc/3.2.3/lwjgl-jemalloc-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-openvr/3.2.3/lwjgl-openvr-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-xxhash/3.2.3/lwjgl-xxhash-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-remotery/3.2.3/lwjgl-remotery-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-remotery/3.2.3/lwjgl-remotery-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/core/jackson-databind/2.9.10.1/jackson-databind-2.9.10.1.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/core/jackson-annotations/2.9.10/jackson-annotations-2.9.10.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/core/jackson-core/2.9.10/jackson-core-2.9.10.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/module/jackson-module-kotlin/2.9.9/jackson-module-kotlin-2.9.9.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/dataformat/jackson-dataformat-yaml/2.9.9/jackson-dataformat-yaml-2.9.9.jar:/home/aryaman/.m2/repository/graphics/scenery/spirvcrossj/0.7.0-1.1.106.0/spirvcrossj-0.7.0-1.1.106.0.jar:/home/aryaman/.m2/repository/graphics/scenery/spirvcrossj/0.7.0-1.1.106.0/spirvcrossj-0.7.0-1.1.106.0-natives-linux.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-junit/1.3.61/kotlin-test-junit-1.3.61.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-annotations-common/1.3.61/kotlin-test-annotations-common-1.3.61.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test/1.3.61/kotlin-test-1.3.61.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-common/1.3.61/kotlin-test-common-1.3.61.jar:/home/aryaman/.m2/repository/org/zeromq/jeromq/0.4.3/jeromq-0.4.3.jar:/home/aryaman/.m2/repository/org/zeromq/jnacl/0.1.0/jnacl-0.1.0.jar:/home/aryaman/.m2/repository/com/esotericsoftware/kryo/4.0.2/kryo-4.0.2.jar:/home/aryaman/.m2/repository/com/esotericsoftware/reflectasm/1.11.3/reflectasm-1.11.3.jar:/home/aryaman/.m2/repository/org/ow2/asm/asm/5.0.4/asm-5.0.4.jar:/home/aryaman/.m2/repository/com/esotericsoftware/minlog/1.3.0/minlog-1.3.0.jar:/home/aryaman/.m2/repository/org/objenesis/objenesis/2.5.1/objenesis-2.5.1.jar:/home/aryaman/.m2/repository/org/msgpack/msgpack-core/0.8.16/msgpack-core-0.8.16.jar:/home/aryaman/.m2/repository/org/msgpack/jackson-dataformat-msgpack/0.8.16/jackson-dataformat-msgpack-0.8.16.jar:/home/aryaman/.m2/repository/graphics/scenery/jvrpn/1.1.0/jvrpn-1.1.0.jar:/home/aryaman/.m2/repository/graphics/scenery/jvrpn/1.1.0/jvrpn-1.1.0-natives-linux.jar:/home/aryaman/.m2/repository/io/scif/scifio/0.37.3/scifio-0.37.3.jar:/home/aryaman/.m2/repository/io/scif/scifio-jai-imageio/1.1.1/scifio-jai-imageio-1.1.1.jar:/home/aryaman/.m2/repository/net/imagej/imagej-common/0.28.2/imagej-common-0.28.2.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-roi/0.7.0/imglib2-roi-0.7.0.jar:/home/aryaman/.m2/repository/edu/ucar/udunits/4.3.18/udunits-4.3.18.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2/5.6.3/imglib2-5.6.3.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-cache/1.0.0-beta-11/imglib2-cache-1.0.0-beta-11.jar:/home/aryaman/.m2/repository/com/github/ben-manes/caffeine/caffeine/2.4.0/caffeine-2.4.0.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp-presets/ffmpeg/4.1-1.4.4/ffmpeg-4.1-1.4.4.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp/1.4.4/javacpp-1.4.4.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp-presets/ffmpeg/4.1-1.4.4/ffmpeg-4.1-1.4.4-linux-x86.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp-presets/ffmpeg/4.1-1.4.4/ffmpeg-4.1-1.4.4-linux-x86_64.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp-presets/ffmpeg/4.1-1.4.4/ffmpeg-4.1-1.4.4-macosx-x86_64.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp-presets/ffmpeg/4.1-1.4.4/ffmpeg-4.1-1.4.4-windows-x86_64.jar:/home/aryaman/.m2/repository/org/reflections/reflections/0.9.11/reflections-0.9.11.jar:/home/aryaman/.m2/repository/com/google/guava/guava/21.0/guava-21.0.jar:/home/aryaman/.m2/repository/org/javassist/javassist/3.24.0-GA/javassist-3.24.0-GA.jar:/home/aryaman/.m2/repository/sc/fiji/bigvolumeviewer/0.1.6/bigvolumeviewer-0.1.6.jar:/home/aryaman/.m2/repository/org/joml/joml/1.9.12/joml-1.9.12.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-math3/3.6.1/commons-math3-3.6.1.jar:/home/aryaman/.m2/repository/sc/fiji/bigdataviewer-core/5.1.0/bigdataviewer-core-5.1.0.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-realtransform/2.2.1/imglib2-realtransform-2.2.1.jar:/home/aryaman/.m2/repository/gov/nist/math/jama/1.0.3/jama-1.0.3.jar:/home/aryaman/.m2/repository/jitk/jitk-tps/3.0.1/jitk-tps-3.0.1.jar:/home/aryaman/.m2/repository/com/googlecode/efficient-java-matrix-library/ejml/0.24/ejml-0.24.jar:/home/aryaman/.m2/repository/log4j/log4j/1.2.17/log4j-1.2.17.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-ui/2.0.0/imglib2-ui-2.0.0.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-algorithm/0.11.1/imglib2-algorithm-0.11.1.jar:/home/aryaman/.m2/repository/org/ojalgo/ojalgo/45.1.1/ojalgo-45.1.1.jar:/home/aryaman/.m2/repository/sc/fiji/spim_data/2.2.2/spim_data-2.2.2.jar:/home/aryaman/.m2/repository/cisd/jhdf5/14.12.6/jhdf5-14.12.6.jar:/home/aryaman/.m2/repository/commons-lang/commons-lang/2.6/commons-lang-2.6.jar:/home/aryaman/.m2/repository/org/jdom/jdom2/2.0.6/jdom2-2.0.6.jar:/home/aryaman/.m2/repository/sc/fiji/bigdataviewer-vistools/1.0.0-beta-11/bigdataviewer-vistools-1.0.0-beta-11.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-ij/2.0.0-beta-44/imglib2-ij-2.0.0-beta-44.jar:/home/aryaman/.m2/repository/net/imagej/ij/1.52o/ij-1.52o.jar:/home/aryaman/.m2/repository/net/imagej/imagej/2.0.0-rc-71/imagej-2.0.0-rc-71.jar:/home/aryaman/.m2/repository/net/imagej/imagej-notebook/0.7.1/imagej-notebook-0.7.1.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-text/1.6/commons-text-1.6.jar:/home/aryaman/.m2/repository/org/knowm/xchart/xchart/3.5.4/xchart-3.5.4.jar:/home/aryaman/.m2/repository/de/erichseifert/vectorgraphics2d/VectorGraphics2D/0.13/VectorGraphics2D-0.13.jar:/home/aryaman/.m2/repository/net/imagej/imagej-ops/0.43.2/imagej-ops-0.43.2.jar:/home/aryaman/.m2/repository/net/imagej/imagej-mesh/0.7.1/imagej-mesh-0.7.1.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-algorithm-fft/0.2.0/imglib2-algorithm-fft-0.2.0.jar:/home/aryaman/.m2/repository/edu/mines/mines-jtk/20151125/mines-jtk-20151125.jar:/home/aryaman/.m2/repository/net/imagej/imagej-updater/0.9.3/imagej-updater-0.9.3.jar:/home/aryaman/.m2/repository/net/imagej/imagej-plugins-batch/0.1.1/imagej-plugins-batch-0.1.1.jar:/home/aryaman/.m2/repository/net/imagej/imagej-deprecated/0.1.2/imagej-deprecated-0.1.2.jar:/home/aryaman/.m2/repository/net/imagej/imagej-plugins-commands/0.8.1/imagej-plugins-commands-0.8.1.jar:/home/aryaman/.m2/repository/net/iharder/base64/2.3.8/base64-2.3.8.jar:/home/aryaman/.m2/repository/net/imagej/imagej-plugins-tools/0.3.1/imagej-plugins-tools-0.3.1.jar:/home/aryaman/.m2/repository/net/imagej/imagej-plugins-uploader-ssh/0.3.2/imagej-plugins-uploader-ssh-0.3.2.jar:/home/aryaman/.m2/repository/com/jcraft/jsch/0.1.55/jsch-0.1.55.jar:/home/aryaman/.m2/repository/net/imagej/imagej-plugins-uploader-webdav/0.2.2/imagej-plugins-uploader-webdav-0.2.2.jar:/home/aryaman/.m2/repository/net/imagej/imagej-scripting/0.8.2/imagej-scripting-0.8.2.jar:/home/aryaman/.m2/repository/org/scijava/scripting-beanshell/0.3.3/scripting-beanshell-0.3.3.jar:/home/aryaman/.m2/repository/org/apache-extras/beanshell/bsh/2.0b6/bsh-2.0b6.jar:/home/aryaman/.m2/repository/org/scijava/scripting-clojure/0.1.6/scripting-clojure-0.1.6.jar:/home/aryaman/.m2/repository/org/clojure/clojure/1.8.0/clojure-1.8.0.jar:/home/aryaman/.m2/repository/org/scijava/scripting-groovy/0.2.7/scripting-groovy-0.2.7.jar:/home/aryaman/.m2/repository/org/codehaus/groovy/groovy/2.4.8/groovy-2.4.8.jar:/home/aryaman/.m2/repository/org/apache/ivy/ivy/2.4.0/ivy-2.4.0.jar:/home/aryaman/.m2/repository/org/codehaus/gpars/gpars/1.2.1/gpars-1.2.1.jar:/home/aryaman/.m2/repository/org/multiverse/multiverse-core/0.7.0/multiverse-core-0.7.0.jar:/home/aryaman/.m2/repository/org/codehaus/jsr166-mirror/jsr166y/1.7.0/jsr166y-1.7.0.jar:/home/aryaman/.m2/repository/org/scijava/scripting-jruby/0.3.0/scripting-jruby-0.3.0.jar:/home/aryaman/.m2/repository/org/jruby/jruby-core/9.1.17.0/jruby-core-9.1.17.0.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-netdb/1.1.6/jnr-netdb-1.1.6.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-enxio/0.16/jnr-enxio-0.16.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-x86asm/1.0.2/jnr-x86asm-1.0.2.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-unixsocket/0.17/jnr-unixsocket-0.17.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-posix/3.0.44/jnr-posix-3.0.44.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-constants/0.9.9/jnr-constants-0.9.9.jar:/home/aryaman/.m2/repository/com/github/jnr/jffi/1.2.16/jffi-1.2.16.jar:/home/aryaman/.m2/repository/com/github/jnr/jffi/1.2.16/jffi-1.2.16-native.jar:/home/aryaman/.m2/repository/org/jruby/joni/joni/2.1.15/joni-2.1.15.jar:/home/aryaman/.m2/repository/org/jruby/extras/bytelist/1.0.15/bytelist-1.0.15.jar:/home/aryaman/.m2/repository/org/jruby/jcodings/jcodings/1.0.27/jcodings-1.0.27.jar:/home/aryaman/.m2/repository/org/jruby/dirgra/0.3/dirgra-0.3.jar:/home/aryaman/.m2/repository/com/headius/invokebinder/1.10/invokebinder-1.10.jar:/home/aryaman/.m2/repository/com/headius/options/1.4/options-1.4.jar:/home/aryaman/.m2/repository/com/headius/unsafe-fences/1.0/unsafe-fences-1.0.jar:/home/aryaman/.m2/repository/com/jcraft/jzlib/1.1.3/jzlib-1.1.3.jar:/home/aryaman/.m2/repository/com/martiansoftware/nailgun-server/0.9.1/nailgun-server-0.9.1.jar:/home/aryaman/.m2/repository/joda-time/joda-time/2.10.1/joda-time-2.10.1.jar:/home/aryaman/.m2/repository/com/headius/modulator/1.0/modulator-1.0.jar:/home/aryaman/.m2/repository/org/jruby/jruby-stdlib/9.1.17.0/jruby-stdlib-9.1.17.0.jar:/home/aryaman/.m2/repository/org/scijava/scripting-jython/0.4.2/scripting-jython-0.4.2.jar:/home/aryaman/.m2/repository/org/scijava/jython-shaded/2.7.1.1/jython-shaded-2.7.1.1.jar:/home/aryaman/.m2/repository/org/scijava/scripting-renjin/0.2.2/scripting-renjin-0.2.2.jar:/home/aryaman/.m2/repository/org/renjin/renjin-script-engine/0.8.1906/renjin-script-engine-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/renjin-core/0.8.1906/renjin-core-0.8.1906.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-math/2.2/commons-math-2.2.jar:/home/aryaman/.m2/repository/org/renjin/gcc-runtime/0.8.1906/gcc-runtime-0.8.1906.jar:/home/aryaman/.m2/repository/org/netlib/netlib-java/0.9.3-renjin-patched-2/netlib-java-0.9.3-renjin-patched-2.jar:/home/aryaman/.m2/repository/org/netlib/lapack/0.8/lapack-0.8.jar:/home/aryaman/.m2/repository/org/netlib/f2jutil/0.8/f2jutil-0.8.jar:/home/aryaman/.m2/repository/org/netlib/xerbla/0.8/xerbla-0.8.jar:/home/aryaman/.m2/repository/org/netlib/blas/0.8/blas-0.8.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-vfs2/2.0/commons-vfs2-2.0.jar:/home/aryaman/.m2/repository/commons-logging/commons-logging/1.2/commons-logging-1.2.jar:/home/aryaman/.m2/repository/org/apache/maven/scm/maven-scm-api/1.4/maven-scm-api-1.4.jar:/home/aryaman/.m2/repository/org/codehaus/plexus/plexus-utils/1.5.6/plexus-utils-1.5.6.jar:/home/aryaman/.m2/repository/org/apache/maven/scm/maven-scm-provider-svnexe/1.4/maven-scm-provider-svnexe-1.4.jar:/home/aryaman/.m2/repository/org/apache/maven/scm/maven-scm-provider-svn-commons/1.4/maven-scm-provider-svn-commons-1.4.jar:/home/aryaman/.m2/repository/regexp/regexp/1.3/regexp-1.3.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-compress/1.4.1/commons-compress-1.4.1.jar:/home/aryaman/.m2/repository/org/tukaani/xz/1.0/xz-1.0.jar:/home/aryaman/.m2/repository/edu/emory/mathcs/jtransforms/2.4/jtransforms-2.4.jar:/home/aryaman/.m2/repository/net/sf/jung/jung-api/2.0.1/jung-api-2.0.1.jar:/home/aryaman/.m2/repository/net/sourceforge/collections/collections-generic/4.01/collections-generic-4.01.jar:/home/aryaman/.m2/repository/net/sf/jung/jung-graph-impl/2.0.1/jung-graph-impl-2.0.1.jar:/home/aryaman/.m2/repository/org/ow2/asm/asm-util/5.0.4/asm-util-5.0.4.jar:/home/aryaman/.m2/repository/org/ow2/asm/asm-tree/5.0.4/asm-tree-5.0.4.jar:/home/aryaman/.m2/repository/org/ow2/asm/asm-commons/5.0.4/asm-commons-5.0.4.jar:/home/aryaman/.m2/repository/com/sun/codemodel/codemodel/2.6/codemodel-2.6.jar:/home/aryaman/.m2/repository/org/renjin/stats/0.8.1906/stats-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/renjin-appl/0.8.1906/renjin-appl-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/renjin-gnur-runtime/0.8.1906/renjin-gnur-runtime-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/methods/0.8.1906/methods-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/datasets/0.8.1906/datasets-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/utils/0.8.1906/utils-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/grDevices/0.8.1906/grDevices-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/graphics/0.8.1906/graphics-0.8.1906.jar:/home/aryaman/.m2/repository/org/scijava/scripting-scala/0.2.1/scripting-scala-0.2.1.jar:/home/aryaman/.m2/repository/org/scala-lang/scala-compiler/2.12.1/scala-compiler-2.12.1.jar:/home/aryaman/.m2/repository/org/scala-lang/scala-library/2.12.1/scala-library-2.12.1.jar:/home/aryaman/.m2/repository/org/scala-lang/scala-reflect/2.12.1/scala-reflect-2.12.1.jar:/home/aryaman/.m2/repository/org/scala-lang/modules/scala-xml_2.12/1.0.6/scala-xml_2.12-1.0.6.jar:/home/aryaman/.m2/repository/net/imagej/imagej-ui-swing/0.22.0/imagej-ui-swing-0.22.0.jar:/home/aryaman/.m2/repository/org/jfree/jfreechart/1.5.0/jfreechart-1.5.0.jar:/home/aryaman/.m2/repository/org/jhotdraw/jhotdraw/7.6.0/jhotdraw-7.6.0.jar:/home/aryaman/.m2/repository/net/imagej/imagej-ui-awt/0.3.1/imagej-ui-awt-0.3.1.jar:/home/aryaman/.m2/repository/org/scijava/scijava-plugins-commands/0.2.3/scijava-plugins-commands-0.2.3.jar:/home/aryaman/.m2/repository/org/scijava/scijava-plugins-platforms/0.3.1/scijava-plugins-platforms-0.3.1.jar:/home/aryaman/.m2/repository/org/scijava/scijava-plugins-text-markdown/0.1.3/scijava-plugins-text-markdown-0.1.3.jar:/home/aryaman/.m2/repository/org/markdownj/markdownj/0.3.0-1.0.2b4/markdownj-0.3.0-1.0.2b4.jar:/home/aryaman/.m2/repository/org/scijava/scijava-plugins-text-plain/0.1.3/scijava-plugins-text-plain-0.1.3.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-lang3/3.8.1/commons-lang3-3.8.1.jar:/home/aryaman/.m2/repository/org/antlr/ST4/4.0.8/ST4-4.0.8.jar:/home/aryaman/.m2/repository/org/antlr/antlr-runtime/3.5.2/antlr-runtime-3.5.2.jar:/home/aryaman/.m2/repository/com/github/lwjglx/lwjgl3-awt/45a85af/lwjgl3-awt-45a85af.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-jawt/3.2.3/lwjgl-jawt-3.2.3.jar:/home/aryaman/Repositories/scenery/mpi.jar:/home/aryaman/Repositories/scenery-insitu/target/test-classes:/home/aryaman/Repositories/scenery-insitu/target/classes:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-jdk8/1.3.21/kotlin-stdlib-jdk8-1.3.21.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib/1.3.21/kotlin-stdlib-1.3.21.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-common/1.3.21/kotlin-stdlib-common-1.3.21.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-jdk7/1.3.21/kotlin-stdlib-jdk7-1.3.21.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-reflect/1.3.21/kotlin-reflect-1.3.21.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlinx/kotlinx-coroutines-core/1.1.1/kotlinx-coroutines-core-1.1.1.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlinx/kotlinx-coroutines-core-common/1.1.1/kotlinx-coroutines-core-common-1.1.1.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-junit/1.3.21/kotlin-test-junit-1.3.21.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-annotations-common/1.3.21/kotlin-test-annotations-common-1.3.21.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test/1.3.21/kotlin-test-1.3.21.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-common/1.3.21/kotlin-test-common-1.3.21.jar:/home/aryaman/.m2/repository/org/scijava/scijava-common/2.76.1/scijava-common-2.76.1.jar:/home/aryaman/.m2/repository/org/scijava/script-editor/0.4.0/script-editor-0.4.0.jar:/home/aryaman/.m2/repository/org/scijava/scijava-table/0.2.0/scijava-table-0.2.0.jar:/home/aryaman/.m2/repository/org/scijava/scijava-search/0.5.0/scijava-search-0.5.0.jar:/home/aryaman/.m2/repository/net/imagej/imagej-common/0.28.0/imagej-common-0.28.0.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-realtransform/2.1.0/imglib2-realtransform-2.1.0.jar:/home/aryaman/.m2/repository/jitk/jitk-tps/3.0.0/jitk-tps-3.0.0.jar:/home/aryaman/.m2/repository/net/imagej/ij/1.52i/ij-1.52i.jar:/home/aryaman/.m2/repository/net/imagej/imagej/2.0.0-rc-69/imagej-2.0.0-rc-69.jar:/home/aryaman/.m2/repository/net/imagej/imagej-notebook/0.3.0/imagej-notebook-0.3.0.jar:/home/aryaman/.m2/repository/net/imagej/imagej-ops/0.43.1/imagej-ops-0.43.1.jar:/home/aryaman/.m2/repository/org/scijava/jython-shaded/2.7.1/jython-shaded-2.7.1.jar:/home/aryaman/.m2/repository/org/msgpack/msgpack-core/0.8.20/msgpack-core-0.8.20.jar:/home/aryaman/.m2/repository/org/msgpack/jackson-dataformat-msgpack/0.8.20/jackson-dataformat-msgpack-0.8.20.jar";
-//    "/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.base:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.compiler:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.datatransfer:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.desktop:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.instrument:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.logging:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.management:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.management.rmi:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.naming:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.net.http:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.prefs:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.rmi:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.scripting:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.se:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.security.jgss:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.security.sasl:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.smartcardio:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.sql:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.sql.rowset:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.transaction.xa:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.xml:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/java.xml.crypto:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.accessibility:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.aot:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.attach:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.charsets:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.compiler:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.crypto.cryptoki:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.crypto.ec:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.dynalink:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.editpad:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.hotspot.agent:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.httpserver:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.internal.ed:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.internal.jvmstat:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.internal.le:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.internal.opt:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.internal.vm.ci:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.internal.vm.compiler:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.internal.vm.compiler.management:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jartool:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.javadoc:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jcmd:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jconsole:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jdeps:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jdi:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jdwp.agent:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jfr:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jlink:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jshell:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jsobject:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.jstatd:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.localedata:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.management:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.management.agent:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.management.jfr:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.naming.dns:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.naming.rmi:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.net:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.pack:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.rmic:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.scripting.nashorn:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.scripting.nashorn.shell:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.sctp:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.security.auth:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.security.jgss:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.unsupported:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.unsupported.desktop:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.xml.dom:/usr/lib/jvm/adoptopenjdk-11-hotspot-amd64!/jdk.zipfs:/home/aryaman/Repositories/scenery/target/test-classes:/home/aryaman/Repositories/scenery/target/classes:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-jdk8/1.4-M1/kotlin-stdlib-jdk8-1.4-M1.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib/1.4-M1/kotlin-stdlib-1.4-M1.jar:/home/aryaman/.m2/repository/org/jetbrains/annotations/13.0/annotations-13.0.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-jdk7/1.4-M1/kotlin-stdlib-jdk7-1.4-M1.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-reflect/1.4-M1/kotlin-reflect-1.4-M1.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlinx/kotlinx-coroutines-core/1.3.5-1.4-M1/kotlinx-coroutines-core-1.3.5-1.4-M1.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-common/1.4-M1/kotlin-stdlib-common-1.4-M1.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-linux-amd64.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-linux-i586.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-macosx-universal.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-windows-amd64.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-linux-amd64.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-linux-i586.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-macosx-universal.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-windows-amd64.jar:/home/aryaman/.m2/repository/junit/junit/4.13/junit-4.13.jar:/home/aryaman/.m2/repository/org/hamcrest/hamcrest-core/1.3/hamcrest-core-1.3.jar:/home/aryaman/.m2/repository/org/slf4j/slf4j-api/1.7.25/slf4j-api-1.7.25.jar:/home/aryaman/.m2/repository/org/slf4j/slf4j-simple/1.7.25/slf4j-simple-1.7.25.jar:/home/aryaman/.m2/repository/net/clearvolume/cleargl/2.2.10/cleargl-2.2.10.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt-main/2.3.2/gluegen-rt-main-2.3.2.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-android-aarch64.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-android-armv6.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-linux-armv6.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-linux-armv6hf.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-solaris-amd64.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-solaris-i586.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-windows-i586.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all-main/2.3.2/jogl-all-main-2.3.2.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-android-aarch64.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-android-armv6.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-linux-armv6.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-linux-armv6hf.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-solaris-amd64.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-solaris-i586.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-windows-i586.jar:/home/aryaman/.m2/repository/org/joml/joml/1.9.22/joml-1.9.22.jar:/home/aryaman/.m2/repository/com/github/scenerygraphics/vector/958f2e6/vector-958f2e6.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2/5.9.2/imglib2-5.9.2.jar:/home/aryaman/.m2/repository/net/clearcontrol/coremem/0.4.6/coremem-0.4.6.jar:/home/aryaman/.m2/repository/com/nativelibs4java/bridj/0.7.0/bridj-0.7.0.jar:/home/aryaman/.m2/repository/net/java/jinput/jinput/2.0.9/jinput-2.0.9.jar:/home/aryaman/.m2/repository/net/java/jinput/jinput/2.0.9/jinput-2.0.9-natives-all.jar:/home/aryaman/.m2/repository/org/scijava/scijava-common/2.83.2/scijava-common-2.83.2.jar:/home/aryaman/.m2/repository/org/scijava/parsington/2.0.0/parsington-2.0.0.jar:/home/aryaman/.m2/repository/org/bushe/eventbus/1.4/eventbus-1.4.jar:/home/aryaman/.m2/repository/org/scijava/script-editor/0.5.3/script-editor-0.5.3.jar:/home/aryaman/.m2/repository/org/scijava/scripting-java/0.4.1/scripting-java-0.4.1.jar:/home/aryaman/.m2/repository/org/scijava/minimaven/2.2.2/minimaven-2.2.2.jar:/home/aryaman/.m2/repository/com/fifesoft/rsyntaxtextarea/3.1.1/rsyntaxtextarea-3.1.1.jar:/home/aryaman/.m2/repository/com/fifesoft/languagesupport/3.1.0/languagesupport-3.1.0.jar:/home/aryaman/.m2/repository/com/fifesoft/autocomplete/3.1.0/autocomplete-3.1.0.jar:/home/aryaman/.m2/repository/org/mozilla/rhino/1.7.6/rhino-1.7.6.jar:/home/aryaman/.m2/repository/com/miglayout/miglayout/3.7.4/miglayout-3.7.4-swing.jar:/home/aryaman/.m2/repository/org/scijava/batch-processor/0.4.1/batch-processor-0.4.1.jar:/home/aryaman/.m2/repository/org/scijava/scijava-table/0.5.0/scijava-table-0.5.0.jar:/home/aryaman/.m2/repository/org/scijava/scijava-ui-swing/0.13.2/scijava-ui-swing-0.13.2.jar:/home/aryaman/.m2/repository/org/scijava/swing-checkbox-tree/1.0.2/swing-checkbox-tree-1.0.2.jar:/home/aryaman/.m2/repository/net/sourceforge/jdatepicker/jdatepicker/1.3.2/jdatepicker-1.3.2.jar:/home/aryaman/.m2/repository/org/scijava/scijava-search/0.7.0/scijava-search-0.7.0.jar:/home/aryaman/.m2/repository/org/ocpsoft/prettytime/prettytime/4.0.1.Final/prettytime-4.0.1.Final.jar:/home/aryaman/.m2/repository/commons-io/commons-io/2.6/commons-io-2.6.jar:/home/aryaman/.m2/repository/org/scijava/ui-behaviour/2.0.1/ui-behaviour-2.0.1.jar:/home/aryaman/.m2/repository/com/google/code/gson/gson/2.8.6/gson-2.8.6.jar:/home/aryaman/.m2/repository/org/yaml/snakeyaml/1.26/snakeyaml-1.26.jar:/home/aryaman/.m2/repository/net/sf/trove4j/trove4j/3.0.3/trove4j-3.0.3.jar:/home/aryaman/.m2/repository/org/scijava/scripting-javascript/0.5.0/scripting-javascript-0.5.0.jar:/home/aryaman/.m2/repository/org/scijava/scripting-jython/1.0.0/scripting-jython-1.0.0.jar:/home/aryaman/.m2/repository/org/python/jython-slim/2.7.2/jython-slim-2.7.2.jar:/home/aryaman/.m2/repository/org/antlr/antlr/3.5.2/antlr-3.5.2.jar:/home/aryaman/.m2/repository/javax/servlet/javax.servlet-api/3.1.0/javax.servlet-api-3.1.0.jar:/home/aryaman/.m2/repository/org/antlr/antlr-runtime/3.5.2/antlr-runtime-3.5.2.jar:/home/aryaman/.m2/repository/org/bouncycastle/bcpkix-jdk15on/1.62/bcpkix-jdk15on-1.62.jar:/home/aryaman/.m2/repository/org/bouncycastle/bcprov-jdk15on/1.62/bcprov-jdk15on-1.62.jar:/home/aryaman/.m2/repository/org/ow2/asm/asm/7.1/asm-7.1.jar:/home/aryaman/.m2/repository/org/ow2/asm/asm-commons/7.1/asm-commons-7.1.jar:/home/aryaman/.m2/repository/org/ow2/asm/asm-tree/7.1/asm-tree-7.1.jar:/home/aryaman/.m2/repository/org/ow2/asm/asm-analysis/7.1/asm-analysis-7.1.jar:/home/aryaman/.m2/repository/org/ow2/asm/asm-util/7.1/asm-util-7.1.jar:/home/aryaman/.m2/repository/com/google/guava/failureaccess/1.0.1/failureaccess-1.0.1.jar:/home/aryaman/.m2/repository/com/ibm/icu/icu4j/59.1/icu4j-59.1.jar:/home/aryaman/.m2/repository/com/carrotsearch/java-sizeof/0.0.5/java-sizeof-0.0.5.jar:/home/aryaman/.m2/repository/com/github/jnr/jffi/1.2.20/jffi-1.2.20.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-netdb/1.1.6/jnr-netdb-1.1.6.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-ffi/2.1.10/jnr-ffi-2.1.10.jar:/home/aryaman/.m2/repository/com/github/jnr/jffi/1.2.19/jffi-1.2.19-native.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-a64asm/1.0.0/jnr-a64asm-1.0.0.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-x86asm/1.0.2/jnr-x86asm-1.0.2.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-posix/3.0.50/jnr-posix-3.0.50.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-constants/0.9.12/jnr-constants-0.9.12.jar:/home/aryaman/.m2/repository/jline/jline/2.14.5/jline-2.14.5.jar:/home/aryaman/.m2/repository/io/netty/netty-buffer/4.1.45.Final/netty-buffer-4.1.45.Final.jar:/home/aryaman/.m2/repository/io/netty/netty-codec/4.1.45.Final/netty-codec-4.1.45.Final.jar:/home/aryaman/.m2/repository/io/netty/netty-common/4.1.45.Final/netty-common-4.1.45.Final.jar:/home/aryaman/.m2/repository/io/netty/netty-handler/4.1.45.Final/netty-handler-4.1.45.Final.jar:/home/aryaman/.m2/repository/io/netty/netty-resolver/4.1.45.Final/netty-resolver-4.1.45.Final.jar:/home/aryaman/.m2/repository/io/netty/netty-transport/4.1.45.Final/netty-transport-4.1.45.Final.jar:/home/aryaman/.m2/repository/org/apache/ant/ant/1.9.7/ant-1.9.7.jar:/home/aryaman/.m2/repository/org/apache/ant/ant-launcher/1.9.7/ant-launcher-1.9.7.jar:/home/aryaman/.m2/repository/net/java/dev/jna/jna/4.5.2/jna-4.5.2.jar:/home/aryaman/.m2/repository/net/java/dev/jna/jna-platform/4.5.2/jna-platform-4.5.2.jar:/home/aryaman/.m2/repository/org/jocl/jocl/2.0.1/jocl-2.0.1.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl/3.2.3/lwjgl-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-glfw/3.2.3/lwjgl-glfw-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-jemalloc/3.2.3/lwjgl-jemalloc-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-vulkan/3.2.3/lwjgl-vulkan-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-opengl/3.2.3/lwjgl-opengl-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-openvr/3.2.3/lwjgl-openvr-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-xxhash/3.2.3/lwjgl-xxhash-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl/3.2.3/lwjgl-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-opengl/3.2.3/lwjgl-opengl-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-glfw/3.2.3/lwjgl-glfw-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-jemalloc/3.2.3/lwjgl-jemalloc-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-openvr/3.2.3/lwjgl-openvr-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-xxhash/3.2.3/lwjgl-xxhash-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-remotery/3.2.3/lwjgl-remotery-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-remotery/3.2.3/lwjgl-remotery-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/core/jackson-databind/2.9.10.1/jackson-databind-2.9.10.1.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/core/jackson-annotations/2.9.10/jackson-annotations-2.9.10.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/core/jackson-core/2.9.10/jackson-core-2.9.10.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/module/jackson-module-kotlin/2.9.9/jackson-module-kotlin-2.9.9.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/dataformat/jackson-dataformat-yaml/2.9.9/jackson-dataformat-yaml-2.9.9.jar:/home/aryaman/.m2/repository/graphics/scenery/spirvcrossj/0.7.0-1.1.106.0/spirvcrossj-0.7.0-1.1.106.0.jar:/home/aryaman/.m2/repository/graphics/scenery/spirvcrossj/0.7.0-1.1.106.0/spirvcrossj-0.7.0-1.1.106.0-natives-linux.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-junit/1.4-M1/kotlin-test-junit-1.4-M1.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-annotations-common/1.4-M1/kotlin-test-annotations-common-1.4-M1.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test/1.4-M1/kotlin-test-1.4-M1.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-common/1.4-M1/kotlin-test-common-1.4-M1.jar:/home/aryaman/.m2/repository/org/zeromq/jeromq/0.4.3/jeromq-0.4.3.jar:/home/aryaman/.m2/repository/org/zeromq/jnacl/0.1.0/jnacl-0.1.0.jar:/home/aryaman/.m2/repository/com/esotericsoftware/kryo/4.0.2/kryo-4.0.2.jar:/home/aryaman/.m2/repository/com/esotericsoftware/reflectasm/1.11.3/reflectasm-1.11.3.jar:/home/aryaman/.m2/repository/com/esotericsoftware/minlog/1.3.0/minlog-1.3.0.jar:/home/aryaman/.m2/repository/org/objenesis/objenesis/2.5.1/objenesis-2.5.1.jar:/home/aryaman/.m2/repository/org/msgpack/msgpack-core/0.8.16/msgpack-core-0.8.16.jar:/home/aryaman/.m2/repository/org/msgpack/jackson-dataformat-msgpack/0.8.16/jackson-dataformat-msgpack-0.8.16.jar:/home/aryaman/.m2/repository/graphics/scenery/jvrpn/1.1.0/jvrpn-1.1.0.jar:/home/aryaman/.m2/repository/graphics/scenery/jvrpn/1.1.0/jvrpn-1.1.0-natives-linux.jar:/home/aryaman/.m2/repository/io/scif/scifio/0.40.0/scifio-0.40.0.jar:/home/aryaman/.m2/repository/io/scif/scifio-jai-imageio/1.1.1/scifio-jai-imageio-1.1.1.jar:/home/aryaman/.m2/repository/net/imagej/imagej-common/0.33.0/imagej-common-0.33.0.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-roi/0.10.2/imglib2-roi-0.10.2.jar:/home/aryaman/.m2/repository/edu/ucar/udunits/4.3.18/udunits-4.3.18.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-cache/1.0.0-beta-13/imglib2-cache-1.0.0-beta-13.jar:/home/aryaman/.m2/repository/com/github/ben-manes/caffeine/caffeine/2.4.0/caffeine-2.4.0.jar:/home/aryaman/.m2/repository/org/bytedeco/ffmpeg/4.2.1-1.5.2/ffmpeg-4.2.1-1.5.2.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp/1.5.2/javacpp-1.5.2.jar:/home/aryaman/.m2/repository/org/bytedeco/ffmpeg/4.2.1-1.5.2/ffmpeg-4.2.1-1.5.2-linux-x86_64.jar:/home/aryaman/.m2/repository/org/bytedeco/ffmpeg/4.2.1-1.5.2/ffmpeg-4.2.1-1.5.2-macosx-x86_64.jar:/home/aryaman/.m2/repository/org/bytedeco/ffmpeg/4.2.1-1.5.2/ffmpeg-4.2.1-1.5.2-windows-x86_64.jar:/home/aryaman/.m2/repository/org/reflections/reflections/0.9.11/reflections-0.9.11.jar:/home/aryaman/.m2/repository/com/google/guava/guava/21.0/guava-21.0.jar:/home/aryaman/.m2/repository/org/javassist/javassist/3.27.0-GA/javassist-3.27.0-GA.jar:/home/aryaman/.m2/repository/sc/fiji/bigvolumeviewer/0.1.8/bigvolumeviewer-0.1.8.jar:/home/aryaman/.m2/repository/commons-lang/commons-lang/2.6/commons-lang-2.6.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-algorithm/0.11.2/imglib2-algorithm-0.11.2.jar:/home/aryaman/.m2/repository/gov/nist/math/jama/1.0.3/jama-1.0.3.jar:/home/aryaman/.m2/repository/org/ojalgo/ojalgo/45.1.1/ojalgo-45.1.1.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-realtransform/3.0.0/imglib2-realtransform-3.0.0.jar:/home/aryaman/.m2/repository/jitk/jitk-tps/3.0.1/jitk-tps-3.0.1.jar:/home/aryaman/.m2/repository/com/googlecode/efficient-java-matrix-library/ejml/0.24/ejml-0.24.jar:/home/aryaman/.m2/repository/log4j/log4j/1.2.17/log4j-1.2.17.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-ui/2.0.1/imglib2-ui-2.0.1.jar:/home/aryaman/.m2/repository/org/antlr/ST4/4.0.8/ST4-4.0.8.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-math3/3.6.1/commons-math3-3.6.1.jar:/home/aryaman/.m2/repository/org/jdom/jdom2/2.0.6/jdom2-2.0.6.jar:/home/aryaman/.m2/repository/sc/fiji/bigdataviewer-core/9.0.2/bigdataviewer-core-9.0.2.jar:/home/aryaman/.m2/repository/cisd/jhdf5/14.12.6/jhdf5-14.12.6.jar:/home/aryaman/.m2/repository/org/scijava/scijava-listeners/1.0.0-beta-2/scijava-listeners-1.0.0-beta-2.jar:/home/aryaman/.m2/repository/sc/fiji/bigdataviewer-vistools/1.0.0-beta-20/bigdataviewer-vistools-1.0.0-beta-20.jar:/home/aryaman/.m2/repository/sc/fiji/spim_data/2.2.4/spim_data-2.2.4.jar:/home/aryaman/.m2/repository/org/lwjglx/lwjgl3-awt/0.1.7/lwjgl3-awt-0.1.7.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-jawt/3.2.3/lwjgl-jawt-3.2.3.jar:/home/aryaman/Repositories/scenery/mpi.jar:/home/aryaman/.m2/repository/net/imagej/imagej/2.0.0-rc-72/imagej-2.0.0-rc-72.jar:/home/aryaman/.m2/repository/net/imagej/imagej-notebook/0.7.1/imagej-notebook-0.7.1.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-text/1.8/commons-text-1.8.jar:/home/aryaman/.m2/repository/org/knowm/xchart/xchart/3.5.4/xchart-3.5.4.jar:/home/aryaman/.m2/repository/de/erichseifert/vectorgraphics2d/VectorGraphics2D/0.13/VectorGraphics2D-0.13.jar:/home/aryaman/.m2/repository/net/imagej/imagej-ops/0.45.5/imagej-ops-0.45.5.jar:/home/aryaman/.m2/repository/net/imagej/imagej-mesh/0.8.0/imagej-mesh-0.8.0.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-algorithm-fft/0.2.0/imglib2-algorithm-fft-0.2.0.jar:/home/aryaman/.m2/repository/edu/mines/mines-jtk/20151125/mines-jtk-20151125.jar:/home/aryaman/.m2/repository/net/imagej/imagej-updater/0.10.4/imagej-updater-0.10.4.jar:/home/aryaman/.m2/repository/net/imagej/imagej-plugins-batch/0.1.1/imagej-plugins-batch-0.1.1.jar:/home/aryaman/.m2/repository/net/imagej/imagej-deprecated/0.1.4/imagej-deprecated-0.1.4.jar:/home/aryaman/.m2/repository/net/imagej/imagej-plugins-commands/0.8.1/imagej-plugins-commands-0.8.1.jar:/home/aryaman/.m2/repository/com/github/sbridges/object-inspector/object-inspector/0.1/object-inspector-0.1.jar:/home/aryaman/.m2/repository/net/iharder/base64/2.3.8/base64-2.3.8.jar:/home/aryaman/.m2/repository/net/imagej/imagej-plugins-tools/0.3.1/imagej-plugins-tools-0.3.1.jar:/home/aryaman/.m2/repository/net/imagej/imagej-plugins-uploader-ssh/0.3.2/imagej-plugins-uploader-ssh-0.3.2.jar:/home/aryaman/.m2/repository/com/jcraft/jsch/0.1.55/jsch-0.1.55.jar:/home/aryaman/.m2/repository/net/imagej/imagej-plugins-uploader-webdav/0.3.2/imagej-plugins-uploader-webdav-0.3.2.jar:/home/aryaman/.m2/repository/org/apache/jackrabbit/jackrabbit-webdav/2.21.1/jackrabbit-webdav-2.21.1.jar:/home/aryaman/.m2/repository/org/apache/httpcomponents/httpclient/4.5.12/httpclient-4.5.12.jar:/home/aryaman/.m2/repository/commons-codec/commons-codec/1.14/commons-codec-1.14.jar:/home/aryaman/.m2/repository/org/apache/httpcomponents/httpcore/4.4.13/httpcore-4.4.13.jar:/home/aryaman/.m2/repository/net/imagej/imagej-scripting/0.8.2/imagej-scripting-0.8.2.jar:/home/aryaman/.m2/repository/org/scijava/scripting-beanshell/0.3.4/scripting-beanshell-0.3.4.jar:/home/aryaman/.m2/repository/org/apache-extras/beanshell/bsh/2.0b6/bsh-2.0b6.jar:/home/aryaman/.m2/repository/org/scijava/scripting-clojure/0.1.6/scripting-clojure-0.1.6.jar:/home/aryaman/.m2/repository/org/clojure/clojure/1.8.0/clojure-1.8.0.jar:/home/aryaman/.m2/repository/org/scijava/scripting-groovy/0.3.0/scripting-groovy-0.3.0.jar:/home/aryaman/.m2/repository/org/codehaus/groovy/groovy/3.0.4/groovy-3.0.4.jar:/home/aryaman/.m2/repository/org/codehaus/groovy/groovy-dateutil/3.0.4/groovy-dateutil-3.0.4.jar:/home/aryaman/.m2/repository/org/codehaus/groovy/groovy-json/3.0.4/groovy-json-3.0.4.jar:/home/aryaman/.m2/repository/org/codehaus/groovy/groovy-swing/3.0.4/groovy-swing-3.0.4.jar:/home/aryaman/.m2/repository/org/codehaus/groovy/groovy-templates/3.0.4/groovy-templates-3.0.4.jar:/home/aryaman/.m2/repository/org/codehaus/groovy/groovy-xml/3.0.4/groovy-xml-3.0.4.jar:/home/aryaman/.m2/repository/org/apache/ivy/ivy/2.4.0/ivy-2.4.0.jar:/home/aryaman/.m2/repository/org/codehaus/gpars/gpars/1.2.1/gpars-1.2.1.jar:/home/aryaman/.m2/repository/org/multiverse/multiverse-core/0.7.0/multiverse-core-0.7.0.jar:/home/aryaman/.m2/repository/org/codehaus/jsr166-mirror/jsr166y/1.7.0/jsr166y-1.7.0.jar:/home/aryaman/.m2/repository/org/scijava/scripting-jruby/0.3.1/scripting-jruby-0.3.1.jar:/home/aryaman/.m2/repository/org/jruby/jruby-core/9.1.17.0/jruby-core-9.1.17.0.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-enxio/0.16/jnr-enxio-0.16.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-unixsocket/0.17/jnr-unixsocket-0.17.jar:/home/aryaman/.m2/repository/org/jruby/joni/joni/2.1.15/joni-2.1.15.jar:/home/aryaman/.m2/repository/org/jruby/extras/bytelist/1.0.15/bytelist-1.0.15.jar:/home/aryaman/.m2/repository/org/jruby/jcodings/jcodings/1.0.27/jcodings-1.0.27.jar:/home/aryaman/.m2/repository/org/jruby/dirgra/0.3/dirgra-0.3.jar:/home/aryaman/.m2/repository/com/headius/invokebinder/1.10/invokebinder-1.10.jar:/home/aryaman/.m2/repository/com/headius/options/1.4/options-1.4.jar:/home/aryaman/.m2/repository/com/headius/unsafe-fences/1.0/unsafe-fences-1.0.jar:/home/aryaman/.m2/repository/com/jcraft/jzlib/1.1.3/jzlib-1.1.3.jar:/home/aryaman/.m2/repository/com/martiansoftware/nailgun-server/0.9.1/nailgun-server-0.9.1.jar:/home/aryaman/.m2/repository/joda-time/joda-time/2.10.6/joda-time-2.10.6.jar:/home/aryaman/.m2/repository/com/headius/modulator/1.0/modulator-1.0.jar:/home/aryaman/.m2/repository/org/jruby/jruby-stdlib/9.1.17.0/jruby-stdlib-9.1.17.0.jar:/home/aryaman/.m2/repository/org/scijava/scripting-renjin/0.2.3/scripting-renjin-0.2.3.jar:/home/aryaman/.m2/repository/org/renjin/renjin-script-engine/0.8.1906/renjin-script-engine-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/renjin-core/0.8.1906/renjin-core-0.8.1906.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-math/2.2/commons-math-2.2.jar:/home/aryaman/.m2/repository/org/renjin/gcc-runtime/0.8.1906/gcc-runtime-0.8.1906.jar:/home/aryaman/.m2/repository/org/netlib/netlib-java/0.9.3-renjin-patched-2/netlib-java-0.9.3-renjin-patched-2.jar:/home/aryaman/.m2/repository/org/netlib/lapack/0.8/lapack-0.8.jar:/home/aryaman/.m2/repository/org/netlib/f2jutil/0.8/f2jutil-0.8.jar:/home/aryaman/.m2/repository/org/netlib/xerbla/0.8/xerbla-0.8.jar:/home/aryaman/.m2/repository/org/netlib/blas/0.8/blas-0.8.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-vfs2/2.0/commons-vfs2-2.0.jar:/home/aryaman/.m2/repository/commons-logging/commons-logging/1.2/commons-logging-1.2.jar:/home/aryaman/.m2/repository/org/apache/maven/scm/maven-scm-api/1.4/maven-scm-api-1.4.jar:/home/aryaman/.m2/repository/org/codehaus/plexus/plexus-utils/1.5.6/plexus-utils-1.5.6.jar:/home/aryaman/.m2/repository/org/apache/maven/scm/maven-scm-provider-svnexe/1.4/maven-scm-provider-svnexe-1.4.jar:/home/aryaman/.m2/repository/org/apache/maven/scm/maven-scm-provider-svn-commons/1.4/maven-scm-provider-svn-commons-1.4.jar:/home/aryaman/.m2/repository/regexp/regexp/1.3/regexp-1.3.jar:/home/aryaman/.m2/repository/edu/emory/mathcs/jtransforms/2.4/jtransforms-2.4.jar:/home/aryaman/.m2/repository/net/sf/jung/jung-api/2.0.1/jung-api-2.0.1.jar:/home/aryaman/.m2/repository/net/sourceforge/collections/collections-generic/4.01/collections-generic-4.01.jar:/home/aryaman/.m2/repository/net/sf/jung/jung-graph-impl/2.0.1/jung-graph-impl-2.0.1.jar:/home/aryaman/.m2/repository/com/sun/codemodel/codemodel/2.6/codemodel-2.6.jar:/home/aryaman/.m2/repository/org/renjin/stats/0.8.1906/stats-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/renjin-appl/0.8.1906/renjin-appl-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/renjin-gnur-runtime/0.8.1906/renjin-gnur-runtime-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/methods/0.8.1906/methods-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/datasets/0.8.1906/datasets-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/utils/0.8.1906/utils-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/grDevices/0.8.1906/grDevices-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/graphics/0.8.1906/graphics-0.8.1906.jar:/home/aryaman/.m2/repository/org/scijava/scripting-scala/0.2.2/scripting-scala-0.2.2.jar:/home/aryaman/.m2/repository/org/scala-lang/scala-compiler/2.12.1/scala-compiler-2.12.1.jar:/home/aryaman/.m2/repository/org/scala-lang/scala-library/2.12.1/scala-library-2.12.1.jar:/home/aryaman/.m2/repository/org/scala-lang/scala-reflect/2.12.1/scala-reflect-2.12.1.jar:/home/aryaman/.m2/repository/org/scala-lang/modules/scala-xml_2.12/1.0.6/scala-xml_2.12-1.0.6.jar:/home/aryaman/.m2/repository/net/imagej/imagej-ui-swing/0.23.2/imagej-ui-swing-0.23.2.jar:/home/aryaman/.m2/repository/org/scijava/scijava-ui-awt/0.1.7/scijava-ui-awt-0.1.7.jar:/home/aryaman/.m2/repository/com/miglayout/miglayout-swing/5.2/miglayout-swing-5.2.jar:/home/aryaman/.m2/repository/com/miglayout/miglayout-core/5.2/miglayout-core-5.2.jar:/home/aryaman/.m2/repository/org/jfree/jfreechart/1.5.0/jfreechart-1.5.0.jar:/home/aryaman/.m2/repository/org/jhotdraw/jhotdraw/7.6.0/jhotdraw-7.6.0.jar:/home/aryaman/.m2/repository/net/imagej/imagej-ui-awt/0.3.1/imagej-ui-awt-0.3.1.jar:/home/aryaman/.m2/repository/org/scijava/scijava-io-http/0.2.0/scijava-io-http-0.2.0.jar:/home/aryaman/.m2/repository/com/squareup/okhttp3/okhttp/3.6.0/okhttp-3.6.0.jar:/home/aryaman/.m2/repository/com/squareup/okio/okio/1.11.0/okio-1.11.0.jar:/home/aryaman/.m2/repository/org/scijava/scijava-plugins-commands/0.2.3/scijava-plugins-commands-0.2.3.jar:/home/aryaman/.m2/repository/org/scijava/scijava-plugins-io-table/0.2.0/scijava-plugins-io-table-0.2.0.jar:/home/aryaman/.m2/repository/org/scijava/scijava-plugins-platforms/0.3.1/scijava-plugins-platforms-0.3.1.jar:/home/aryaman/.m2/repository/org/scijava/scijava-plugins-text-markdown/0.1.3/scijava-plugins-text-markdown-0.1.3.jar:/home/aryaman/.m2/repository/org/markdownj/markdownj/0.3.0-1.0.2b4/markdownj-0.3.0-1.0.2b4.jar:/home/aryaman/.m2/repository/org/scijava/scijava-plugins-text-plain/0.1.3/scijava-plugins-text-plain-0.1.3.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-lang3/3.10/commons-lang3-3.10.jar:/home/aryaman/.m2/repository/net/imagej/ij/1.52p/ij-1.52p.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-ij/2.0.0-beta-46/imglib2-ij-2.0.0-beta-46.jar:/home/aryaman/.m2/repository/org/janelia/saalfeldlab/n5/2.1.3/n5-2.1.3.jar:/home/aryaman/.m2/repository/org/tukaani/xz/1.8/xz-1.8.jar:/home/aryaman/.m2/repository/org/lz4/lz4-java/1.7.1/lz4-java-1.7.1.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-compress/1.20/commons-compress-1.20.jar:/home/aryaman/.m2/repository/org/janelia/saalfeldlab/n5-imglib2/3.4.1/n5-imglib2-3.4.1.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-label-multisets/0.9.0/imglib2-label-multisets-0.9.0.jar:/home/aryaman/Repositories/scenery-insitu/target/test-classes:/home/aryaman/Repositories/scenery-insitu/target/classes:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-jdk8/1.3.21/kotlin-stdlib-jdk8-1.3.21.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib/1.3.21/kotlin-stdlib-1.3.21.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-common/1.3.21/kotlin-stdlib-common-1.3.21.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-jdk7/1.3.21/kotlin-stdlib-jdk7-1.3.21.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-reflect/1.3.21/kotlin-reflect-1.3.21.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlinx/kotlinx-coroutines-core/1.1.1/kotlinx-coroutines-core-1.1.1.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlinx/kotlinx-coroutines-core-common/1.1.1/kotlinx-coroutines-core-common-1.1.1.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-junit/1.3.21/kotlin-test-junit-1.3.21.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-annotations-common/1.3.21/kotlin-test-annotations-common-1.3.21.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test/1.3.21/kotlin-test-1.3.21.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-common/1.3.21/kotlin-test-common-1.3.21.jar:/home/aryaman/.m2/repository/junit/junit/4.12/junit-4.12.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2/5.6.3/imglib2-5.6.3.jar:/home/aryaman/.m2/repository/org/scijava/scijava-common/2.76.1/scijava-common-2.76.1.jar:/home/aryaman/.m2/repository/org/scijava/parsington/1.0.4/parsington-1.0.4.jar:/home/aryaman/.m2/repository/org/scijava/script-editor/0.4.0/script-editor-0.4.0.jar:/home/aryaman/.m2/repository/org/scijava/minimaven/2.2.1/minimaven-2.2.1.jar:/home/aryaman/.m2/repository/com/fifesoft/rsyntaxtextarea/2.6.1/rsyntaxtextarea-2.6.1.jar:/home/aryaman/.m2/repository/com/fifesoft/languagesupport/2.6.0/languagesupport-2.6.0.jar:/home/aryaman/.m2/repository/com/fifesoft/autocomplete/2.6.1/autocomplete-2.6.1.jar:/home/aryaman/.m2/repository/org/scijava/batch-processor/0.3.0/batch-processor-0.3.0.jar:/home/aryaman/.m2/repository/org/scijava/scijava-table/0.2.0/scijava-table-0.2.0.jar:/home/aryaman/.m2/repository/org/scijava/scijava-ui-swing/0.12.0/scijava-ui-swing-0.12.0.jar:/home/aryaman/.m2/repository/org/scijava/scijava-ui-awt/0.1.6/scijava-ui-awt-0.1.6.jar:/home/aryaman/.m2/repository/org/scijava/scijava-search/0.5.0/scijava-search-0.5.0.jar:/home/aryaman/.m2/repository/org/scijava/ui-behaviour/1.7.4/ui-behaviour-1.7.4.jar:/home/aryaman/.m2/repository/com/google/code/gson/gson/2.8.5/gson-2.8.5.jar:/home/aryaman/.m2/repository/org/yaml/snakeyaml/1.23/snakeyaml-1.23.jar:/home/aryaman/.m2/repository/org/scijava/scripting-javascript/0.4.4/scripting-javascript-0.4.4.jar:/home/aryaman/.m2/repository/org/scijava/scripting-jython/0.4.2/scripting-jython-0.4.2.jar:/home/aryaman/.m2/repository/org/scijava/jython-shaded/2.7.1/jython-shaded-2.7.1.jar:/home/aryaman/.m2/repository/org/ow2/asm/asm/5.0.4/asm-5.0.4.jar:/home/aryaman/.m2/repository/io/scif/scifio/0.37.3/scifio-0.37.3.jar:/home/aryaman/.m2/repository/net/imagej/imagej-common/0.28.0/imagej-common-0.28.0.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-roi/0.7.0/imglib2-roi-0.7.0.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-cache/1.0.0-beta-11/imglib2-cache-1.0.0-beta-11.jar:/home/aryaman/.m2/repository/org/javassist/javassist/3.24.0-GA/javassist-3.24.0-GA.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-algorithm/0.11.1/imglib2-algorithm-0.11.1.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-realtransform/2.1.0/imglib2-realtransform-2.1.0.jar:/home/aryaman/.m2/repository/jitk/jitk-tps/3.0.0/jitk-tps-3.0.0.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-ui/2.0.0/imglib2-ui-2.0.0.jar:/home/aryaman/.m2/repository/sc/fiji/bigdataviewer-core/5.1.0/bigdataviewer-core-5.1.0.jar:/home/aryaman/.m2/repository/sc/fiji/bigdataviewer-vistools/1.0.0-beta-11/bigdataviewer-vistools-1.0.0-beta-11.jar:/home/aryaman/.m2/repository/sc/fiji/spim_data/2.2.2/spim_data-2.2.2.jar:/home/aryaman/.m2/repository/org/lz4/lz4-java/1.5.0/lz4-java-1.5.0.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-compress/1.19/commons-compress-1.19.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-label-multisets/0.8.1/imglib2-label-multisets-0.8.1.jar:/home/aryaman/.m2/repository/org/msgpack/msgpack-core/0.8.20/msgpack-core-0.8.20.jar:/home/aryaman/.m2/repository/org/msgpack/jackson-dataformat-msgpack/0.8.20/jackson-dataformat-msgpack-0.8.20.jar:/home/aryaman/Repositories/insituclient/target/test-classes:/home/aryaman/Repositories/insituclient/target/classes:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-jdk8/1.3.60/kotlin-stdlib-jdk8-1.3.60.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib/1.3.60/kotlin-stdlib-1.3.60.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-common/1.3.60/kotlin-stdlib-common-1.3.60.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-jdk7/1.3.60/kotlin-stdlib-jdk7-1.3.60.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-reflect/1.3.60/kotlin-reflect-1.3.60.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-junit/1.3.60/kotlin-test-junit-1.3.60.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-annotations-common/1.3.60/kotlin-test-annotations-common-1.3.60.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/core/jackson-databind/2.9.9.3/jackson-databind-2.9.9.3.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/core/jackson-annotations/2.9.0/jackson-annotations-2.9.0.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/core/jackson-core/2.9.9/jackson-core-2.9.9.jar:/home/aryaman/.m2/repository/org/scijava/minimaven/2.2.0/minimaven-2.2.0.jar:/home/aryaman/.m2/repository/com/fifesoft/autocomplete/2.6.0/autocomplete-2.6.0.jar:/home/aryaman/.m2/repository/org/scijava/scijava-ui-swing/0.11.0/scijava-ui-swing-0.11.0.jar:/home/aryaman/.m2/repository/net/imagej/imagej-common/0.31.0/imagej-common-0.31.0.jar:/home/aryaman/.m2/repository/com/google/guava/guava/20.0/guava-20.0.jar:/home/aryaman/.m2/repository/org/javassist/javassist/3.21.0-GA/javassist-3.21.0-GA.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-realtransform/2.2.1/imglib2-realtransform-2.2.1.jar:/home/aryaman/.m2/repository/sc/fiji/bigdataviewer-core/8.0.0/bigdataviewer-core-8.0.0.jar:/home/aryaman/.m2/repository/sc/fiji/bigdataviewer-vistools/1.0.0-beta-18/bigdataviewer-vistools-1.0.0-beta-18.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test/1.3.60/kotlin-test-1.3.60.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-common/1.3.60/kotlin-test-common-1.3.60.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp-presets/ffmpeg/4.1-1.4.4/ffmpeg-4.1-1.4.4.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp/1.4.4/javacpp-1.4.4.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp-presets/ffmpeg/4.1-1.4.4/ffmpeg-4.1-1.4.4-linux-x86.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp-presets/ffmpeg/4.1-1.4.4/ffmpeg-4.1-1.4.4-linux-x86_64.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp-presets/ffmpeg/4.1-1.4.4/ffmpeg-4.1-1.4.4-macosx-x86_64.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp-presets/ffmpeg/4.1-1.4.4/ffmpeg-4.1-1.4.4-windows-x86_64.jar";
-    // "/usr/lib/jvm/jdk-14.0.1!/java.base:/usr/lib/jvm/jdk-14.0.1!/java.compiler:/usr/lib/jvm/jdk-14.0.1!/java.datatransfer:/usr/lib/jvm/jdk-14.0.1!/java.desktop:/usr/lib/jvm/jdk-14.0.1!/java.instrument:/usr/lib/jvm/jdk-14.0.1!/java.logging:/usr/lib/jvm/jdk-14.0.1!/java.management:/usr/lib/jvm/jdk-14.0.1!/java.management.rmi:/usr/lib/jvm/jdk-14.0.1!/java.naming:/usr/lib/jvm/jdk-14.0.1!/java.net.http:/usr/lib/jvm/jdk-14.0.1!/java.prefs:/usr/lib/jvm/jdk-14.0.1!/java.rmi:/usr/lib/jvm/jdk-14.0.1!/java.scripting:/usr/lib/jvm/jdk-14.0.1!/java.se:/usr/lib/jvm/jdk-14.0.1!/java.security.jgss:/usr/lib/jvm/jdk-14.0.1!/java.security.sasl:/usr/lib/jvm/jdk-14.0.1!/java.smartcardio:/usr/lib/jvm/jdk-14.0.1!/java.sql:/usr/lib/jvm/jdk-14.0.1!/java.sql.rowset:/usr/lib/jvm/jdk-14.0.1!/java.transaction.xa:/usr/lib/jvm/jdk-14.0.1!/java.xml:/usr/lib/jvm/jdk-14.0.1!/java.xml.crypto:/usr/lib/jvm/jdk-14.0.1!/jdk.accessibility:/usr/lib/jvm/jdk-14.0.1!/jdk.aot:/usr/lib/jvm/jdk-14.0.1!/jdk.attach:/usr/lib/jvm/jdk-14.0.1!/jdk.charsets:/usr/lib/jvm/jdk-14.0.1!/jdk.compiler:/usr/lib/jvm/jdk-14.0.1!/jdk.crypto.cryptoki:/usr/lib/jvm/jdk-14.0.1!/jdk.crypto.ec:/usr/lib/jvm/jdk-14.0.1!/jdk.dynalink:/usr/lib/jvm/jdk-14.0.1!/jdk.editpad:/usr/lib/jvm/jdk-14.0.1!/jdk.hotspot.agent:/usr/lib/jvm/jdk-14.0.1!/jdk.httpserver:/usr/lib/jvm/jdk-14.0.1!/jdk.incubator.foreign:/usr/lib/jvm/jdk-14.0.1!/jdk.incubator.jpackage:/usr/lib/jvm/jdk-14.0.1!/jdk.internal.ed:/usr/lib/jvm/jdk-14.0.1!/jdk.internal.jvmstat:/usr/lib/jvm/jdk-14.0.1!/jdk.internal.le:/usr/lib/jvm/jdk-14.0.1!/jdk.internal.opt:/usr/lib/jvm/jdk-14.0.1!/jdk.internal.vm.ci:/usr/lib/jvm/jdk-14.0.1!/jdk.internal.vm.compiler:/usr/lib/jvm/jdk-14.0.1!/jdk.internal.vm.compiler.management:/usr/lib/jvm/jdk-14.0.1!/jdk.jartool:/usr/lib/jvm/jdk-14.0.1!/jdk.javadoc:/usr/lib/jvm/jdk-14.0.1!/jdk.jcmd:/usr/lib/jvm/jdk-14.0.1!/jdk.jconsole:/usr/lib/jvm/jdk-14.0.1!/jdk.jdeps:/usr/lib/jvm/jdk-14.0.1!/jdk.jdi:/usr/lib/jvm/jdk-14.0.1!/jdk.jdwp.agent:/usr/lib/jvm/jdk-14.0.1!/jdk.jfr:/usr/lib/jvm/jdk-14.0.1!/jdk.jlink:/usr/lib/jvm/jdk-14.0.1!/jdk.jshell:/usr/lib/jvm/jdk-14.0.1!/jdk.jsobject:/usr/lib/jvm/jdk-14.0.1!/jdk.jstatd:/usr/lib/jvm/jdk-14.0.1!/jdk.localedata:/usr/lib/jvm/jdk-14.0.1!/jdk.management:/usr/lib/jvm/jdk-14.0.1!/jdk.management.agent:/usr/lib/jvm/jdk-14.0.1!/jdk.management.jfr:/usr/lib/jvm/jdk-14.0.1!/jdk.naming.dns:/usr/lib/jvm/jdk-14.0.1!/jdk.naming.rmi:/usr/lib/jvm/jdk-14.0.1!/jdk.net:/usr/lib/jvm/jdk-14.0.1!/jdk.nio.mapmode:/usr/lib/jvm/jdk-14.0.1!/jdk.rmic:/usr/lib/jvm/jdk-14.0.1!/jdk.scripting.nashorn:/usr/lib/jvm/jdk-14.0.1!/jdk.scripting.nashorn.shell:/usr/lib/jvm/jdk-14.0.1!/jdk.sctp:/usr/lib/jvm/jdk-14.0.1!/jdk.security.auth:/usr/lib/jvm/jdk-14.0.1!/jdk.security.jgss:/usr/lib/jvm/jdk-14.0.1!/jdk.unsupported:/usr/lib/jvm/jdk-14.0.1!/jdk.unsupported.desktop:/usr/lib/jvm/jdk-14.0.1!/jdk.xml.dom:/usr/lib/jvm/jdk-14.0.1!/jdk.zipfs:/home/aryaman/Repositories/scenery/target/test-classes:/home/aryaman/Repositories/scenery/target/classes:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-jdk8/1.3.61/kotlin-stdlib-jdk8-1.3.61.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib/1.3.61/kotlin-stdlib-1.3.61.jar:/home/aryaman/.m2/repository/org/jetbrains/annotations/13.0/annotations-13.0.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-jdk7/1.3.61/kotlin-stdlib-jdk7-1.3.61.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-reflect/1.3.61/kotlin-reflect-1.3.61.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlinx/kotlinx-coroutines-core/1.3.2/kotlinx-coroutines-core-1.3.2.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-common/1.3.50/kotlin-stdlib-common-1.3.50.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-linux-amd64.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-linux-i586.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-macosx-universal.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-windows-amd64.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-linux-amd64.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-linux-i586.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-macosx-universal.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-windows-amd64.jar:/home/aryaman/.m2/repository/junit/junit/4.12/junit-4.12.jar:/home/aryaman/.m2/repository/org/hamcrest/hamcrest-core/1.3/hamcrest-core-1.3.jar:/home/aryaman/.m2/repository/org/slf4j/slf4j-api/1.7.25/slf4j-api-1.7.25.jar:/home/aryaman/.m2/repository/org/slf4j/slf4j-simple/1.7.25/slf4j-simple-1.7.25.jar:/home/aryaman/.m2/repository/net/clearvolume/cleargl/2.2.10/cleargl-2.2.10.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt-main/2.3.2/gluegen-rt-main-2.3.2.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-android-aarch64.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-android-armv6.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-linux-armv6.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-linux-armv6hf.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-solaris-amd64.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-solaris-i586.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-windows-i586.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all-main/2.3.2/jogl-all-main-2.3.2.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-android-aarch64.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-android-armv6.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-linux-armv6.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-linux-armv6hf.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-solaris-amd64.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-solaris-i586.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-windows-i586.jar:/home/aryaman/.m2/repository/net/clearcontrol/coremem/0.4.6/coremem-0.4.6.jar:/home/aryaman/.m2/repository/com/nativelibs4java/bridj/0.7.0/bridj-0.7.0.jar:/home/aryaman/.m2/repository/net/java/jinput/jinput/2.0.9/jinput-2.0.9.jar:/home/aryaman/.m2/repository/net/java/jinput/jinput/2.0.9/jinput-2.0.9-natives-all.jar:/home/aryaman/.m2/repository/org/scijava/scijava-common/2.77.0/scijava-common-2.77.0.jar:/home/aryaman/.m2/repository/org/scijava/parsington/1.0.4/parsington-1.0.4.jar:/home/aryaman/.m2/repository/org/bushe/eventbus/1.4/eventbus-1.4.jar:/home/aryaman/.m2/repository/org/scijava/script-editor/0.5.3/script-editor-0.5.3.jar:/home/aryaman/.m2/repository/org/scijava/scripting-java/0.4.1/scripting-java-0.4.1.jar:/home/aryaman/.m2/repository/org/scijava/minimaven/2.2.1/minimaven-2.2.1.jar:/home/aryaman/.m2/repository/com/fifesoft/rsyntaxtextarea/2.6.1/rsyntaxtextarea-2.6.1.jar:/home/aryaman/.m2/repository/com/fifesoft/languagesupport/2.6.0/languagesupport-2.6.0.jar:/home/aryaman/.m2/repository/com/fifesoft/autocomplete/2.6.1/autocomplete-2.6.1.jar:/home/aryaman/.m2/repository/org/mozilla/rhino/1.7.6/rhino-1.7.6.jar:/home/aryaman/.m2/repository/com/miglayout/miglayout/3.7.4/miglayout-3.7.4-swing.jar:/home/aryaman/.m2/repository/org/scijava/batch-processor/0.3.0/batch-processor-0.3.0.jar:/home/aryaman/.m2/repository/org/scijava/scijava-table/0.4.0/scijava-table-0.4.0.jar:/home/aryaman/.m2/repository/org/scijava/scijava-ui-swing/0.12.0/scijava-ui-swing-0.12.0.jar:/home/aryaman/.m2/repository/org/scijava/scijava-ui-awt/0.1.6/scijava-ui-awt-0.1.6.jar:/home/aryaman/.m2/repository/org/scijava/swing-checkbox-tree/1.0.2/swing-checkbox-tree-1.0.2.jar:/home/aryaman/.m2/repository/com/github/sbridges/object-inspector/object-inspector/0.1/object-inspector-0.1.jar:/home/aryaman/.m2/repository/net/sourceforge/jdatepicker/jdatepicker/1.3.2/jdatepicker-1.3.2.jar:/home/aryaman/.m2/repository/org/scijava/scijava-search/0.6.0/scijava-search-0.6.0.jar:/home/aryaman/.m2/repository/org/ocpsoft/prettytime/prettytime/4.0.1.Final/prettytime-4.0.1.Final.jar:/home/aryaman/.m2/repository/commons-io/commons-io/2.6/commons-io-2.6.jar:/home/aryaman/.m2/repository/org/scijava/ui-behaviour/1.7.4/ui-behaviour-1.7.4.jar:/home/aryaman/.m2/repository/com/google/code/gson/gson/2.8.5/gson-2.8.5.jar:/home/aryaman/.m2/repository/org/yaml/snakeyaml/1.23/snakeyaml-1.23.jar:/home/aryaman/.m2/repository/net/sf/trove4j/trove4j/3.0.3/trove4j-3.0.3.jar:/home/aryaman/.m2/repository/org/scijava/scripting-javascript/0.4.4/scripting-javascript-0.4.4.jar:/home/aryaman/.m2/repository/net/java/dev/jna/jna/4.5.2/jna-4.5.2.jar:/home/aryaman/.m2/repository/net/java/dev/jna/jna-platform/4.5.2/jna-platform-4.5.2.jar:/home/aryaman/.m2/repository/org/jocl/jocl/2.0.1/jocl-2.0.1.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl/3.2.3/lwjgl-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-glfw/3.2.3/lwjgl-glfw-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-jemalloc/3.2.3/lwjgl-jemalloc-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-vulkan/3.2.3/lwjgl-vulkan-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-opengl/3.2.3/lwjgl-opengl-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-openvr/3.2.3/lwjgl-openvr-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-xxhash/3.2.3/lwjgl-xxhash-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl/3.2.3/lwjgl-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-opengl/3.2.3/lwjgl-opengl-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-glfw/3.2.3/lwjgl-glfw-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-jemalloc/3.2.3/lwjgl-jemalloc-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-openvr/3.2.3/lwjgl-openvr-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-xxhash/3.2.3/lwjgl-xxhash-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-remotery/3.2.3/lwjgl-remotery-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-remotery/3.2.3/lwjgl-remotery-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/core/jackson-databind/2.9.10.1/jackson-databind-2.9.10.1.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/core/jackson-annotations/2.9.10/jackson-annotations-2.9.10.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/core/jackson-core/2.9.10/jackson-core-2.9.10.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/module/jackson-module-kotlin/2.9.9/jackson-module-kotlin-2.9.9.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/dataformat/jackson-dataformat-yaml/2.9.9/jackson-dataformat-yaml-2.9.9.jar:/home/aryaman/.m2/repository/graphics/scenery/spirvcrossj/0.7.0-1.1.106.0/spirvcrossj-0.7.0-1.1.106.0.jar:/home/aryaman/.m2/repository/graphics/scenery/spirvcrossj/0.7.0-1.1.106.0/spirvcrossj-0.7.0-1.1.106.0-natives-linux.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-junit/1.3.61/kotlin-test-junit-1.3.61.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-annotations-common/1.3.61/kotlin-test-annotations-common-1.3.61.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test/1.3.61/kotlin-test-1.3.61.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-common/1.3.61/kotlin-test-common-1.3.61.jar:/home/aryaman/.m2/repository/org/zeromq/jeromq/0.4.3/jeromq-0.4.3.jar:/home/aryaman/.m2/repository/org/zeromq/jnacl/0.1.0/jnacl-0.1.0.jar:/home/aryaman/.m2/repository/com/esotericsoftware/kryo/4.0.2/kryo-4.0.2.jar:/home/aryaman/.m2/repository/com/esotericsoftware/reflectasm/1.11.3/reflectasm-1.11.3.jar:/home/aryaman/.m2/repository/org/ow2/asm/asm/5.0.4/asm-5.0.4.jar:/home/aryaman/.m2/repository/com/esotericsoftware/minlog/1.3.0/minlog-1.3.0.jar:/home/aryaman/.m2/repository/org/objenesis/objenesis/2.5.1/objenesis-2.5.1.jar:/home/aryaman/.m2/repository/org/msgpack/msgpack-core/0.8.16/msgpack-core-0.8.16.jar:/home/aryaman/.m2/repository/org/msgpack/jackson-dataformat-msgpack/0.8.16/jackson-dataformat-msgpack-0.8.16.jar:/home/aryaman/.m2/repository/graphics/scenery/jvrpn/1.1.0/jvrpn-1.1.0.jar:/home/aryaman/.m2/repository/graphics/scenery/jvrpn/1.1.0/jvrpn-1.1.0-natives-linux.jar:/home/aryaman/.m2/repository/io/scif/scifio/0.37.3/scifio-0.37.3.jar:/home/aryaman/.m2/repository/io/scif/scifio-jai-imageio/1.1.1/scifio-jai-imageio-1.1.1.jar:/home/aryaman/.m2/repository/net/imagej/imagej-common/0.28.2/imagej-common-0.28.2.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-roi/0.7.0/imglib2-roi-0.7.0.jar:/home/aryaman/.m2/repository/edu/ucar/udunits/4.3.18/udunits-4.3.18.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2/5.6.3/imglib2-5.6.3.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-cache/1.0.0-beta-11/imglib2-cache-1.0.0-beta-11.jar:/home/aryaman/.m2/repository/com/github/ben-manes/caffeine/caffeine/2.4.0/caffeine-2.4.0.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp-presets/ffmpeg/4.1-1.4.4/ffmpeg-4.1-1.4.4.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp/1.4.4/javacpp-1.4.4.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp-presets/ffmpeg/4.1-1.4.4/ffmpeg-4.1-1.4.4-linux-x86.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp-presets/ffmpeg/4.1-1.4.4/ffmpeg-4.1-1.4.4-linux-x86_64.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp-presets/ffmpeg/4.1-1.4.4/ffmpeg-4.1-1.4.4-macosx-x86_64.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp-presets/ffmpeg/4.1-1.4.4/ffmpeg-4.1-1.4.4-windows-x86_64.jar:/home/aryaman/.m2/repository/org/reflections/reflections/0.9.11/reflections-0.9.11.jar:/home/aryaman/.m2/repository/com/google/guava/guava/21.0/guava-21.0.jar:/home/aryaman/.m2/repository/org/javassist/javassist/3.24.0-GA/javassist-3.24.0-GA.jar:/home/aryaman/.m2/repository/sc/fiji/bigvolumeviewer/0.1.6/bigvolumeviewer-0.1.6.jar:/home/aryaman/.m2/repository/org/joml/joml/1.9.12/joml-1.9.12.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-math3/3.6.1/commons-math3-3.6.1.jar:/home/aryaman/.m2/repository/sc/fiji/bigdataviewer-core/5.1.0/bigdataviewer-core-5.1.0.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-realtransform/2.2.1/imglib2-realtransform-2.2.1.jar:/home/aryaman/.m2/repository/gov/nist/math/jama/1.0.3/jama-1.0.3.jar:/home/aryaman/.m2/repository/jitk/jitk-tps/3.0.1/jitk-tps-3.0.1.jar:/home/aryaman/.m2/repository/com/googlecode/efficient-java-matrix-library/ejml/0.24/ejml-0.24.jar:/home/aryaman/.m2/repository/log4j/log4j/1.2.17/log4j-1.2.17.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-ui/2.0.0/imglib2-ui-2.0.0.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-algorithm/0.11.1/imglib2-algorithm-0.11.1.jar:/home/aryaman/.m2/repository/org/ojalgo/ojalgo/45.1.1/ojalgo-45.1.1.jar:/home/aryaman/.m2/repository/sc/fiji/spim_data/2.2.2/spim_data-2.2.2.jar:/home/aryaman/.m2/repository/cisd/jhdf5/14.12.6/jhdf5-14.12.6.jar:/home/aryaman/.m2/repository/commons-lang/commons-lang/2.6/commons-lang-2.6.jar:/home/aryaman/.m2/repository/org/jdom/jdom2/2.0.6/jdom2-2.0.6.jar:/home/aryaman/.m2/repository/sc/fiji/bigdataviewer-vistools/1.0.0-beta-11/bigdataviewer-vistools-1.0.0-beta-11.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-ij/2.0.0-beta-44/imglib2-ij-2.0.0-beta-44.jar:/home/aryaman/.m2/repository/net/imagej/ij/1.52o/ij-1.52o.jar:/home/aryaman/.m2/repository/net/imagej/imagej/2.0.0-rc-71/imagej-2.0.0-rc-71.jar:/home/aryaman/.m2/repository/net/imagej/imagej-notebook/0.7.1/imagej-notebook-0.7.1.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-text/1.6/commons-text-1.6.jar:/home/aryaman/.m2/repository/org/knowm/xchart/xchart/3.5.4/xchart-3.5.4.jar:/home/aryaman/.m2/repository/de/erichseifert/vectorgraphics2d/VectorGraphics2D/0.13/VectorGraphics2D-0.13.jar:/home/aryaman/.m2/repository/net/imagej/imagej-ops/0.43.2/imagej-ops-0.43.2.jar:/home/aryaman/.m2/repository/net/imagej/imagej-mesh/0.7.1/imagej-mesh-0.7.1.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-algorithm-fft/0.2.0/imglib2-algorithm-fft-0.2.0.jar:/home/aryaman/.m2/repository/edu/mines/mines-jtk/20151125/mines-jtk-20151125.jar:/home/aryaman/.m2/repository/net/imagej/imagej-updater/0.9.3/imagej-updater-0.9.3.jar:/home/aryaman/.m2/repository/net/imagej/imagej-plugins-batch/0.1.1/imagej-plugins-batch-0.1.1.jar:/home/aryaman/.m2/repository/net/imagej/imagej-deprecated/0.1.2/imagej-deprecated-0.1.2.jar:/home/aryaman/.m2/repository/net/imagej/imagej-plugins-commands/0.8.1/imagej-plugins-commands-0.8.1.jar:/home/aryaman/.m2/repository/net/iharder/base64/2.3.8/base64-2.3.8.jar:/home/aryaman/.m2/repository/net/imagej/imagej-plugins-tools/0.3.1/imagej-plugins-tools-0.3.1.jar:/home/aryaman/.m2/repository/net/imagej/imagej-plugins-uploader-ssh/0.3.2/imagej-plugins-uploader-ssh-0.3.2.jar:/home/aryaman/.m2/repository/com/jcraft/jsch/0.1.55/jsch-0.1.55.jar:/home/aryaman/.m2/repository/net/imagej/imagej-plugins-uploader-webdav/0.2.2/imagej-plugins-uploader-webdav-0.2.2.jar:/home/aryaman/.m2/repository/net/imagej/imagej-scripting/0.8.2/imagej-scripting-0.8.2.jar:/home/aryaman/.m2/repository/org/scijava/scripting-beanshell/0.3.3/scripting-beanshell-0.3.3.jar:/home/aryaman/.m2/repository/org/apache-extras/beanshell/bsh/2.0b6/bsh-2.0b6.jar:/home/aryaman/.m2/repository/org/scijava/scripting-clojure/0.1.6/scripting-clojure-0.1.6.jar:/home/aryaman/.m2/repository/org/clojure/clojure/1.8.0/clojure-1.8.0.jar:/home/aryaman/.m2/repository/org/scijava/scripting-groovy/0.2.7/scripting-groovy-0.2.7.jar:/home/aryaman/.m2/repository/org/codehaus/groovy/groovy/2.4.8/groovy-2.4.8.jar:/home/aryaman/.m2/repository/org/apache/ivy/ivy/2.4.0/ivy-2.4.0.jar:/home/aryaman/.m2/repository/org/codehaus/gpars/gpars/1.2.1/gpars-1.2.1.jar:/home/aryaman/.m2/repository/org/multiverse/multiverse-core/0.7.0/multiverse-core-0.7.0.jar:/home/aryaman/.m2/repository/org/codehaus/jsr166-mirror/jsr166y/1.7.0/jsr166y-1.7.0.jar:/home/aryaman/.m2/repository/org/scijava/scripting-jruby/0.3.0/scripting-jruby-0.3.0.jar:/home/aryaman/.m2/repository/org/jruby/jruby-core/9.1.17.0/jruby-core-9.1.17.0.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-netdb/1.1.6/jnr-netdb-1.1.6.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-enxio/0.16/jnr-enxio-0.16.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-x86asm/1.0.2/jnr-x86asm-1.0.2.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-unixsocket/0.17/jnr-unixsocket-0.17.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-posix/3.0.44/jnr-posix-3.0.44.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-constants/0.9.9/jnr-constants-0.9.9.jar:/home/aryaman/.m2/repository/com/github/jnr/jffi/1.2.16/jffi-1.2.16.jar:/home/aryaman/.m2/repository/com/github/jnr/jffi/1.2.16/jffi-1.2.16-native.jar:/home/aryaman/.m2/repository/org/jruby/joni/joni/2.1.15/joni-2.1.15.jar:/home/aryaman/.m2/repository/org/jruby/extras/bytelist/1.0.15/bytelist-1.0.15.jar:/home/aryaman/.m2/repository/org/jruby/jcodings/jcodings/1.0.27/jcodings-1.0.27.jar:/home/aryaman/.m2/repository/org/jruby/dirgra/0.3/dirgra-0.3.jar:/home/aryaman/.m2/repository/com/headius/invokebinder/1.10/invokebinder-1.10.jar:/home/aryaman/.m2/repository/com/headius/options/1.4/options-1.4.jar:/home/aryaman/.m2/repository/com/headius/unsafe-fences/1.0/unsafe-fences-1.0.jar:/home/aryaman/.m2/repository/com/jcraft/jzlib/1.1.3/jzlib-1.1.3.jar:/home/aryaman/.m2/repository/com/martiansoftware/nailgun-server/0.9.1/nailgun-server-0.9.1.jar:/home/aryaman/.m2/repository/joda-time/joda-time/2.10.1/joda-time-2.10.1.jar:/home/aryaman/.m2/repository/com/headius/modulator/1.0/modulator-1.0.jar:/home/aryaman/.m2/repository/org/jruby/jruby-stdlib/9.1.17.0/jruby-stdlib-9.1.17.0.jar:/home/aryaman/.m2/repository/org/scijava/scripting-jython/0.4.2/scripting-jython-0.4.2.jar:/home/aryaman/.m2/repository/org/scijava/jython-shaded/2.7.1.1/jython-shaded-2.7.1.1.jar:/home/aryaman/.m2/repository/org/scijava/scripting-renjin/0.2.2/scripting-renjin-0.2.2.jar:/home/aryaman/.m2/repository/org/renjin/renjin-script-engine/0.8.1906/renjin-script-engine-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/renjin-core/0.8.1906/renjin-core-0.8.1906.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-math/2.2/commons-math-2.2.jar:/home/aryaman/.m2/repository/org/renjin/gcc-runtime/0.8.1906/gcc-runtime-0.8.1906.jar:/home/aryaman/.m2/repository/org/netlib/netlib-java/0.9.3-renjin-patched-2/netlib-java-0.9.3-renjin-patched-2.jar:/home/aryaman/.m2/repository/org/netlib/lapack/0.8/lapack-0.8.jar:/home/aryaman/.m2/repository/org/netlib/f2jutil/0.8/f2jutil-0.8.jar:/home/aryaman/.m2/repository/org/netlib/xerbla/0.8/xerbla-0.8.jar:/home/aryaman/.m2/repository/org/netlib/blas/0.8/blas-0.8.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-vfs2/2.0/commons-vfs2-2.0.jar:/home/aryaman/.m2/repository/commons-logging/commons-logging/1.2/commons-logging-1.2.jar:/home/aryaman/.m2/repository/org/apache/maven/scm/maven-scm-api/1.4/maven-scm-api-1.4.jar:/home/aryaman/.m2/repository/org/codehaus/plexus/plexus-utils/1.5.6/plexus-utils-1.5.6.jar:/home/aryaman/.m2/repository/org/apache/maven/scm/maven-scm-provider-svnexe/1.4/maven-scm-provider-svnexe-1.4.jar:/home/aryaman/.m2/repository/org/apache/maven/scm/maven-scm-provider-svn-commons/1.4/maven-scm-provider-svn-commons-1.4.jar:/home/aryaman/.m2/repository/regexp/regexp/1.3/regexp-1.3.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-compress/1.4.1/commons-compress-1.4.1.jar:/home/aryaman/.m2/repository/org/tukaani/xz/1.0/xz-1.0.jar:/home/aryaman/.m2/repository/edu/emory/mathcs/jtransforms/2.4/jtransforms-2.4.jar:/home/aryaman/.m2/repository/net/sf/jung/jung-api/2.0.1/jung-api-2.0.1.jar:/home/aryaman/.m2/repository/net/sourceforge/collections/collections-generic/4.01/collections-generic-4.01.jar:/home/aryaman/.m2/repository/net/sf/jung/jung-graph-impl/2.0.1/jung-graph-impl-2.0.1.jar:/home/aryaman/.m2/repository/org/ow2/asm/asm-util/5.0.4/asm-util-5.0.4.jar:/home/aryaman/.m2/repository/org/ow2/asm/asm-tree/5.0.4/asm-tree-5.0.4.jar:/home/aryaman/.m2/repository/org/ow2/asm/asm-commons/5.0.4/asm-commons-5.0.4.jar:/home/aryaman/.m2/repository/com/sun/codemodel/codemodel/2.6/codemodel-2.6.jar:/home/aryaman/.m2/repository/org/renjin/stats/0.8.1906/stats-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/renjin-appl/0.8.1906/renjin-appl-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/renjin-gnur-runtime/0.8.1906/renjin-gnur-runtime-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/methods/0.8.1906/methods-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/datasets/0.8.1906/datasets-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/utils/0.8.1906/utils-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/grDevices/0.8.1906/grDevices-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/graphics/0.8.1906/graphics-0.8.1906.jar:/home/aryaman/.m2/repository/org/scijava/scripting-scala/0.2.1/scripting-scala-0.2.1.jar:/home/aryaman/.m2/repository/org/scala-lang/scala-compiler/2.12.1/scala-compiler-2.12.1.jar:/home/aryaman/.m2/repository/org/scala-lang/scala-library/2.12.1/scala-library-2.12.1.jar:/home/aryaman/.m2/repository/org/scala-lang/scala-reflect/2.12.1/scala-reflect-2.12.1.jar:/home/aryaman/.m2/repository/org/scala-lang/modules/scala-xml_2.12/1.0.6/scala-xml_2.12-1.0.6.jar:/home/aryaman/.m2/repository/net/imagej/imagej-ui-swing/0.22.0/imagej-ui-swing-0.22.0.jar:/home/aryaman/.m2/repository/org/jfree/jfreechart/1.5.0/jfreechart-1.5.0.jar:/home/aryaman/.m2/repository/org/jhotdraw/jhotdraw/7.6.0/jhotdraw-7.6.0.jar:/home/aryaman/.m2/repository/net/imagej/imagej-ui-awt/0.3.1/imagej-ui-awt-0.3.1.jar:/home/aryaman/.m2/repository/org/scijava/scijava-plugins-commands/0.2.3/scijava-plugins-commands-0.2.3.jar:/home/aryaman/.m2/repository/org/scijava/scijava-plugins-platforms/0.3.1/scijava-plugins-platforms-0.3.1.jar:/home/aryaman/.m2/repository/org/scijava/scijava-plugins-text-markdown/0.1.3/scijava-plugins-text-markdown-0.1.3.jar:/home/aryaman/.m2/repository/org/markdownj/markdownj/0.3.0-1.0.2b4/markdownj-0.3.0-1.0.2b4.jar:/home/aryaman/.m2/repository/org/scijava/scijava-plugins-text-plain/0.1.3/scijava-plugins-text-plain-0.1.3.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-lang3/3.8.1/commons-lang3-3.8.1.jar:/home/aryaman/.m2/repository/org/antlr/ST4/4.0.8/ST4-4.0.8.jar:/home/aryaman/.m2/repository/org/antlr/antlr-runtime/3.5.2/antlr-runtime-3.5.2.jar:/home/aryaman/.m2/repository/com/github/lwjglx/lwjgl3-awt/45a85af/lwjgl3-awt-45a85af.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-jawt/3.2.3/lwjgl-jawt-3.2.3.jar:/home/aryaman/Repositories/scenery/mpi.jar:/home/aryaman/Repositories/scenery-insitu/target/test-classes:/home/aryaman/Repositories/scenery-insitu/target/classes:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-jdk8/1.3.21/kotlin-stdlib-jdk8-1.3.21.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib/1.3.21/kotlin-stdlib-1.3.21.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-common/1.3.21/kotlin-stdlib-common-1.3.21.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-jdk7/1.3.21/kotlin-stdlib-jdk7-1.3.21.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-reflect/1.3.21/kotlin-reflect-1.3.21.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlinx/kotlinx-coroutines-core/1.1.1/kotlinx-coroutines-core-1.1.1.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlinx/kotlinx-coroutines-core-common/1.1.1/kotlinx-coroutines-core-common-1.1.1.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-junit/1.3.21/kotlin-test-junit-1.3.21.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-annotations-common/1.3.21/kotlin-test-annotations-common-1.3.21.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test/1.3.21/kotlin-test-1.3.21.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-common/1.3.21/kotlin-test-common-1.3.21.jar:/home/aryaman/.m2/repository/org/scijava/scijava-common/2.76.1/scijava-common-2.76.1.jar:/home/aryaman/.m2/repository/org/scijava/script-editor/0.4.0/script-editor-0.4.0.jar:/home/aryaman/.m2/repository/org/scijava/scijava-table/0.2.0/scijava-table-0.2.0.jar:/home/aryaman/.m2/repository/org/scijava/scijava-search/0.5.0/scijava-search-0.5.0.jar:/home/aryaman/.m2/repository/net/imagej/imagej-common/0.28.0/imagej-common-0.28.0.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-realtransform/2.1.0/imglib2-realtransform-2.1.0.jar:/home/aryaman/.m2/repository/jitk/jitk-tps/3.0.0/jitk-tps-3.0.0.jar:/home/aryaman/.m2/repository/net/imagej/ij/1.52i/ij-1.52i.jar:/home/aryaman/.m2/repository/net/imagej/imagej/2.0.0-rc-69/imagej-2.0.0-rc-69.jar:/home/aryaman/.m2/repository/net/imagej/imagej-notebook/0.3.0/imagej-notebook-0.3.0.jar:/home/aryaman/.m2/repository/net/imagej/imagej-ops/0.43.1/imagej-ops-0.43.1.jar:/home/aryaman/.m2/repository/org/scijava/jython-shaded/2.7.1/jython-shaded-2.7.1.jar:/home/aryaman/.m2/repository/org/msgpack/msgpack-core/0.8.20/msgpack-core-0.8.20.jar:/home/aryaman/.m2/repository/org/msgpack/jackson-dataformat-msgpack/0.8.20/jackson-dataformat-msgpack-0.8.20.jar";
-    //"-Djava.class.path=/home/aryaman/jdk8u242-b08/jre/lib/charsets.jar:/home/aryaman/jdk8u242-b08/jre/lib/ext/cldrdata.jar:/home/aryaman/jdk8u242-b08/jre/lib/ext/dnsns.jar:/home/aryaman/jdk8u242-b08/jre/lib/ext/jaccess.jar:/home/aryaman/jdk8u242-b08/jre/lib/ext/localedata.jar:/home/aryaman/jdk8u242-b08/jre/lib/ext/nashorn.jar:/home/aryaman/jdk8u242-b08/jre/lib/ext/sunec.jar:/home/aryaman/jdk8u242-b08/jre/lib/ext/sunjce_provider.jar:/home/aryaman/jdk8u242-b08/jre/lib/ext/sunpkcs11.jar:/home/aryaman/jdk8u242-b08/jre/lib/ext/zipfs.jar:/home/aryaman/jdk8u242-b08/jre/lib/jce.jar:/home/aryaman/jdk8u242-b08/jre/lib/jsse.jar:/home/aryaman/jdk8u242-b08/jre/lib/management-agent.jar:/home/aryaman/jdk8u242-b08/jre/lib/resources.jar:/home/aryaman/jdk8u242-b08/jre/lib/rt.jar:/home/aryaman/Repositories/scenery/target/test-classes:/home/aryaman/Repositories/scenery/target/classes:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-jdk8/1.3.61/kotlin-stdlib-jdk8-1.3.61.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib/1.3.61/kotlin-stdlib-1.3.61.jar:/home/aryaman/.m2/repository/org/jetbrains/annotations/13.0/annotations-13.0.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-jdk7/1.3.61/kotlin-stdlib-jdk7-1.3.61.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-reflect/1.3.61/kotlin-reflect-1.3.61.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlinx/kotlinx-coroutines-core/1.3.2/kotlinx-coroutines-core-1.3.2.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-common/1.3.50/kotlin-stdlib-common-1.3.50.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-linux-amd64.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-linux-i586.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-macosx-universal.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-windows-amd64.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-linux-amd64.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-linux-i586.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-macosx-universal.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-windows-amd64.jar:/home/aryaman/.m2/repository/junit/junit/4.12/junit-4.12.jar:/home/aryaman/.m2/repository/org/hamcrest/hamcrest-core/1.3/hamcrest-core-1.3.jar:/home/aryaman/.m2/repository/org/slf4j/slf4j-api/1.7.25/slf4j-api-1.7.25.jar:/home/aryaman/.m2/repository/org/slf4j/slf4j-simple/1.7.25/slf4j-simple-1.7.25.jar:/home/aryaman/.m2/repository/net/clearvolume/cleargl/2.2.10/cleargl-2.2.10.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt-main/2.3.2/gluegen-rt-main-2.3.2.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-android-aarch64.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-android-armv6.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-linux-armv6.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-linux-armv6hf.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-solaris-amd64.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-solaris-i586.jar:/home/aryaman/.m2/repository/org/jogamp/gluegen/gluegen-rt/2.3.2/gluegen-rt-2.3.2-natives-windows-i586.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all-main/2.3.2/jogl-all-main-2.3.2.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-android-aarch64.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-android-armv6.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-linux-armv6.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-linux-armv6hf.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-solaris-amd64.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-solaris-i586.jar:/home/aryaman/.m2/repository/org/jogamp/jogl/jogl-all/2.3.2/jogl-all-2.3.2-natives-windows-i586.jar:/home/aryaman/.m2/repository/net/clearcontrol/coremem/0.4.6/coremem-0.4.6.jar:/home/aryaman/.m2/repository/com/nativelibs4java/bridj/0.7.0/bridj-0.7.0.jar:/home/aryaman/.m2/repository/net/java/jinput/jinput/2.0.9/jinput-2.0.9.jar:/home/aryaman/.m2/repository/net/java/jinput/jinput/2.0.9/jinput-2.0.9-natives-all.jar:/home/aryaman/.m2/repository/org/scijava/scijava-common/2.77.0/scijava-common-2.77.0.jar:/home/aryaman/.m2/repository/org/scijava/parsington/1.0.4/parsington-1.0.4.jar:/home/aryaman/.m2/repository/org/bushe/eventbus/1.4/eventbus-1.4.jar:/home/aryaman/.m2/repository/org/scijava/script-editor/0.5.3/script-editor-0.5.3.jar:/home/aryaman/.m2/repository/org/scijava/scripting-java/0.4.1/scripting-java-0.4.1.jar:/home/aryaman/.m2/repository/org/scijava/minimaven/2.2.1/minimaven-2.2.1.jar:/home/aryaman/.m2/repository/com/fifesoft/rsyntaxtextarea/2.6.1/rsyntaxtextarea-2.6.1.jar:/home/aryaman/.m2/repository/com/fifesoft/languagesupport/2.6.0/languagesupport-2.6.0.jar:/home/aryaman/.m2/repository/com/fifesoft/autocomplete/2.6.1/autocomplete-2.6.1.jar:/home/aryaman/.m2/repository/org/mozilla/rhino/1.7.6/rhino-1.7.6.jar:/home/aryaman/.m2/repository/com/miglayout/miglayout/3.7.4/miglayout-3.7.4-swing.jar:/home/aryaman/.m2/repository/org/scijava/batch-processor/0.3.0/batch-processor-0.3.0.jar:/home/aryaman/.m2/repository/org/scijava/scijava-table/0.4.0/scijava-table-0.4.0.jar:/home/aryaman/.m2/repository/org/scijava/scijava-ui-swing/0.12.0/scijava-ui-swing-0.12.0.jar:/home/aryaman/.m2/repository/org/scijava/scijava-ui-awt/0.1.6/scijava-ui-awt-0.1.6.jar:/home/aryaman/.m2/repository/org/scijava/swing-checkbox-tree/1.0.2/swing-checkbox-tree-1.0.2.jar:/home/aryaman/.m2/repository/com/github/sbridges/object-inspector/object-inspector/0.1/object-inspector-0.1.jar:/home/aryaman/.m2/repository/net/sourceforge/jdatepicker/jdatepicker/1.3.2/jdatepicker-1.3.2.jar:/home/aryaman/.m2/repository/org/scijava/scijava-search/0.6.0/scijava-search-0.6.0.jar:/home/aryaman/.m2/repository/org/ocpsoft/prettytime/prettytime/4.0.1.Final/prettytime-4.0.1.Final.jar:/home/aryaman/.m2/repository/commons-io/commons-io/2.6/commons-io-2.6.jar:/home/aryaman/.m2/repository/org/scijava/ui-behaviour/1.7.4/ui-behaviour-1.7.4.jar:/home/aryaman/.m2/repository/com/google/code/gson/gson/2.8.5/gson-2.8.5.jar:/home/aryaman/.m2/repository/org/yaml/snakeyaml/1.23/snakeyaml-1.23.jar:/home/aryaman/.m2/repository/net/sf/trove4j/trove4j/3.0.3/trove4j-3.0.3.jar:/home/aryaman/.m2/repository/org/scijava/scripting-javascript/0.4.4/scripting-javascript-0.4.4.jar:/home/aryaman/.m2/repository/net/java/dev/jna/jna/4.5.2/jna-4.5.2.jar:/home/aryaman/.m2/repository/net/java/dev/jna/jna-platform/4.5.2/jna-platform-4.5.2.jar:/home/aryaman/.m2/repository/org/jocl/jocl/2.0.1/jocl-2.0.1.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl/3.2.3/lwjgl-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-glfw/3.2.3/lwjgl-glfw-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-jemalloc/3.2.3/lwjgl-jemalloc-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-vulkan/3.2.3/lwjgl-vulkan-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-opengl/3.2.3/lwjgl-opengl-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-openvr/3.2.3/lwjgl-openvr-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-xxhash/3.2.3/lwjgl-xxhash-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl/3.2.3/lwjgl-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-opengl/3.2.3/lwjgl-opengl-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-glfw/3.2.3/lwjgl-glfw-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-jemalloc/3.2.3/lwjgl-jemalloc-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-openvr/3.2.3/lwjgl-openvr-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-xxhash/3.2.3/lwjgl-xxhash-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-remotery/3.2.3/lwjgl-remotery-3.2.3.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-remotery/3.2.3/lwjgl-remotery-3.2.3-natives-linux.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/core/jackson-databind/2.9.10.1/jackson-databind-2.9.10.1.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/core/jackson-annotations/2.9.10/jackson-annotations-2.9.10.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/core/jackson-core/2.9.10/jackson-core-2.9.10.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/module/jackson-module-kotlin/2.9.9/jackson-module-kotlin-2.9.9.jar:/home/aryaman/.m2/repository/com/fasterxml/jackson/dataformat/jackson-dataformat-yaml/2.9.9/jackson-dataformat-yaml-2.9.9.jar:/home/aryaman/.m2/repository/graphics/scenery/spirvcrossj/0.7.0-1.1.106.0/spirvcrossj-0.7.0-1.1.106.0.jar:/home/aryaman/.m2/repository/graphics/scenery/spirvcrossj/0.7.0-1.1.106.0/spirvcrossj-0.7.0-1.1.106.0-natives-linux.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-junit/1.3.61/kotlin-test-junit-1.3.61.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-annotations-common/1.3.61/kotlin-test-annotations-common-1.3.61.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test/1.3.61/kotlin-test-1.3.61.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-common/1.3.61/kotlin-test-common-1.3.61.jar:/home/aryaman/.m2/repository/org/zeromq/jeromq/0.4.3/jeromq-0.4.3.jar:/home/aryaman/.m2/repository/org/zeromq/jnacl/0.1.0/jnacl-0.1.0.jar:/home/aryaman/.m2/repository/com/esotericsoftware/kryo/4.0.2/kryo-4.0.2.jar:/home/aryaman/.m2/repository/com/esotericsoftware/reflectasm/1.11.3/reflectasm-1.11.3.jar:/home/aryaman/.m2/repository/org/ow2/asm/asm/5.0.4/asm-5.0.4.jar:/home/aryaman/.m2/repository/com/esotericsoftware/minlog/1.3.0/minlog-1.3.0.jar:/home/aryaman/.m2/repository/org/objenesis/objenesis/2.5.1/objenesis-2.5.1.jar:/home/aryaman/.m2/repository/org/msgpack/msgpack-core/0.8.16/msgpack-core-0.8.16.jar:/home/aryaman/.m2/repository/org/msgpack/jackson-dataformat-msgpack/0.8.16/jackson-dataformat-msgpack-0.8.16.jar:/home/aryaman/.m2/repository/graphics/scenery/jvrpn/1.1.0/jvrpn-1.1.0.jar:/home/aryaman/.m2/repository/graphics/scenery/jvrpn/1.1.0/jvrpn-1.1.0-natives-linux.jar:/home/aryaman/.m2/repository/io/scif/scifio/0.37.3/scifio-0.37.3.jar:/home/aryaman/.m2/repository/io/scif/scifio-jai-imageio/1.1.1/scifio-jai-imageio-1.1.1.jar:/home/aryaman/.m2/repository/net/imagej/imagej-common/0.28.2/imagej-common-0.28.2.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-roi/0.7.0/imglib2-roi-0.7.0.jar:/home/aryaman/.m2/repository/edu/ucar/udunits/4.3.18/udunits-4.3.18.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2/5.6.3/imglib2-5.6.3.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-cache/1.0.0-beta-11/imglib2-cache-1.0.0-beta-11.jar:/home/aryaman/.m2/repository/com/github/ben-manes/caffeine/caffeine/2.4.0/caffeine-2.4.0.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp-presets/ffmpeg/4.1-1.4.4/ffmpeg-4.1-1.4.4.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp/1.4.4/javacpp-1.4.4.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp-presets/ffmpeg/4.1-1.4.4/ffmpeg-4.1-1.4.4-linux-x86.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp-presets/ffmpeg/4.1-1.4.4/ffmpeg-4.1-1.4.4-linux-x86_64.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp-presets/ffmpeg/4.1-1.4.4/ffmpeg-4.1-1.4.4-macosx-x86_64.jar:/home/aryaman/.m2/repository/org/bytedeco/javacpp-presets/ffmpeg/4.1-1.4.4/ffmpeg-4.1-1.4.4-windows-x86_64.jar:/home/aryaman/.m2/repository/org/reflections/reflections/0.9.11/reflections-0.9.11.jar:/home/aryaman/.m2/repository/com/google/guava/guava/21.0/guava-21.0.jar:/home/aryaman/.m2/repository/org/javassist/javassist/3.24.0-GA/javassist-3.24.0-GA.jar:/home/aryaman/.m2/repository/sc/fiji/bigvolumeviewer/0.1.6/bigvolumeviewer-0.1.6.jar:/home/aryaman/.m2/repository/org/joml/joml/1.9.12/joml-1.9.12.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-math3/3.6.1/commons-math3-3.6.1.jar:/home/aryaman/.m2/repository/sc/fiji/bigdataviewer-core/5.1.0/bigdataviewer-core-5.1.0.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-realtransform/2.2.1/imglib2-realtransform-2.2.1.jar:/home/aryaman/.m2/repository/gov/nist/math/jama/1.0.3/jama-1.0.3.jar:/home/aryaman/.m2/repository/jitk/jitk-tps/3.0.1/jitk-tps-3.0.1.jar:/home/aryaman/.m2/repository/com/googlecode/efficient-java-matrix-library/ejml/0.24/ejml-0.24.jar:/home/aryaman/.m2/repository/log4j/log4j/1.2.17/log4j-1.2.17.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-ui/2.0.0/imglib2-ui-2.0.0.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-algorithm/0.11.1/imglib2-algorithm-0.11.1.jar:/home/aryaman/.m2/repository/org/ojalgo/ojalgo/45.1.1/ojalgo-45.1.1.jar:/home/aryaman/.m2/repository/sc/fiji/spim_data/2.2.2/spim_data-2.2.2.jar:/home/aryaman/.m2/repository/cisd/jhdf5/14.12.6/jhdf5-14.12.6.jar:/home/aryaman/.m2/repository/commons-lang/commons-lang/2.6/commons-lang-2.6.jar:/home/aryaman/.m2/repository/org/jdom/jdom2/2.0.6/jdom2-2.0.6.jar:/home/aryaman/.m2/repository/sc/fiji/bigdataviewer-vistools/1.0.0-beta-11/bigdataviewer-vistools-1.0.0-beta-11.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-ij/2.0.0-beta-44/imglib2-ij-2.0.0-beta-44.jar:/home/aryaman/.m2/repository/net/imagej/ij/1.52o/ij-1.52o.jar:/home/aryaman/.m2/repository/net/imagej/imagej/2.0.0-rc-71/imagej-2.0.0-rc-71.jar:/home/aryaman/.m2/repository/net/imagej/imagej-notebook/0.7.1/imagej-notebook-0.7.1.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-text/1.6/commons-text-1.6.jar:/home/aryaman/.m2/repository/org/knowm/xchart/xchart/3.5.4/xchart-3.5.4.jar:/home/aryaman/.m2/repository/de/erichseifert/vectorgraphics2d/VectorGraphics2D/0.13/VectorGraphics2D-0.13.jar:/home/aryaman/.m2/repository/net/imagej/imagej-ops/0.43.2/imagej-ops-0.43.2.jar:/home/aryaman/.m2/repository/net/imagej/imagej-mesh/0.7.1/imagej-mesh-0.7.1.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-algorithm-fft/0.2.0/imglib2-algorithm-fft-0.2.0.jar:/home/aryaman/.m2/repository/edu/mines/mines-jtk/20151125/mines-jtk-20151125.jar:/home/aryaman/.m2/repository/net/imagej/imagej-updater/0.9.3/imagej-updater-0.9.3.jar:/home/aryaman/.m2/repository/net/imagej/imagej-plugins-batch/0.1.1/imagej-plugins-batch-0.1.1.jar:/home/aryaman/.m2/repository/net/imagej/imagej-deprecated/0.1.2/imagej-deprecated-0.1.2.jar:/home/aryaman/.m2/repository/net/imagej/imagej-plugins-commands/0.8.1/imagej-plugins-commands-0.8.1.jar:/home/aryaman/.m2/repository/net/iharder/base64/2.3.8/base64-2.3.8.jar:/home/aryaman/.m2/repository/net/imagej/imagej-plugins-tools/0.3.1/imagej-plugins-tools-0.3.1.jar:/home/aryaman/.m2/repository/net/imagej/imagej-plugins-uploader-ssh/0.3.2/imagej-plugins-uploader-ssh-0.3.2.jar:/home/aryaman/.m2/repository/com/jcraft/jsch/0.1.55/jsch-0.1.55.jar:/home/aryaman/.m2/repository/net/imagej/imagej-plugins-uploader-webdav/0.2.2/imagej-plugins-uploader-webdav-0.2.2.jar:/home/aryaman/.m2/repository/net/imagej/imagej-scripting/0.8.2/imagej-scripting-0.8.2.jar:/home/aryaman/.m2/repository/org/scijava/scripting-beanshell/0.3.3/scripting-beanshell-0.3.3.jar:/home/aryaman/.m2/repository/org/apache-extras/beanshell/bsh/2.0b6/bsh-2.0b6.jar:/home/aryaman/.m2/repository/org/scijava/scripting-clojure/0.1.6/scripting-clojure-0.1.6.jar:/home/aryaman/.m2/repository/org/clojure/clojure/1.8.0/clojure-1.8.0.jar:/home/aryaman/.m2/repository/org/scijava/scripting-groovy/0.2.7/scripting-groovy-0.2.7.jar:/home/aryaman/.m2/repository/org/codehaus/groovy/groovy/2.4.8/groovy-2.4.8.jar:/home/aryaman/.m2/repository/org/apache/ivy/ivy/2.4.0/ivy-2.4.0.jar:/home/aryaman/.m2/repository/org/codehaus/gpars/gpars/1.2.1/gpars-1.2.1.jar:/home/aryaman/.m2/repository/org/multiverse/multiverse-core/0.7.0/multiverse-core-0.7.0.jar:/home/aryaman/.m2/repository/org/codehaus/jsr166-mirror/jsr166y/1.7.0/jsr166y-1.7.0.jar:/home/aryaman/.m2/repository/org/scijava/scripting-jruby/0.3.0/scripting-jruby-0.3.0.jar:/home/aryaman/.m2/repository/org/jruby/jruby-core/9.1.17.0/jruby-core-9.1.17.0.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-netdb/1.1.6/jnr-netdb-1.1.6.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-enxio/0.16/jnr-enxio-0.16.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-x86asm/1.0.2/jnr-x86asm-1.0.2.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-unixsocket/0.17/jnr-unixsocket-0.17.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-posix/3.0.44/jnr-posix-3.0.44.jar:/home/aryaman/.m2/repository/com/github/jnr/jnr-constants/0.9.9/jnr-constants-0.9.9.jar:/home/aryaman/.m2/repository/com/github/jnr/jffi/1.2.16/jffi-1.2.16.jar:/home/aryaman/.m2/repository/com/github/jnr/jffi/1.2.16/jffi-1.2.16-native.jar:/home/aryaman/.m2/repository/org/jruby/joni/joni/2.1.15/joni-2.1.15.jar:/home/aryaman/.m2/repository/org/jruby/extras/bytelist/1.0.15/bytelist-1.0.15.jar:/home/aryaman/.m2/repository/org/jruby/jcodings/jcodings/1.0.27/jcodings-1.0.27.jar:/home/aryaman/.m2/repository/org/jruby/dirgra/0.3/dirgra-0.3.jar:/home/aryaman/.m2/repository/com/headius/invokebinder/1.10/invokebinder-1.10.jar:/home/aryaman/.m2/repository/com/headius/options/1.4/options-1.4.jar:/home/aryaman/.m2/repository/com/headius/unsafe-fences/1.0/unsafe-fences-1.0.jar:/home/aryaman/.m2/repository/com/jcraft/jzlib/1.1.3/jzlib-1.1.3.jar:/home/aryaman/.m2/repository/com/martiansoftware/nailgun-server/0.9.1/nailgun-server-0.9.1.jar:/home/aryaman/.m2/repository/joda-time/joda-time/2.10.1/joda-time-2.10.1.jar:/home/aryaman/.m2/repository/com/headius/modulator/1.0/modulator-1.0.jar:/home/aryaman/.m2/repository/org/jruby/jruby-stdlib/9.1.17.0/jruby-stdlib-9.1.17.0.jar:/home/aryaman/.m2/repository/org/scijava/scripting-jython/0.4.2/scripting-jython-0.4.2.jar:/home/aryaman/.m2/repository/org/scijava/jython-shaded/2.7.1.1/jython-shaded-2.7.1.1.jar:/home/aryaman/.m2/repository/org/scijava/scripting-renjin/0.2.2/scripting-renjin-0.2.2.jar:/home/aryaman/.m2/repository/org/renjin/renjin-script-engine/0.8.1906/renjin-script-engine-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/renjin-core/0.8.1906/renjin-core-0.8.1906.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-math/2.2/commons-math-2.2.jar:/home/aryaman/.m2/repository/org/renjin/gcc-runtime/0.8.1906/gcc-runtime-0.8.1906.jar:/home/aryaman/.m2/repository/org/netlib/netlib-java/0.9.3-renjin-patched-2/netlib-java-0.9.3-renjin-patched-2.jar:/home/aryaman/.m2/repository/org/netlib/lapack/0.8/lapack-0.8.jar:/home/aryaman/.m2/repository/org/netlib/f2jutil/0.8/f2jutil-0.8.jar:/home/aryaman/.m2/repository/org/netlib/xerbla/0.8/xerbla-0.8.jar:/home/aryaman/.m2/repository/org/netlib/blas/0.8/blas-0.8.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-vfs2/2.0/commons-vfs2-2.0.jar:/home/aryaman/.m2/repository/commons-logging/commons-logging/1.2/commons-logging-1.2.jar:/home/aryaman/.m2/repository/org/apache/maven/scm/maven-scm-api/1.4/maven-scm-api-1.4.jar:/home/aryaman/.m2/repository/org/codehaus/plexus/plexus-utils/1.5.6/plexus-utils-1.5.6.jar:/home/aryaman/.m2/repository/org/apache/maven/scm/maven-scm-provider-svnexe/1.4/maven-scm-provider-svnexe-1.4.jar:/home/aryaman/.m2/repository/org/apache/maven/scm/maven-scm-provider-svn-commons/1.4/maven-scm-provider-svn-commons-1.4.jar:/home/aryaman/.m2/repository/regexp/regexp/1.3/regexp-1.3.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-compress/1.4.1/commons-compress-1.4.1.jar:/home/aryaman/.m2/repository/org/tukaani/xz/1.0/xz-1.0.jar:/home/aryaman/.m2/repository/edu/emory/mathcs/jtransforms/2.4/jtransforms-2.4.jar:/home/aryaman/.m2/repository/net/sf/jung/jung-api/2.0.1/jung-api-2.0.1.jar:/home/aryaman/.m2/repository/net/sourceforge/collections/collections-generic/4.01/collections-generic-4.01.jar:/home/aryaman/.m2/repository/net/sf/jung/jung-graph-impl/2.0.1/jung-graph-impl-2.0.1.jar:/home/aryaman/.m2/repository/org/ow2/asm/asm-util/5.0.4/asm-util-5.0.4.jar:/home/aryaman/.m2/repository/org/ow2/asm/asm-tree/5.0.4/asm-tree-5.0.4.jar:/home/aryaman/.m2/repository/org/ow2/asm/asm-commons/5.0.4/asm-commons-5.0.4.jar:/home/aryaman/.m2/repository/com/sun/codemodel/codemodel/2.6/codemodel-2.6.jar:/home/aryaman/.m2/repository/org/renjin/stats/0.8.1906/stats-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/renjin-appl/0.8.1906/renjin-appl-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/renjin-gnur-runtime/0.8.1906/renjin-gnur-runtime-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/methods/0.8.1906/methods-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/datasets/0.8.1906/datasets-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/utils/0.8.1906/utils-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/grDevices/0.8.1906/grDevices-0.8.1906.jar:/home/aryaman/.m2/repository/org/renjin/graphics/0.8.1906/graphics-0.8.1906.jar:/home/aryaman/.m2/repository/org/scijava/scripting-scala/0.2.1/scripting-scala-0.2.1.jar:/home/aryaman/.m2/repository/org/scala-lang/scala-compiler/2.12.1/scala-compiler-2.12.1.jar:/home/aryaman/.m2/repository/org/scala-lang/scala-library/2.12.1/scala-library-2.12.1.jar:/home/aryaman/.m2/repository/org/scala-lang/scala-reflect/2.12.1/scala-reflect-2.12.1.jar:/home/aryaman/.m2/repository/org/scala-lang/modules/scala-xml_2.12/1.0.6/scala-xml_2.12-1.0.6.jar:/home/aryaman/.m2/repository/net/imagej/imagej-ui-swing/0.22.0/imagej-ui-swing-0.22.0.jar:/home/aryaman/.m2/repository/org/jfree/jfreechart/1.5.0/jfreechart-1.5.0.jar:/home/aryaman/.m2/repository/org/jhotdraw/jhotdraw/7.6.0/jhotdraw-7.6.0.jar:/home/aryaman/.m2/repository/net/imagej/imagej-ui-awt/0.3.1/imagej-ui-awt-0.3.1.jar:/home/aryaman/.m2/repository/org/scijava/scijava-plugins-commands/0.2.3/scijava-plugins-commands-0.2.3.jar:/home/aryaman/.m2/repository/org/scijava/scijava-plugins-platforms/0.3.1/scijava-plugins-platforms-0.3.1.jar:/home/aryaman/.m2/repository/org/scijava/scijava-plugins-text-markdown/0.1.3/scijava-plugins-text-markdown-0.1.3.jar:/home/aryaman/.m2/repository/org/markdownj/markdownj/0.3.0-1.0.2b4/markdownj-0.3.0-1.0.2b4.jar:/home/aryaman/.m2/repository/org/scijava/scijava-plugins-text-plain/0.1.3/scijava-plugins-text-plain-0.1.3.jar:/home/aryaman/.m2/repository/org/apache/commons/commons-lang3/3.8.1/commons-lang3-3.8.1.jar:/home/aryaman/.m2/repository/org/antlr/ST4/4.0.8/ST4-4.0.8.jar:/home/aryaman/.m2/repository/org/antlr/antlr-runtime/3.5.2/antlr-runtime-3.5.2.jar:/home/aryaman/.m2/repository/com/github/lwjglx/lwjgl3-awt/45a85af/lwjgl3-awt-45a85af.jar:/home/aryaman/.m2/repository/org/lwjgl/lwjgl-jawt/3.2.3/lwjgl-jawt-3.2.3.jar:/home/aryaman/Repositories/scenery/mpi.jar:/home/aryaman/Repositories/scenery-insitu/target/test-classes:/home/aryaman/Repositories/scenery-insitu/target/classes:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-jdk8/1.3.21/kotlin-stdlib-jdk8-1.3.21.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib/1.3.21/kotlin-stdlib-1.3.21.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-common/1.3.21/kotlin-stdlib-common-1.3.21.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-stdlib-jdk7/1.3.21/kotlin-stdlib-jdk7-1.3.21.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-reflect/1.3.21/kotlin-reflect-1.3.21.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlinx/kotlinx-coroutines-core/1.1.1/kotlinx-coroutines-core-1.1.1.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlinx/kotlinx-coroutines-core-common/1.1.1/kotlinx-coroutines-core-common-1.1.1.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-junit/1.3.21/kotlin-test-junit-1.3.21.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-annotations-common/1.3.21/kotlin-test-annotations-common-1.3.21.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test/1.3.21/kotlin-test-1.3.21.jar:/home/aryaman/.m2/repository/org/jetbrains/kotlin/kotlin-test-common/1.3.21/kotlin-test-common-1.3.21.jar:/home/aryaman/.m2/repository/org/scijava/scijava-common/2.76.1/scijava-common-2.76.1.jar:/home/aryaman/.m2/repository/org/scijava/script-editor/0.4.0/script-editor-0.4.0.jar:/home/aryaman/.m2/repository/org/scijava/scijava-table/0.2.0/scijava-table-0.2.0.jar:/home/aryaman/.m2/repository/org/scijava/scijava-search/0.5.0/scijava-search-0.5.0.jar:/home/aryaman/.m2/repository/net/imagej/imagej-common/0.28.0/imagej-common-0.28.0.jar:/home/aryaman/.m2/repository/net/imglib2/imglib2-realtransform/2.1.0/imglib2-realtransform-2.1.0.jar:/home/aryaman/.m2/repository/jitk/jitk-tps/3.0.0/jitk-tps-3.0.0.jar:/home/aryaman/.m2/repository/net/imagej/ij/1.52i/ij-1.52i.jar:/home/aryaman/.m2/repository/net/imagej/imagej/2.0.0-rc-69/imagej-2.0.0-rc-69.jar:/home/aryaman/.m2/repository/net/imagej/imagej-notebook/0.3.0/imagej-notebook-0.3.0.jar:/home/aryaman/.m2/repository/net/imagej/imagej-ops/0.43.1/imagej-ops-0.43.1.jar:/home/aryaman/.m2/repository/org/scijava/jython-shaded/2.7.1/jython-shaded-2.7.1.jar:/home/aryaman/.m2/repository/org/msgpack/msgpack-core/0.8.20/msgpack-core-0.8.20.jar:/home/aryaman/.m2/repository/org/msgpack/jackson-dataformat-msgpack/0.8.20/jackson-dataformat-msgpack-0.8.20.jar";
-    //    options[1].optionString = (char *)"-Djava.library.path=/Users/argupta/Documents/scenery-insitu/src/test/kotlin/graphics/scenery/insitu/";fscenery
+
     if(USE_VULKAN) {
         options[1].optionString = (char *)
                 "-Dscenery.Renderer=VulkanRenderer";
@@ -71,45 +87,34 @@ InVis::InVis(int wSize, int cPartners, MPI_Comm vComm, bool isHead) {
         options[1].optionString = (char *)
                 "-Dscenery.Renderer=OpenGLRenderer";
     }
-    if(isHead) {
-        options[2].optionString = (char *)
-                "-Dorg.slf4j.simpleLogger.defaultLogLevel=info";
+
+    options[2].optionString = (char *)
+            "-Dorg.slf4j.simpleLogger.defaultLogLevel=info";
 //        options[3].optionString = (char *)
 //                "-Dscenery.ENABLE_VULKAN_RENDERDOC_CAPTURE=1";
-    }
-    else {
-        options[2].optionString = (char *)
-                "-Dorg.slf4j.simpleLogger.defaultLogLevel=info";
-//        options[3].optionString = (char *)
-//                "-Dscenery.ENABLE_VULKAN_RENDERDOC_CAPTURE=0";
-    }
 
     options[3].optionString = (char *)
-            "-Dscenery.Headless=true";
+            "-Dscenery.Headless=false";
 
-    vm_args.version = JNI_VERSION_1_6;             // minimum Java version
-    vm_args.nOptions = 4;                          // number of options
+    vm_args.version = JNI_VERSION_1_6;
+    vm_args.nOptions = 4;
     vm_args.options = options;
-    vm_args.ignoreUnrecognized = false;     // invalid options make the JVM init fail
-    //=============== load and initialize Java VM and JNI interface =============
+    vm_args.ignoreUnrecognized = false;
+
     JNIEnv *env;
     jint rc = JNI_CreateJavaVM(&jvm, (void **) &env, &vm_args);  // YES !!
     std::cout<<"Hello world"<<std::endl;
-    delete options;    // we then no longer need the initialisation options.
+    delete options;
     if (rc != JNI_OK) {
         // TODO: error processing...
         std::cin.get();
         std::exit(EXIT_FAILURE);
     }
 
-    //=============== Display JVM version =======================================
     std::cout << "JVM load succeeded: Version " << std::endl;
     jint ver = env->GetVersion();
     std::cout << ((ver >> 16) & 0x0f) << "." << (ver & 0x0f) << std::endl;
 
-    std::string className;
-    if(isHead) className = "graphics/scenery/insitu/Head";
-    else className = "graphics/scenery/insitu/InVisRenderer";
 
     jclass localClass;
     localClass = env->FindClass(className.c_str());  // try to find the class
@@ -124,7 +129,7 @@ InVis::InVis(int wSize, int cPartners, MPI_Comm vComm, bool isHead) {
 
     // if class found, continue
     std::cout << "Class found " << className << std::endl;
-    jmethodID ctor = env->GetMethodID(localClass, "<init>", "()V");  // FIND AN OBJECT CONSTRUCTOR
+    jmethodID ctor = env->GetMethodID(localClass, "<init>", "()V");  // find constructor
     if (ctor == nullptr) {
         if( env->ExceptionOccurred() ) {
             env->ExceptionDescribe();
@@ -155,6 +160,22 @@ InVis::InVis(int wSize, int cPartners, MPI_Comm vComm, bool isHead) {
     env->DeleteLocalRef(localObj);
 
     std::cout << "Global references have been created and local ones deleted" << std::endl;
+}
+
+vis_type InVis::getVisType() {
+    ShmBuffer * buf = new ShmBuffer(datatype_path, 0, false);
+    buf->update_key(true);
+    ptr_with_size pws = buf->attach();
+    int * ptr = (int *)pws.ptr;
+    if(*ptr == 1) {
+        return vis_type::grid;
+    } else if (*ptr == 2) {
+        return  vis_type::particles;
+    }
+    else {
+        std::cerr<<"Found unidentified data when searching for data type flag"<<std::endl;
+        std::exit(EXIT_FAILURE);
+    }
 }
 
 void InVis::updateMemory(jmethodID methodID, int memKey, bool pos) {
@@ -247,7 +268,7 @@ void sendImage(JNIEnv *e, jobject clazz, jobject image) {
     MPI_Request req;
     MPI_Status stat;
     buf = (jbyte *) e->GetDirectBufferAddress(image);
-    MPI_Isend((void *)buf, imgSize, MPI_BYTE, 0, 0, comm, &req);
+    MPI_Isend((void *)buf, imgSize, MPI_BYTE, 0, 0, renComm, &req);
     MPI_Wait(&req, &stat); //TODO: put wait before Isend or remove wait
 }
 
@@ -255,11 +276,11 @@ void InVis::receiveImages() {
     JNIEnv *env;
     jvm->AttachCurrentThread(reinterpret_cast<void **>(&env), NULL);
 
-    char buf[commSize-1][imageSize];
-    MPI_Request requests[commSize-1];
-    MPI_Status statuses[commSize-1];
+    char buf[rCommSize - 1][imageSize];
+    MPI_Request requests[rCommSize - 1];
+    MPI_Status statuses[rCommSize - 1];
     jobjectArray bytebuffers;
-    jsize size = commSize - 1;
+    jsize size = rCommSize - 1;
     bytebuffers = (env)->NewObjectArray(size, (env)->FindClass("java/nio/ByteBuffer"), 0);
     std::cout << "Looking for function composite"<<std::endl;
 
@@ -276,19 +297,19 @@ void InVis::receiveImages() {
     std::cout << "Function composite has been found!"<<std::endl;
 
     while (true) {
-        for(int i = 1; i < commSize; i++) {
+        for(int i = 1; i < rCommSize; i++) {
             MPI_Irecv(buf[i-1], imageSize, MPI_BYTE, i, 0, visComm, &requests[i-1]);
         }
-        MPI_Waitall(commSize-1, requests, statuses);
+        MPI_Waitall(rCommSize - 1, requests, statuses);
         jobject bb;
-        for(int i = 1; i < commSize; i++) {
+        for(int i = 1; i < rCommSize; i++) {
             bb = (env)->NewDirectByteBuffer((void*) buf[i-1], imageSize);
 //            int temp = (env)-> GetDirectBufferCapacity(bb);
 //            std::cout<<"Bytebuffer containing color and depth is of size "<<temp<<std::endl;
             (env)->SetObjectArrayElement(bytebuffers, i-1, bb);
         }
         std::cout << "Function composite will be called"<<std::endl;
-        (env)-> CallVoidMethod(obj, compositeMethod, bytebuffers, commSize);
+        (env)-> CallVoidMethod(obj, compositeMethod, bytebuffers, rCommSize);
     }
     jvm->DetachCurrentThread();
 }
@@ -309,6 +330,259 @@ void InVis::updateCamera() {
     jvm->DetachCurrentThread();
 }
 
+void InVis::getGridMemory() {
+    JNIEnv *env;
+    jvm->AttachCurrentThread(reinterpret_cast<void **>(&env), NULL);
+
+    jmethodID updateDataMethod = env->GetMethodID(clazz, "updateData", "(II[Ljava/nio/ByteBuffer;[I[I[I)V");
+    if(updateDataMethod == nullptr) {
+        if (env->ExceptionOccurred()) {
+            env->ExceptionDescribe();
+        } else {
+            std::cout << "ERROR: function updateData not found!";
+//            std::exit(EXIT_FAILURE);
+        }
+    }
+
+//    std::cout<<"GOing to sleep"<<std::endl;
+//    sleep(5);
+    for(int i = 0; i < computePartners; i++) {
+        // Read the metadata of the grids managed by computePartner i
+        std::cout<<"Attempting to read metadata managed by computePartner " << i << std::endl;
+        ShmBuffer *GBoxBuf;
+        std::string pName = grid_shm_path + std::to_string(i);
+
+        GBoxBuf = new ShmBuffer(pName, 0, true);
+        ptr_with_size pws{};
+
+        std::cout<<"Calling update key function"<<std::endl;
+        GBoxBuf->update_key(true);
+        pws = GBoxBuf->attach();
+
+        size_t metaSize = pws.size - 512 - 32;
+        int numMetaElements  = metaSize / sizeof(long);
+        std::cout << "Got metadata for compute partner " << i << ". The size and no of elements are:" << metaSize << " and " << numMetaElements << std::endl;
+        long *metaData = (long*)pws.ptr;
+
+        for(int j = 0; j < numMetaElements; j++) {
+            std::cout<<j+1<<"th element is "<<metaData[j]<<"\t";
+        }
+        std::cout<<std::endl;
+
+        int elementsInGBox = 15; // origin -> 3 + grid dimensions -> 6 + domain dimensions -> 6
+        int numGrids = numMetaElements / elementsInGBox;
+
+        int origins[3 * numGrids];
+        jintArray jOrigins = env->NewIntArray(3 * numGrids);
+        int gridDims[6 * numGrids];
+        jintArray jGridDims = env->NewIntArray(6 * numGrids);
+        int domainDims[6 * numGrids];
+        jintArray jDomainDims = env->NewIntArray(6 * numGrids);
+
+        jobjectArray bytebuffers; // Each element will be a direct byte buffer wrapping a grid that is controlled by this computePartner
+        jsize size = numGrids;
+        bytebuffers = (env)->NewObjectArray(size, (env)->FindClass("java/nio/ByteBuffer"), 0);
+
+        int it = 0;
+        for(int j = 0; j < numGrids; j++) {
+            gridDims[j * 6] = (int)metaData[it];
+            it ++;
+            gridDims[j * 6 + 1] = (int)metaData[it];
+            it++;
+            gridDims[j * 6 + 2] = (int)metaData[it];
+            it++;
+            gridDims[j * 6 + 3] = (int)metaData[it];
+            it ++;
+            gridDims[j * 6 + 4] = (int)metaData[it];
+            it++;
+            gridDims[j * 6 + 5] = (int)metaData[it];
+            it++;
+
+            domainDims[j * 6] = (int)metaData[it];
+            it ++;
+            domainDims[j * 6 + 1] = (int)metaData[it];
+            it++;
+            domainDims[j * 6 + 2] = (int)metaData[it];
+            it++;
+            domainDims[j * 6 + 3] = (int)metaData[it];
+            it ++;
+            domainDims[j * 6 + 4] = (int)metaData[it];
+            it++;
+            domainDims[j * 6 + 5] = (int)metaData[it];
+            it++;
+
+            origins[j * 3] = (int)metaData[it];
+            it ++;
+            origins[j * 3 + 1] = (int)metaData[it];
+            it++;
+            origins[j * 3 + 2] = (int)metaData[it];
+            it++;
+
+            gridBuffers[i].push_back(new ShmBuffer(pName, j+1, true));
+            std::cout<<"1"<<std::endl;
+            gridBuffers[i][j]->update_key(true);
+            std::cout<<"2"<<std::endl;
+
+            ptr_with_size tmp = gridBuffers[i][j]->attach();
+            std::cout<<"2"<<std::endl;
+
+            int width = abs(gridDims[j*6] - gridDims[j*6 + 3]) + 1;
+            std::cout<<"3"<<std::endl;
+
+            int height = abs(gridDims[j*6 + 1] - gridDims[j*6 + 4]) + 1;
+            std::cout<<"4"<<std::endl;
+
+            int depth = abs(gridDims[j*6 + 2] - gridDims[j*6 + 5]) + 1;
+            std::cout<<"5"<<std::endl;
+
+            int gridSize = width * height * depth;
+            std::cout<<"6"<<std::endl;
+
+            std::cout<< "Grid " << j << " of computePartner " << i << " is of size " << gridSize << std::endl;
+
+            jobject bb;
+            bb = env->NewDirectByteBuffer((void *)tmp.ptr, 2*gridSize);
+            env->SetObjectArrayElement(bytebuffers, j, bb);
+            int temp = (env)-> GetDirectBufferCapacity(bb);
+            std::cout<<"Bytebuffer capacity was found to be " <<temp <<std::endl;
+//            env->DeleteLocalRef(bb); TODO: uncomment
+        }
+
+        env->SetIntArrayRegion(jOrigins, 0, 3 * numGrids, origins);
+        env->SetIntArrayRegion(jGridDims, 0, 6 * numGrids, gridDims);
+        env->SetIntArrayRegion(jDomainDims, 0, 6 * numGrids, domainDims);
+        // the meta arguments for the java function updateData are now ready
+        std::cout<<"the meta arguments for the java function updateData are now ready"<<std::endl;
+
+        env->CallVoidMethod(obj, updateDataMethod, i, numGrids, bytebuffers, jOrigins, jGridDims, jDomainDims);
+
+//        env->DeleteLocalRef(bytebuffers);
+//        env->DeleteLocalRef(jOrigins);
+//        env->DeleteLocalRef(jGridDims);
+//        env->DeleteLocalRef(jDomainDims);
+    }
+    jvm->DetachCurrentThread();
+}
+
+void distributeVDIs(JNIEnv *e, jobject clazzObject, jobject subVDICol, jint sizePerProcess, jint commSize) {
+    std::cout<<"In distribute VDIs function. Comm size is "<<commSize<<std::endl;
+
+    void *ptrCol = e->GetDirectBufferAddress(subVDICol);
+//    void *ptrDepth = e->GetDirectBufferAddress(subVDIDepth);
+
+    void *recvBufCol = malloc(sizePerProcess * commSize * 3);
+//    void *recvBufDepth = malloc(sizePerProcess * rCommSize * 2);
+
+    MPI_Alltoall(ptrCol, sizePerProcess*3, MPI_BYTE, recvBufCol, sizePerProcess*3, MPI_BYTE, renComm);
+//    MPI_Alltoall(ptrDepth, sizePerProcess*2, MPI_BYTE, recvBufDepth, sizePerProcess*2, MPI_BYTE, mpiComm);
+
+    jclass clazz = e->GetObjectClass(clazzObject);
+    jmethodID compositeMethod = e->GetMethodID(clazz, "compositeVDIs", "(Ljava/nio/ByteBuffer;I)V");
+
+    jobject bbCol = e->NewDirectByteBuffer(recvBufCol, sizePerProcess * commSize * 3);
+//    jobject bbDepth = e->NewDirectByteBuffer(recvBufDepth, sizePerProcess * rCommSize * 2);
+
+    if(e->ExceptionOccurred()) {
+        e->ExceptionDescribe();
+        e->ExceptionClear();
+    }
+
+    std::cout<<"Finished distributing the VDIs. Calling the Composite method now!"<<std::endl;
+
+    e->CallVoidMethod(clazzObject, compositeMethod, bbCol, sizePerProcess);
+    if(e->ExceptionOccurred()) {
+        e->ExceptionDescribe();
+        e->ExceptionClear();
+    }
+}
+
+void gatherCompositedVDIs(JNIEnv *e, jobject clazzObject, jobject compositedVDIColor, jint root, jint compositedVDILen, jint myRank, jint commSize, jboolean saveFiles) {
+
+    std::cout<<"In Gather function " <<std::endl;
+    void *ptrCol = e->GetDirectBufferAddress(compositedVDIColor);
+//    void *ptrDepth = e->GetDirectBufferAddress(compositedVDIDepth);
+//    std::cout << "Data received as parameter is: " << (char *)ptrCol << std::endl;
+    void *recvBufCol;
+//    void *recvBufDepth;
+    if (myRank == 0) {
+        recvBufCol = malloc(compositedVDILen * commSize * 3);
+//        recvBufDepth = malloc(compositedVDILen * rCommSize * 2);
+    }
+    else {
+        recvBufCol = NULL;
+//        recvBufDepth = NULL;
+    }
+
+    MPI_Gather(ptrCol, compositedVDILen*3, MPI_BYTE, recvBufCol, compositedVDILen*3, MPI_BYTE, root, renComm);
+//    MPI_Gather(ptrDepth, compositedVDILen*2, MPI_BYTE, recvBufDepth, compositedVDILen*2, MPI_BYTE, root, mpiComm);
+    //The data is here now!
+
+    if(myRank == 0) {
+//        //send or store the VDI
+        std::time_t t = std::time(0);
+
+        int timetaken = 0;
+        if(prevtime == 0) {
+            //skip this step
+        }
+        else {
+            timetaken = t - prevtime;
+            totaltime += timetaken;
+            avg_recent += timetaken;
+            count++;
+        }
+        if(count != 0 && count % 5 == 0) {
+            double average_time = (double)totaltime/count;
+            avg_recent = (double)avg_recent/5.0;
+            std::cout<<"Average time for the last 5 VDIs was: "<<avg_recent<<std::endl;
+            avg_recent = 0.0;
+            std::cout<<"Overall average so far is"<<average_time<<" and total VDIs generated so far are "<< count <<std::endl;
+
+        }
+        if(count != 0 && count % 10 == 0) {
+//        if(saveFiles) {
+
+            if(timeVDIwrite != 0) {
+                int time_elapsed = t - timeVDIwrite;
+                std::cout<<"Writing VDI now. Time since the last write was performed is: " << time_elapsed <<std::endl;
+                timeVDIwrite = t;
+            }
+            std::cout<<"Writing the final gathered VDI now"<<std::endl;
+
+            std::string filename = "size:" + std::to_string(commSize) + "Final_VDICol" + std::to_string(count) + ".raw";
+            std::ofstream b_stream(filename.c_str(),
+                                   std::fstream::out | std::fstream::binary);
+//            std::string filenameDepth = "size:" + std::to_string(rCommSize) + "Final_VDIDepth" + std::to_string(count) + ".raw";
+//            std::ofstream b_streamDepth(filenameDepth.c_str(),
+//                                   std::fstream::out | std::fstream::binary);
+
+            if (b_stream)
+            {
+                b_stream.write(static_cast<const char *>(recvBufCol), compositedVDILen * commSize * 3);
+//                b_streamDepth.write(static_cast<const char *>(recvBufDepth), compositedVDILen * rCommSize * 2);
+
+                if (b_stream.good()) {
+                    std::cout<<"Writing was successful"<<std::endl;
+                }
+            }
+
+        }
+        std::time_t t_out = std::time(0);
+        prevtime = t_out;
+//        std::cout << "cpp on rank " << myRank << " color data received is " << (char *)recvBufCol << std::endl;
+//        std::cout << "cpp on rank " << myRank << " depth data received is " << (char *)recvBufDepth << std::endl;
+    }
+
+//    jclass clazz = e->GetObjectClass(clazzObject);
+//    jmethodID updateMethod = e->GetMethodID(clazz, "updateVolumes", "()V");
+//
+//    e->CallVoidMethod(clazzObject, updateMethod);
+//    if(e->ExceptionOccurred()) {
+//        e->ExceptionDescribe();
+//        e->ExceptionClear();
+//    }
+}
+
 void InVis::doRender() {
     JNIEnv *env;
     jvm->GetEnv((void **)&env, JNI_VERSION_1_6);
@@ -323,30 +597,34 @@ void InVis::doRender() {
         }
     }
     env->CallVoidMethod(obj, mainMethod);
+    if(env->ExceptionOccurred()) {
+        env->ExceptionDescribe();
+        env->ExceptionClear();
+    }
 }
 
-void InVis::manageVisHead() {
-    JNIEnv *env;
-    jvm->GetEnv((void **)&env, JNI_VERSION_1_6);
-    jfieldID winSizeField = env->GetFieldID(clazz, "windowSize", "I");
-    env->SetIntField(obj, winSizeField, windowSize);
+//void InVis::manageVisHead() {
+//    JNIEnv *env;
+//    jvm->GetEnv((void **)&env, JNI_VERSION_1_6);
+//    jfieldID winSizeField = env->GetFieldID(clazz, "windowSize", "I");
+//    env->SetIntField(obj, winSizeField, windowSize);
+//
+//    std::cout<<"starting thread: gather images"<<std::endl;
+//    std::thread gatherImages(&InVis::receiveImages, this);
+//
+//    std::cout<<"starting thread: camera"<<std::endl;
+//    std::thread camera(&InVis::updateCamera, this);
+//
+//    std::cout<<"calling function: doRender"<<std::endl;
+//    doRender();
+//
+//    gatherImages.join();
+//    camera.join();
+//
+//    jvm->DestroyJavaVM();
+//}
 
-    std::cout<<"starting thread: gather images"<<std::endl;
-    std::thread gatherImages(&InVis::receiveImages, this);
-
-    std::cout<<"starting thread: camera"<<std::endl;
-    std::thread camera(&InVis::updateCamera, this);
-
-    std::cout<<"calling function: doRender"<<std::endl;
-    doRender();
-
-    gatherImages.join();
-    camera.join();
-
-    jvm->DestroyJavaVM();
-}
-
-void InVis::manageVisRenderer() {
+void InVis::manageRenderer() {
     JNIEnv *env;
     jvm->GetEnv((void **)&env, JNI_VERSION_1_6);
     jfieldID winSizeField = env->GetFieldID(clazz, "windowSize", "I");
@@ -355,11 +633,16 @@ void InVis::manageVisRenderer() {
     jfieldID compPartnersField = env->GetFieldID(clazz, "computePartners", "I");
     env->SetIntField(obj, compPartnersField, computePartners);
 
-    int visRank;
-    MPI_Comm_rank(visComm, &visRank);
+    if(vtype == vis_type::grid) {
+        jfieldID commSizeField = env->GetFieldID(clazz, "commSize", "I");
+        env->SetIntField(obj, commSizeField, rCommSize);
+    }
+
+    int renRank;
+    MPI_Comm_rank(renderComm, &renRank);
 
     jfieldID rankField = env->GetFieldID(clazz, "rank", "I");
-    env->SetIntField(obj, rankField, visRank);
+    env->SetIntField(obj, rankField, renRank);
 
     jmethodID initArraysMethod = env->GetMethodID(clazz, "initializeArrays", "()V");
 
@@ -373,49 +656,85 @@ void InVis::manageVisRenderer() {
     }
     env->CallVoidMethod(obj, initArraysMethod);
 
-    jclass RendererClass;
-    if(USE_VULKAN) {
-        RendererClass = env->FindClass("graphics/scenery/backends/vulkan/VulkanRenderer");  // try to find the clas
-    } else {
-        RendererClass = env->FindClass("graphics/scenery/backends/opengl/OpenGLRenderer");  // try to find the class
-    }
-
-    if(RendererClass == nullptr) {
-        if( env->ExceptionOccurred() ) {
-            env->ExceptionDescribe();
+    if(vtype == vis_type::particles) {
+        jclass RendererClass;
+        if(USE_VULKAN) {
+            RendererClass = env->FindClass("graphics/scenery/backends/vulkan/VulkanRenderer");  // try to find the clas
         } else {
-            std::cerr << "ERROR: Renderer class not found !" <<std::endl;
-            //std::exit(EXIT_FAILURE);
+            RendererClass = env->FindClass("graphics/scenery/backends/opengl/OpenGLRenderer");  // try to find the class
+        }
+
+        if(RendererClass == nullptr) {
+            if( env->ExceptionOccurred() ) {
+                env->ExceptionDescribe();
+            } else {
+                std::cerr << "ERROR: Renderer class not found !" <<std::endl;
+                //std::exit(EXIT_FAILURE);
+            }
+        }
+
+        JNINativeMethod methods[] { { (char *)"sendImage", (char *)"(Ljava/nio/ByteBuffer;)V", (void *)&sendImage } };  // mapping table
+
+        if(env->RegisterNatives(RendererClass, methods, 1) < 0) {                        // register it
+            if( env->ExceptionOccurred() ) {
+                env->ExceptionDescribe();
+            } else {
+                std::cerr << "ERROR: Could not register natives for Renderer !" <<std::endl;
+                //std::exit(EXIT_FAILURE);
+            }
+        } else {
+            std::cout<<"Native registered for Renderer. The return value is: "<< env->RegisterNatives(RendererClass, methods, 1) <<std::endl;
         }
     }
 
-    JNINativeMethod methods[] { { (char *)"sendImage", (char *)"(Ljava/nio/ByteBuffer;)V", (void *)&sendImage } };  // mapping table
+    else {
+        JNINativeMethod methods[] { { (char *)"distributeVDIs", (char *)"(Ljava/nio/ByteBuffer;II)V", (void *)&distributeVDIs },
+                                    { (char *)"gatherCompositedVDIs", (char *)"(Ljava/nio/ByteBuffer;IIIIZ)V", (void *)&gatherCompositedVDIs }
+        };  // mapping table
 
-    if(env->RegisterNatives(RendererClass, methods, 1) < 0) {                        // register it
-        if( env->ExceptionOccurred() ) {
-            env->ExceptionDescribe();
+        if(env->RegisterNatives(clazz, methods, 2) < 0) {                        // register it
+            if( env->ExceptionOccurred() ) {
+                env->ExceptionDescribe();
+            } else {
+                std::cerr << "ERROR: Could not register natives for InVisVolume!" <<std::endl;
+                //std::exit(EXIT_FAILURE);
+            }
         } else {
-            std::cerr << "ERROR: Could not register natives for Renderer !" <<std::endl;
-            //std::exit(EXIT_FAILURE);
+            std::cout<<"Natives registered for InVisVolume. The return value is: "<< env->RegisterNatives(clazz, methods, 2) <<std::endl;
         }
-    } else {
-        std::cout<<"Native registered for Renderer. The return value is: "<< env->RegisterNatives(RendererClass, methods, 1) <<std::endl;
     }
 
-    std::cout<<"starting thread: updatePosData"<<std::endl;
+    if(vtype == vis_type::particles) {
+        std::cout<<"starting thread: updatePosData"<<std::endl;
 
-    std::thread updatePosData(&InVis::getMemoryPos, this);
+        std::thread updatePosData(&InVis::getMemoryPos, this);
 
-    std::cout<<"starting thread: updatePropsData"<<std::endl;
+        std::cout<<"starting thread: updatePropsData"<<std::endl;
 
-    std::thread updatePropsData(&InVis::getMemoryProps, this);
+        std::thread updatePropsData(&InVis::getMemoryProps, this);
 
-    std::cout<<"calling function: doRender"<<std::endl;
+        std::cout<<"calling function: doRender"<<std::endl;
 
-    doRender();
+        doRender();
 
-    updatePosData.join();
-    updatePropsData.join();
+        updatePosData.join();
+        updatePropsData.join();
+    }
+
+    else {
+        std::cout<<"starting thread: updatePosData"<<std::endl;
+
+        std::thread updateGridData(&InVis::getGridMemory, this);
+
+        std::cout<<"calling function: doRender"<<std::endl;
+
+        doRender();
+
+        std::cout<<"Finished with doRender!"<<std::endl;
+
+        updateGridData.join();
+    }
+
 
     jvm->DestroyJavaVM();
 }
